@@ -6,20 +6,32 @@
     "homepage": "http://developers.google.com/api-client-library/php",
     "license": "Apache-2.0",
     "require": {
-        "php": ">=5.2.1"
+        "php": ">=5.4",
+        "google/auth": "0.5",
+        "firebase/php-jwt": "~2.0|~3.0",
+        "monolog/monolog": "^1.17",
+        "phpseclib/phpseclib": "~2.0",
+        "guzzlehttp/guzzle": "~5.2|~6.0",
+        "guzzlehttp/psr7": "1.2.*",
+        "psr/http-message": "1.0.*"
     },
     "require-dev": {
-      "phpunit/phpunit": "3.7.*"
+        "phpunit/phpunit": "~4",
+        "squizlabs/php_codesniffer": "~2.3",
+        "symfony/dom-crawler": "~2.0",
+        "symfony/css-selector": "~2.0"
     },
     "autoload": {
+        "psr-0": {
+            "Google_": "src/"
+        },
         "classmap": [
-            "src/"
+            "src/Google/Service/"
         ]
     },
-    "include-path": ["src/"],
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "2.x-dev"
         }
     }
 }
