@@ -15,48 +15,64 @@
  * limitations under the License.
  */
 
-require_once 'Google/Auth/AssertionCredentials.php';
-require_once 'Google/Cache/File.php';
-require_once 'Google/Cache/Memcache.php';
-require_once 'Google/Config.php';
-require_once 'Google/Collection.php';
-require_once 'Google/Exception.php';
-require_once 'Google/IO/Curl.php';
-require_once 'Google/IO/Stream.php';
-require_once 'Google/Model.php';
-require_once 'Google/Service.php';
-require_once 'Google/Service/Resource.php';
+use Google\Auth\ApplicationDefaultCredentials;
+use Google\Auth\CacheInterface;
+use Google\Auth\CredentialsLoader;
+use Google\Auth\HttpHandler\HttpHandlerFactory;
+use Google\Auth\OAuth2;
+use Google\Auth\Credentials\ServiceAccountCredentials;
+use Google\Auth\Credentials\UserRefreshCredentials;
+use GuzzleHttp\Client;
+use GuzzleHttp\ClientInterface;
+use GuzzleHttp\Ring\Client\StreamHandler;
+use GuzzleHttp\Psr7;
+use Psr\Http\Message\RequestInterface;
+use Psr\Log\LoggerInterface;
+use Monolog\Logger;
+use Monolog\Handler\StreamHandler as MonologStreamHandler;
 
 /**
  * The Google API Client
- * http://code.google.com/p/google-api-php-client/
- *
- * @author Chris Chabot <chabotc@google.com>
- * @author Chirag Shah <chirags@google.com>
+ * https://github.com/google/google-api-php-client
  */
 class Google_Client
 {
-  const LIBVER = "1.0.5-beta";
+  const LIBVER = "2.0.0-alpha";
   const USER_AGENT_SUFFIX = "google-api-php-client/";
+  const OAUTH2_REVOKE_URI = 'https://accounts.google.com/o/oauth2/revoke';
+  const OAUTH2_TOKEN_URI = 'https://www.googleapis.com/oauth2/v4/token';
+  const OAUTH2_AUTH_URL = 'https://accounts.google.com/o/oauth2/auth';
+  const API_BASE_PATH = 'https://www.googleapis.com';
+
   /**
-   * @var Google_Auth_Abstract $auth
+   * @var Google\Auth\OAuth2 $auth
    */
   private $auth;
 
   /**
-   * @var Google_IO_Abstract $io
+   * @var GuzzleHttp\ClientInterface $http
    */
-  private $io;
+  private $http;
 
   /**
-   * @var Google_Cache_Abstract $cache
+   * @var Google\Auth\CacheInterface $cache
    */
   private $cache;
+
+  /**
+   * @var array access token
+   */
+  private $token;
 
   /**
    * @var Google_Config $config
    */
   private $config;
+
+  /**
+   * @var Google_Logger_Abstract $logger
+   */
+  private $logger;
 
   /**
    * @var boolean $deferExecution
@@ -65,51 +81,57 @@ class Google_Client
 
   /** @var array $scopes */
   // Scopes requested by the client
-  protected $requestedScopes = array();
-
-  // definitions of services that are discovered.
-  protected $services = array();
-
-  // Used to track authenticated state, can't discover services after doing authenticate()
-  private $authenticated = false;
+  protected $requestedScopes = [];
 
   /**
    * Construct the Google Client.
    *
    * @param $config Google_Config or string for the ini file to load
    */
-  public function __construct($config = null)
+  public function __construct($config = array())
   {
-    if (! ini_get('date.timezone') &&
-        function_exists('date_default_timezone_set')) {
-      date_default_timezone_set('UTC');
-    }
+    $this->config = array_merge(
+        [
+          'application_name' => '',
 
-    if (is_string($config) && strlen($config)) {
-      $config = new Google_Config($config);
-    } else if ( !($config instanceof Google_Config)) {
-      $config = new Google_Config();
+          // Don't change these unless you're working against a special development
+          // or testing environment.
+          'base_path' => self::API_BASE_PATH,
 
-      if ($this->isAppEngine()) {
-        // Automatically use Memcache if we're in AppEngine.
-        $config->setCacheClass('Google_Cache_Memcache');
-      }
+          // https://developers.google.com/console
+          'client_id' => '',
+          'client_secret' => '',
+          'redirect_uri' => null,
+          'state' => null,
 
-      if (version_compare(phpversion(), "5.3.4", "<=") || $this->isAppEngine()) {
-        // Automatically disable compress.zlib, as currently unsupported.
-        $config->setClassConfig('Google_Http_Request', 'disable_gzip', true);
-      }
-    }
-    
-    if ($config->getIoClass() == Google_Config::USE_AUTO_IO_SELECTION) {
-      if (function_exists('curl_version') && function_exists('curl_exec')) {
-        $config->setIoClass("Google_IO_Curl");
-      } else {
-        $config->setIoClass("Google_IO_Stream");
-      }
-    }
+          // Simple API access key, also from the API console. Ensure you get
+          // a Server key, and not a Browser key.
+          'developer_key' => '',
 
-    $this->config = $config;
+          // For use with Google Cloud Platform
+          // fetch the ApplicationDefaultCredentials, if applicable
+          // @see https://developers.google.com/identity/protocols/application-default-credentials
+          'use_application_default_credentials' => false,
+          'signing_key' => null,
+          'signing_algorithm' => null,
+          'subject' => null,
+
+          // Other OAuth2 parameters.
+          'hd' => '',
+          'prompt' => '',
+          'openid.realm' => '',
+          'include_granted_scopes' => null,
+          'login_hint' => '',
+          'request_visible_actions' => '',
+          'access_type' => 'online',
+          'approval_prompt' => 'auto',
+
+          // Task Runner retry configuration
+          // @see Google_Task_Runner
+          'retry' => array(),
+        ],
+        $config
+    );
   }
 
   /**
@@ -123,134 +145,288 @@ class Google_Client
   }
 
   /**
+   * For backwards compatibility
+   * alias for fetchAccessTokenWithAuthCode
+   *
+   * @param $code string code from accounts.google.com
+   * @return array access token
+   */
+  public function authenticate($code)
+  {
+    return $this->fetchAccessTokenWithAuthCode($code);
+  }
+
+  /**
    * Attempt to exchange a code for an valid authentication token.
    * Helper wrapped around the OAuth 2.0 implementation.
    *
    * @param $code string code from accounts.google.com
-   * @return string token
+   * @return array access token
    */
-  public function authenticate($code)
+  public function fetchAccessTokenWithAuthCode($code)
   {
-    $this->authenticated = true;
-    return $this->getAuth()->authenticate($code);
-  }
-
-  /**
-   * Set the auth config from the JSON string provided.
-   * This structure should match the file downloaded from
-   * the "Download JSON" button on in the Google Developer
-   * Console.
-   * @param string $json the configuration json
-   */
-  public function setAuthConfig($json)
-  {
-    $data = json_decode($json);
-    $key = isset($data->installed) ? 'installed' : 'web';
-    if (!isset($data->$key)) {
-      throw new Google_Exception("Invalid client secret JSON file.");
+    if (strlen($code) == 0) {
+      throw new InvalidArgumentException("Invalid code");
     }
-    $this->setClientId($data->$key->client_id);
-    $this->setClientSecret($data->$key->client_secret);
-    if (isset($data->$key->redirect_uris)) {
-      $this->setRedirectUri($data->$key->redirect_uris[0]);
+
+    $auth = $this->getOAuth2Service();
+    $auth->setCode($code);
+    $auth->setRedirectUri($this->getRedirectUri());
+
+    $httpHandler = HttpHandlerFactory::build($this->getHttpClient());
+    $creds = $auth->fetchAuthToken($httpHandler);
+    if ($creds && isset($creds['access_token'])) {
+      $creds['created'] = time();
+      $this->setAccessToken($creds);
     }
+
+    return $creds;
   }
 
   /**
-   * Set the auth config from the JSON file in the path
-   * provided. This should match the file downloaded from
-   * the "Download JSON" button on in the Google Developer
-   * Console.
-   * @param string $file the file location of the client json
+   * For backwards compatibility
+   * alias for fetchAccessTokenWithAssertion
+   *
+   * @return array access token
    */
-  public function setAuthConfigFile($file)
+  public function refreshTokenWithAssertion()
   {
-    $this->setAuthConfig(file_get_contents($file));
+    return $this->fetchAccessTokenWithAssertion();
   }
 
   /**
-   * @return array
-   * @visible For Testing
+   * Fetches a fresh access token with a given assertion token.
+   * @param $assertionCredentials optional.
+   * @return array access token
    */
-  public function prepareScopes()
+  public function fetchAccessTokenWithAssertion(ClientInterface $authHttp = null)
   {
-    if (empty($this->requestedScopes)) {
-      throw new Google_Auth_Exception("No scopes specified");
+    if (!$this->isUsingApplicationDefaultCredentials()) {
+      throw new DomainException(
+          'set the JSON service account credentials using'
+          . ' Google_Client::setAuthConfig or set the path to your JSON file'
+          . ' with the "GOOGLE_APPLICATION_CREDENTIALS" environment variable'
+          . ' and call Google_Client::useApplicationDefaultCredentials to'
+          . ' refresh a token with assertion.'
+      );
     }
-    $scopes = implode(' ', $this->requestedScopes);
-    return $scopes;
-  }
 
-  /**
-   * Set the OAuth 2.0 access token using the string that resulted from calling createAuthUrl()
-   * or Google_Client#getAccessToken().
-   * @param string $accessToken JSON encoded string containing in the following format:
-   * {"access_token":"TOKEN", "refresh_token":"TOKEN", "token_type":"Bearer",
-   *  "expires_in":3600, "id_token":"TOKEN", "created":1320790426}
-   */
-  public function setAccessToken($accessToken)
-  {
-    if ($accessToken == 'null') {
-      $accessToken = null;
+    $this->getLogger()->log(
+        'info',
+        'OAuth2 access token refresh with Signed JWT assertion grants.'
+    );
+
+    $credentials = $this->createApplicationDefaultCredentials();
+
+    $httpHandler = HttpHandlerFactory::build($authHttp);
+    $accessToken = $credentials->fetchAuthToken($httpHandler);
+    if ($accessToken && isset($accessToken['access_token'])) {
+      $this->setAccessToken($accessToken);
     }
-    $this->getAuth()->setAccessToken($accessToken);
+
+    return $accessToken;
   }
 
-
-
   /**
-   * Set the authenticator object
-   * @param Google_Auth_Abstract $auth
+   * For backwards compatibility
+   * alias for fetchAccessTokenWithRefreshToken
+   *
+   * @param string $refreshToken
+   * @return array access token
    */
-  public function setAuth(Google_Auth_Abstract $auth)
+  public function refreshToken($refreshToken)
   {
-    $this->config->setAuthClass(get_class($auth));
-    $this->auth = $auth;
+    return $this->fetchAccessTokenWithRefreshToken($refreshToken);
   }
 
   /**
-   * Set the IO object
-   * @param Google_Io_Abstract $auth
+   * Fetches a fresh OAuth 2.0 access token with the given refresh token.
+   * @param string $refreshToken
+   * @return array access token
    */
-  public function setIo(Google_Io_Abstract $io)
+  public function fetchAccessTokenWithRefreshToken($refreshToken = null)
   {
-    $this->config->setIoClass(get_class($io));
-    $this->io = $io;
+    if (is_null($refreshToken)) {
+      if (!isset($this->token['refresh_token'])) {
+        throw new LogicException(
+            'refresh token must be passed in or set as part of setAccessToken'
+        );
+      }
+      $refreshToken = $this->token['refresh_token'];
+    }
+    $this->getLogger()->info('OAuth2 access token refresh');
+    $auth = $this->getOAuth2Service();
+    $auth->setRefreshToken($refreshToken);
+
+    $httpHandler = HttpHandlerFactory::build($this->getHttpClient());
+    $creds = $auth->fetchAuthToken($httpHandler);
+    if ($creds && isset($creds['access_token'])) {
+      $creds['created'] = time();
+      $this->setAccessToken($creds);
+    }
+
+    return $creds;
   }
 
   /**
-   * Set the Cache object
-   * @param Google_Cache_Abstract $auth
-   */
-  public function setCache(Google_Cache_Abstract $cache)
-  {
-    $this->config->setCacheClass(get_class($cache));
-    $this->cache = $cache;
-  }
-
-  /**
-   * Construct the OAuth 2.0 authorization request URI.
+   * Create a URL to obtain user authorization.
+   * The authorization endpoint allows the user to first
+   * authenticate, and then grant/deny the access request.
+   * @param string|array $scope The scope is expressed as an array or list of space-delimited strings.
    * @return string
    */
-  public function createAuthUrl()
+  public function createAuthUrl($scope = null)
   {
-    $scopes = $this->prepareScopes();
-    return $this->getAuth()->createAuthUrl($scopes);
+    if (empty($scope)) {
+      $scope = $this->prepareScopes();
+    }
+    if (is_array($scope)) {
+      $scope = implode(' ', $scope);
+    }
+
+    // only accept one of prompt or approval_prompt
+    $approvalPrompt = $this->config['prompt']
+      ? null
+      : $this->config['approval_prompt'];
+
+    // include_granted_scopes should be string "true", string "false", or null
+    $includeGrantedScopes = $this->config['include_granted_scopes'] === null
+      ? null
+      : var_export($this->config['include_granted_scopes'], true);
+
+    $params = array_filter(
+        [
+          'access_type' => $this->config['access_type'],
+          'approval_prompt' => $approvalPrompt,
+          'hd' => $this->config['hd'],
+          'include_granted_scopes' => $includeGrantedScopes,
+          'login_hint' => $this->config['login_hint'],
+          'openid.realm' => $this->config['openid.realm'],
+          'prompt' => $this->config['prompt'],
+          'response_type' => 'code',
+          'scope' => $scope,
+          'state' => $this->config['state'],
+        ]
+    );
+
+    // If the list of scopes contains plus.login, add request_visible_actions
+    // to auth URL.
+    $rva = $this->config['request_visible_actions'];
+    if (strlen($rva) > 0 && false !== strpos($scope, 'plus.login')) {
+        $params['request_visible_actions'] = $rva;
+    }
+
+    $auth = $this->getOAuth2Service();
+
+    return (string) $auth->buildFullAuthorizationUri($params);
   }
 
   /**
-   * Get the OAuth 2.0 access token.
-   * @return string $accessToken JSON encoded string in the following format:
-   * {"access_token":"TOKEN", "refresh_token":"TOKEN", "token_type":"Bearer",
-   *  "expires_in":3600,"id_token":"TOKEN", "created":1320790426}
+   * Adds auth listeners to the HTTP client based on the credentials
+   * set in the Google API Client object
+   *
+   * @param GuzzleHttp\ClientInterface $http the http client object.
+   * @param GuzzleHttp\ClientInterface $authHttp an http client for authentication.
+   * @return GuzzleHttp\ClientInterface the http client object
    */
+  public function authorize(ClientInterface $http = null, ClientInterface $authHttp = null)
+  {
+    $credentials = null;
+    $token = null;
+    $scopes = null;
+    if (is_null($http)) {
+      $http = $this->getHttpClient();
+    }
+
+    // These conditionals represent the decision tree for authentication
+    //   1.  Check for Application Default Credentials
+    //   2.  Check for API Key
+    //   3a. Check for an Access Token
+    //   3b. If access token exists but is expired, try to refresh it
+    if ($this->isUsingApplicationDefaultCredentials()) {
+      $credentials = $this->createApplicationDefaultCredentials();
+    } elseif ($token = $this->getAccessToken()) {
+      $scopes = $this->prepareScopes();
+      // add refresh subscriber to request a new token
+      if ($this->isAccessTokenExpired() && isset($token['refresh_token'])) {
+        $credentials = $this->createUserRefreshCredentials(
+            $scopes,
+            $token['refresh_token']
+        );
+      }
+    }
+
+    $authHandler = $this->getAuthHandler();
+
+    if ($credentials) {
+      $http = $authHandler->attachCredentials($http, $credentials);
+    } elseif ($token) {
+      $http = $authHandler->attachToken($http, $token, (array) $scopes);
+    } elseif ($key = $this->config['developer_key']) {
+      $http = $authHandler->attachKey($http, $key);
+    }
+
+    return $http;
+  }
+
+  /**
+   * Set the configuration to use application default credentials for
+   * authentication
+   *
+   * @see https://developers.google.com/identity/protocols/application-default-credentials
+   * @param boolean $useAppCreds
+   */
+  public function useApplicationDefaultCredentials($useAppCreds = true)
+  {
+    $this->config['use_application_default_credentials'] = $useAppCreds;
+  }
+
+  /**
+   * To prevent useApplicationDefaultCredentials from inappropriately being
+   * called in a conditional
+   *
+   * @see https://developers.google.com/identity/protocols/application-default-credentials
+   */
+  public function isUsingApplicationDefaultCredentials()
+  {
+    return $this->config['use_application_default_credentials'];
+  }
+
+  /**
+   * @param string|array $token
+   * @throws InvalidArgumentException
+   */
+  public function setAccessToken($token)
+  {
+    if (is_string($token)) {
+      if ($json = json_decode($token, true)) {
+        $token = $json;
+      } else {
+        // assume $token is just the token string
+        $token = array(
+          'access_token' => $token,
+        );
+      }
+    }
+    if ($token == null) {
+      throw new InvalidArgumentException('invalid json token');
+    }
+    if (!isset($token['access_token'])) {
+      throw new InvalidArgumentException("Invalid token format");
+    }
+    $this->token = $token;
+  }
+
   public function getAccessToken()
   {
-    $token = $this->getAuth()->getAccessToken();
-    // The response is json encoded, so could be the string null.
-    // It is arguable whether this check should be here or lower
-    // in the library.
-    return (null == $token || 'null' == $token || '[]' == $token) ? null : $token;
+    return $this->token;
+  }
+
+  public function getRefreshToken()
+  {
+    if (isset($this->token['refresh_token'])) {
+      return $this->token['refresh_token'];
+    }
   }
 
   /**
@@ -259,7 +435,89 @@ class Google_Client
    */
   public function isAccessTokenExpired()
   {
-    return $this->getAuth()->isAccessTokenExpired();
+    if (!$this->token) {
+      return true;
+    }
+
+    $created = 0;
+    if (isset($this->token['created'])) {
+      $created = $this->token['created'];
+    } elseif (isset($this->token['id_token'])) {
+      // check the ID token for "iat"
+      // signature verification is not required here, as we are just
+      // using this for convenience to save a round trip request
+      // to the Google API server
+      $idToken = $this->token['id_token'];
+      if (substr_count($idToken, '.') == 2) {
+        $parts = explode('.', $idToken);
+        $payload = json_decode(base64_decode($parts[1]), true);
+        if ($payload && isset($payload['iat'])) {
+          $created = $payload['iat'];
+        }
+      }
+    }
+
+    // If the token is set to expire in the next 30 seconds.
+    $expired = ($created
+      + ($this->token['expires_in'] - 30)) < time();
+
+    return $expired;
+  }
+
+  public function getAuth()
+  {
+    throw new BadMethodCallException(
+        'This function no longer exists. See UPGRADING.md for more information'
+    );
+  }
+
+  public function setAuth($auth)
+  {
+    throw new BadMethodCallException(
+        'This function no longer exists. See UPGRADING.md for more information'
+    );
+  }
+
+  /**
+   * Set the OAuth 2.0 Client ID.
+   * @param string $clientId
+   */
+  public function setClientId($clientId)
+  {
+    $this->config['client_id'] = $clientId;
+  }
+
+  public function getClientId()
+  {
+    return $this->config['client_id'];
+  }
+
+  /**
+   * Set the OAuth 2.0 Client Secret.
+   * @param string $clientSecret
+   */
+  public function setClientSecret($clientSecret)
+  {
+    $this->config['client_secret'] = $clientSecret;
+  }
+
+  public function getClientSecret()
+  {
+    return $this->config['client_secret'];
+  }
+
+  /**
+   * Set the OAuth 2.0 Redirect URI.
+   * @param string $redirectUri
+   */
+  public function setRedirectUri($redirectUri)
+  {
+    $this->config['redirect_uri'] = $redirectUri;
+  }
+
+  public function getRedirectUri()
+  {
+    return $this->config['redirect_uri'];
   }
 
   /**
@@ -269,7 +527,7 @@ class Google_Client
    */
   public function setState($state)
   {
-    $this->getAuth()->setState($state);
+    $this->config['state'] = $state;
   }
 
   /**
@@ -279,7 +537,7 @@ class Google_Client
    */
   public function setAccessType($accessType)
   {
-    $this->config->setAccessType($accessType);
+    $this->config['access_type'] = $accessType;
   }
 
   /**
@@ -289,7 +547,16 @@ class Google_Client
    */
   public function setApprovalPrompt($approvalPrompt)
   {
-    $this->config->setApprovalPrompt($approvalPrompt);
+    $this->config['approval_prompt'] = $approvalPrompt;
+  }
+
+  /**
+   * Set the login hint, email address or sub id.
+   * @param string $loginHint
+   */
+  public function setLoginHint($loginHint)
+  {
+    $this->config['login_hint'] = $loginHint;
   }
 
   /**
@@ -298,34 +565,7 @@ class Google_Client
    */
   public function setApplicationName($applicationName)
   {
-    $this->config->setApplicationName($applicationName);
-  }
-
-  /**
-   * Set the OAuth 2.0 Client ID.
-   * @param string $clientId
-   */
-  public function setClientId($clientId)
-  {
-    $this->config->setClientId($clientId);
-  }
-
-  /**
-   * Set the OAuth 2.0 Client Secret.
-   * @param string $clientSecret
-   */
-  public function setClientSecret($clientSecret)
-  {
-    $this->config->setClientSecret($clientSecret);
-  }
-
-  /**
-   * Set the OAuth 2.0 Redirect URI.
-   * @param string $redirectUri
-   */
-  public function setRedirectUri($redirectUri)
-  {
-    $this->config->setRedirectUri($redirectUri);
+    $this->config['application_name'] = $applicationName;
   }
 
   /**
@@ -341,7 +581,7 @@ class Google_Client
     if (is_array($requestVisibleActions)) {
       $requestVisibleActions = join(" ", $requestVisibleActions);
     }
-    $this->config->setRequestVisibleActions($requestVisibleActions);
+    $this->config['request_visible_actions'] = $requestVisibleActions;
   }
 
   /**
@@ -351,68 +591,95 @@ class Google_Client
    */
   public function setDeveloperKey($developerKey)
   {
-    $this->config->setDeveloperKey($developerKey);
+    $this->config['developer_key'] = $developerKey;
   }
 
   /**
-   * Fetches a fresh OAuth 2.0 access token with the given refresh token.
-   * @param string $refreshToken
-   * @return void
+   * Set the hd (hosted domain) parameter streamlines the login process for
+   * Google Apps hosted accounts. By including the domain of the user, you
+   * restrict sign-in to accounts at that domain.
+   * @param $hd string - the domain to use.
    */
-  public function refreshToken($refreshToken)
+  public function setHostedDomain($hd)
   {
-    return $this->getAuth()->refreshToken($refreshToken);
+    $this->config['hd'] = $hd;
+  }
+  /**
+   * Set the prompt hint. Valid values are none, consent and select_account.
+   * If no value is specified and the user has not previously authorized
+   * access, then the user is shown a consent screen.
+   * @param $prompt string
+   */
+  public function setPrompt($prompt)
+  {
+    $this->config['prompt'] = $prompt;
+  }
+  /**
+   * openid.realm is a parameter from the OpenID 2.0 protocol, not from OAuth
+   * 2.0. It is used in OpenID 2.0 requests to signify the URL-space for which
+   * an authentication request is valid.
+   * @param $realm string - the URL-space to use.
+   */
+  public function setOpenidRealm($realm)
+  {
+    $this->config['openid.realm'] = $realm;
+  }
+  /**
+   * If this is provided with the value true, and the authorization request is
+   * granted, the authorization will include any previous authorizations
+   * granted to this user/application combination for other scopes.
+   * @param $include boolean - the URL-space to use.
+   */
+  public function setIncludeGrantedScopes($include)
+  {
+    $this->config['include_granted_scopes'] = $include;
   }
 
   /**
    * Revoke an OAuth2 access token or refresh token. This method will revoke the current access
    * token, if a token isn't provided.
-   * @throws Google_Auth_Exception
+   *
    * @param string|null $token The token (access token or a refresh token) that should be revoked.
    * @return boolean Returns True if the revocation was successful, otherwise False.
    */
   public function revokeToken($token = null)
   {
-    return $this->getAuth()->revokeToken($token);
+    $tokenRevoker = new Google_AccessToken_Revoke(
+        $this->getHttpClient()
+    );
+
+    return $tokenRevoker->revokeToken($token ?: $this->getAccessToken());
   }
 
   /**
    * Verify an id_token. This method will verify the current id_token, if one
    * isn't provided.
-   * @throws Google_Auth_Exception
-   * @param string|null $token The token (id_token) that should be verified.
-   * @return Google_Auth_LoginTicket Returns an apiLoginTicket if the verification was
-   * successful.
-   */
-  public function verifyIdToken($token = null)
-  {
-    return $this->getAuth()->verifyIdToken($token);
-  }
-
-  /**
-   * Verify a JWT that was signed with your own certificates.
    *
-   * @param $jwt the token
-   * @param $certs array of certificates
-   * @param $required_audience the expected consumer of the token
-   * @param [$issuer] the expected issues, defaults to Google
-   * @param [$max_expiry] the max lifetime of a token, defaults to MAX_TOKEN_LIFETIME_SECS
-   * @return token information if valid, false if not
+   * @throws Google_Exception
+   * @param string|null $idToken The token (id_token) that should be verified.
+   * @return array|false Returns the token payload as an array if the verification was
+   * successful, false otherwise.
    */
-  public function verifySignedJwt($id_token, $cert_location, $audience, $issuer, $max_expiry = null)
+  public function verifyIdToken($idToken = null)
   {
-    $auth = new Google_Auth_OAuth2($this);
-    $certs = $auth->retrieveCertsFromLocation($cert_location);
-    return $auth->verifySignedJwtWithCerts($id_token, $certs, $audience, $issuer, $max_expiry);
-  }
+    $tokenVerifier = new Google_AccessToken_Verify(
+        $this->getHttpClient()
+    );
 
-  /**
-   * @param Google_Auth_AssertionCredentials $creds
-   * @return void
-   */
-  public function setAssertionCredentials(Google_Auth_AssertionCredentials $creds)
-  {
-    $this->getAuth()->setAssertionCredentials($creds);
+    if (is_null($idToken)) {
+      $token = $this->getAccessToken();
+      if (!isset($token['id_token'])) {
+        throw new LogicException(
+            'id_token must be passed in or set as part of setAccessToken'
+        );
+      }
+      $idToken = $token['id_token'];
+    }
+
+    return $tokenVerifier->verifyIdToken(
+        $idToken,
+        $this->getClientId()
+    );
   }
 
   /**
@@ -456,6 +723,42 @@ class Google_Client
   }
 
   /**
+   * @return array
+   * @visible For Testing
+   */
+  public function prepareScopes()
+  {
+    if (empty($this->requestedScopes)) {
+      return null;
+    }
+    $scopes = implode(' ', $this->requestedScopes);
+    return $scopes;
+  }
+
+  /**
+   * Helper method to execute deferred HTTP requests.
+   *
+   * @param $request Psr\Http\Message\RequestInterface|Google_Http_Batch
+   * @throws Google_Exception
+   * @return object of the type of the expected class or Psr\Http\Message\ResponseInterface.
+   */
+  public function execute(RequestInterface $request, $expectedClass = null)
+  {
+    $request = $request->withHeader(
+        'User-Agent',
+        $this->config['application_name']
+        . " " . self::USER_AGENT_SUFFIX
+        . $this->getLibraryVersion()
+    );
+
+    // call the authorize method
+    // this is where most of the grunt work is done
+    $http = $this->authorize();
+
+    return Google_Http_REST::execute($http, $request, $expectedClass, $this->config['retry']);
+  }
+
+  /**
    * Declare whether batch calls should be used. This may increase throughput
    * by making multiple requests in one connection.
    *
@@ -466,6 +769,97 @@ class Google_Client
   {
     // This is actually an alias for setDefer.
     $this->setDefer($useBatch);
+  }
+
+  /**
+   * Are we running in Google AppEngine?
+   * return bool
+   */
+  public function isAppEngine()
+  {
+    return (isset($_SERVER['SERVER_SOFTWARE']) &&
+        strpos($_SERVER['SERVER_SOFTWARE'], 'Google App Engine') !== false);
+  }
+
+  public function setConfig($name, $value)
+  {
+    $this->config[$name] = $value;
+  }
+
+  public function getConfig($name, $default = null)
+  {
+    return isset($this->config[$name]) ? $this->config[$name] : $default;
+  }
+
+  /**
+   * For backwards compatibility
+   * alias for setAuthConfig
+   *
+   * @param string $file the configuration file
+   * @throws Google_Exception
+   */
+  public function setAuthConfigFile($file)
+  {
+    $this->setAuthConfig($file);
+  }
+
+  /**
+   * Set the auth config from new or deprecated JSON config.
+   * This structure should match the file downloaded from
+   * the "Download JSON" button on in the Google Developer
+   * Console.
+   * @param string|array $json the configuration json
+   * @throws Google_Exception
+   */
+  public function setAuthConfig($config)
+  {
+    if (is_string($config)) {
+      if (!file_exists($config)) {
+        throw new InvalidArgumentException('file does not exist');
+      }
+
+      $json = file_get_contents($config);
+
+      if (!$config = json_decode($json, true)) {
+        throw new LogicException('invalid json for auth config');
+      }
+    }
+
+    $key = isset($config['installed']) ? 'installed' : 'web';
+    if (isset($config['type']) && $config['type'] == 'service_account') {
+      // application default credentials
+      $this->useApplicationDefaultCredentials();
+
+      // set the information from the config
+      $this->setClientId($config['client_id']);
+      $this->config['client_email'] = $config['client_email'];
+      $this->config['signing_key'] = $config['private_key'];
+      $this->config['signing_algorithm'] = 'HS256';
+    } elseif (isset($config[$key])) {
+      // old-style
+      $this->setClientId($config[$key]['client_id']);
+      $this->setClientSecret($config[$key]['client_secret']);
+      if (isset($config[$key]['redirect_uris'])) {
+        $this->setRedirectUri($config[$key]['redirect_uris'][0]);
+      }
+    } else {
+      // new-style
+      $this->setClientId($config['client_id']);
+      $this->setClientSecret($config['client_secret']);
+      if (isset($config['redirect_uris'])) {
+        $this->setRedirectUri($config['redirect_uris'][0]);
+      }
+    }
+  }
+
+  /**
+   * Use when the service account has been delegated domain wide access.
+   *
+   * @param string subject an email address account to impersonate
+   */
+  public function setSubject($subject)
+  {
+    $this->config['subject'] = $subject;
   }
 
   /**
@@ -480,31 +874,6 @@ class Google_Client
   }
 
   /**
-   * Helper method to execute deferred HTTP requests.
-   *
-   * @returns object of the type of the expected class or array.
-   */
-  public function execute($request)
-  {
-    if ($request instanceof Google_Http_Request) {
-      $request->setUserAgent(
-          $this->getApplicationName()
-          . " " . self::USER_AGENT_SUFFIX
-          . $this->getLibraryVersion()
-      );
-      if (!$this->getClassConfig("Google_Http_Request", "disable_gzip")) {
-        $request->enableGzip();
-      }
-      $request->maybeMoveParametersToBody();
-      return Google_Http_REST::execute($this, $request);
-    } else if ($request instanceof Google_Http_Batch) {
-      return $request->execute();
-    } else {
-      throw new Google_Exception("Do not know how to execute this type of object.");
-    }
-  }
-
-  /**
    * Whether or not to return raw requests
    * @return boolean
    */
@@ -514,95 +883,192 @@ class Google_Client
   }
 
   /**
-   * @return Google_Auth_Abstract Authentication implementation
+   * @return Google\Auth\OAuth2 implementation
    */
-  public function getAuth()
+  public function getOAuth2Service()
   {
     if (!isset($this->auth)) {
-      $class = $this->config->getAuthClass();
-      $this->auth = new $class($this);
+      $this->auth = $this->createOAuth2Service();
     }
+
     return $this->auth;
   }
 
   /**
-   * @return Google_IO_Abstract IO implementation
+   * create a default google auth object
    */
-  public function getIo()
+  protected function createOAuth2Service()
   {
-    if (!isset($this->io)) {
-      $class = $this->config->getIoClass();
-      $this->io = new $class($this);
-    }
-    return $this->io;
+    $auth = new OAuth2(
+        [
+          'clientId'          => $this->getClientId(),
+          'clientSecret'      => $this->getClientSecret(),
+          'authorizationUri'   => self::OAUTH2_AUTH_URL,
+          'tokenCredentialUri' => self::OAUTH2_TOKEN_URI,
+          'redirectUri'       => $this->getRedirectUri(),
+          'issuer'            => $this->config['client_id'],
+          'signingKey'        => $this->config['signing_key'],
+          'signingAlgorithm'  => $this->config['signing_algorithm'],
+        ]
+    );
+
+    return $auth;
   }
 
   /**
-   * @return Google_Cache_Abstract Cache implementation
+   * Set the Cache object
+   * @param Google\Auth\CacheInterface $cache
+   */
+  public function setCache(CacheInterface $cache)
+  {
+    $this->cache = $cache;
+  }
+
+  /**
+   * @return Google\Auth\CacheInterface Cache implementation
    */
   public function getCache()
   {
     if (!isset($this->cache)) {
-      $class = $this->config->getCacheClass();
-      $this->cache = new $class($this);
+      $this->cache = $this->createDefaultCache();
     }
+
     return $this->cache;
   }
 
-  /**
-   * Retrieve custom configuration for a specific class.
-   * @param $class string|object - class or instance of class to retrieve
-   * @param $key string optional - key to retrieve
-   */
-  public function getClassConfig($class, $key = null)
+  protected function createDefaultCache()
   {
-    if (!is_string($class)) {
-      $class = get_class($class);
+    if ($this->isAppEngine()) {
+      $cache = new Google_Cache_Memcache();
+    } else {
+      $cacheDir = sys_get_temp_dir() . '/google-api-php-client';
+      $cache = new Google_Cache_File($cacheDir);
     }
-    return $this->config->getClassConfig($class, $key);
+
+    return $cache;
   }
 
   /**
-   * Set configuration specific to a given class.
-   * $config->setClassConfig('Google_Cache_File',
-   *   array('directory' => '/tmp/cache'));
-   * @param $class The class name for the configuration
-   * @param $config string key or an array of configuration values
-   * @param $value optional - if $config is a key, the value
-   *
+   * Set the Logger object
+   * @param Psr\Log\LoggerInterface $logger
    */
-  public function setClassConfig($class, $config, $value = null)
+  public function setLogger(LoggerInterface $logger)
   {
-    if (!is_string($class)) {
-      $class = get_class($class);
+    $this->logger = $logger;
+  }
+
+  /**
+   * @return Psr\Log\LoggerInterface implementation
+   */
+  public function getLogger()
+  {
+    if (!isset($this->logger)) {
+      $this->logger = $this->createDefaultLogger();
     }
-    return $this->config->setClassConfig($class, $config, $value);
 
+    return $this->logger;
+  }
+
+  protected function createDefaultLogger()
+  {
+    $logger = new Logger('google-api-php-client');
+    $logger->pushHandler(new MonologStreamHandler('php://stderr', Logger::NOTICE));
+
+    return $logger;
   }
 
   /**
-   * @return string the base URL to use for calls to the APIs
+   * Set the Http Client object
+   * @param GuzzleHttp\ClientInterface $http
    */
-  public function getBasePath()
+  public function setHttpClient(ClientInterface $http)
   {
-    return $this->config->getBasePath();
+    $this->http = $http;
   }
 
   /**
-   * @return string the name of the application
+   * @return GuzzleHttp\ClientInterface implementation
    */
-  public function getApplicationName()
+  public function getHttpClient()
   {
-    return $this->config->getApplicationName();
+    if (is_null($this->http)) {
+      $this->http = $this->createDefaultHttpClient();
+    }
+
+    return $this->http;
   }
 
-  /**
-   * Are we running in Google AppEngine?
-   * return bool
-   */
-  public function isAppEngine()
+  protected function createDefaultHttpClient()
   {
-    return (isset($_SERVER['SERVER_SOFTWARE']) &&
-        strpos($_SERVER['SERVER_SOFTWARE'], 'Google App Engine') !== false);
+    $options = ['exceptions' => false];
+
+    $version = ClientInterface::VERSION;
+    if ('5' === $version[0]) {
+      $options = [
+        'base_url' => $this->config['base_path'],
+        'defaults' => $options,
+      ];
+      if ($this->isAppEngine()) {
+        // set StreamHandler on AppEngine by default
+        $options['handler']  = new StreamHandler();
+        $options['defaults']['verify'] = '/etc/ca-certificates.crt';
+      }
+    } else {
+      // guzzle 6
+      $options['base_uri'] = $this->config['base_path'];
+    }
+
+    return new Client($options);
+  }
+
+  private function createApplicationDefaultCredentials()
+  {
+    $scopes = $this->prepareScopes();
+    $sub = $this->config['subject'];
+    $signingKey = $this->config['signing_key'];
+
+    // create credentials using values supplied in setAuthConfig
+    if ($signingKey) {
+      $serviceAccountCredentials = array(
+        'client_id' => $this->config['client_id'],
+        'client_email' => $this->config['client_email'],
+        'private_key' => $signingKey,
+        'type' => 'service_account',
+      );
+      $keyStream = Psr7\stream_for(json_encode($serviceAccountCredentials));
+      $credentials = CredentialsLoader::makeCredentials($scopes, $keyStream);
+    } else {
+      $credentials = ApplicationDefaultCredentials::getCredentials($scopes);
+    }
+
+    // for service account domain-wide authority (impersonating a user)
+    // @see https://developers.google.com/identity/protocols/OAuth2ServiceAccount
+    if ($sub) {
+      if (!$credentials instanceof ServiceAccountCredentials) {
+        throw new DomainException('domain-wide authority requires service account credentials');
+      }
+
+      $credentials->setSub($sub);
+    }
+
+    return $credentials;
+  }
+
+  protected function getAuthHandler()
+  {
+    return Google_AuthHandler_AuthHandlerFactory::build($this->getCache());
+  }
+
+  private function createUserRefreshCredentials($scope, $refreshToken)
+  {
+    $creds = array_filter(
+        array(
+          'client_id' => $this->getClientId(),
+          'client_secret' => $this->getClientSecret(),
+          'refresh_token' => $refreshToken,
+        )
+    );
+
+    return new UserRefreshCredentials($scope, $creds);
   }
 }
