@@ -14,54 +14,67 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-include_once "templates/base.php";
-session_start();
 
-set_include_path("../src/" . PATH_SEPARATOR . get_include_path());
-require_once 'Google/Client.php';
+include_once __DIR__ . '/../vendor/autoload.php';
+include_once "templates/base.php";
+
+echo pageHeader("Retrieving An Id Token");
+
+/*************************************************
+ * Ensure you've downloaded your oauth credentials
+ ************************************************/
+if (!$oauth_credentials = getOAuthCredentialsFile()) {
+  return missingOAuth2CredentialsWarning();
+}
 
 /************************************************
-  ATTENTION: Fill in these values! Make sure
-  the redirect URI is to this page, e.g:
-  http://localhost:8080/user-example.php
+ * NOTICE:
+ * The redirect URI is to the current page, e.g:
+ * http://localhost:8080/idtoken.php
  ************************************************/
-$client_id = '<YOUR_CLIENT_ID>';
-$client_secret = '<YOUR_CLIENT_SECRET>';
-$redirect_uri = '<YOUR_REDIRECT_URI>';
+$redirect_uri = 'http://' . $_SERVER['HTTP_HOST'] . $_SERVER['PHP_SELF'];
 
 $client = new Google_Client();
-$client->setClientId($client_id);
-$client->setClientSecret($client_secret);
+$client->setAuthConfig($oauth_credentials);
 $client->setRedirectUri($redirect_uri);
 $client->setScopes('email');
 
 /************************************************
-  If we're logging out we just need to clear our
-  local access token in this case
+ * If we're logging out we just need to clear our
+ * local access token in this case
  ************************************************/
 if (isset($_REQUEST['logout'])) {
-  unset($_SESSION['access_token']);
+  unset($_SESSION['id_token']);
 }
 
+
 /************************************************
-  If we have a code back from the OAuth 2.0 flow,
-  we need to exchange that with the authenticate()
-  function. We store the resultant access token
-  bundle in the session, and redirect to ourself.
+ * If we have a code back from the OAuth 2.0 flow,
+ * we need to exchange that with the
+ * Google_Client::fetchAccessTokenWithAuthCode()
+ * function. We store the resultant access token
+ * bundle in the session, and redirect to ourself.
  ************************************************/
 if (isset($_GET['code'])) {
-  $client->authenticate($_GET['code']);
-  $_SESSION['access_token'] = $client->getAccessToken();
-  $redirect = 'http://' . $_SERVER['HTTP_HOST'] . $_SERVER['PHP_SELF'];
-  header('Location: ' . filter_var($redirect, FILTER_SANITIZE_URL));
+  $token = $client->fetchAccessTokenWithAuthCode($_GET['code']);
+  $client->setAccessToken($token);
+
+  // store in the session also
+  $_SESSION['id_token'] = $token;
+
+  // redirect back to the example
+  header('Location: ' . filter_var($redirect_uri, FILTER_SANITIZE_URL));
 }
 
 /************************************************
   If we have an access token, we can make
   requests, else we generate an authentication URL.
  ************************************************/
-if (isset($_SESSION['access_token']) && $_SESSION['access_token']) {
-  $client->setAccessToken($_SESSION['access_token']);
+if (
+  !empty($_SESSION['id_token'])
+  && isset($_SESSION['id_token']['id_token'])
+) {
+  $client->setAccessToken($_SESSION['id_token']);
 } else {
   $authUrl = $client->createAuthUrl();
 }
@@ -75,32 +88,21 @@ if (isset($_SESSION['access_token']) && $_SESSION['access_token']) {
   and that can be cached.
  ************************************************/
 if ($client->getAccessToken()) {
-  $_SESSION['access_token'] = $client->getAccessToken();
-  $token_data = $client->verifyIdToken()->getAttributes();
-}
-
-echo pageHeader("User Query - Retrieving An Id Token");
-if (
-    $client_id == '<YOUR_CLIENT_ID>'
-    || $client_secret == '<YOUR_CLIENT_SECRET>'
-    || $redirect_uri == '<YOUR_REDIRECT_URI>') {
-  echo missingClientSecretsWarning();
+  $token_data = $client->verifyIdToken();
 }
 ?>
-<div class="box">
-  <div class="request">
-    <?php if (isset($authUrl)): ?>
-      <a class='login' href='<?php echo $authUrl; ?>'>Connect Me!</a>
-    <?php else: ?>
-      <a class='logout' href='?logout'>Logout</a>
-    <?php endif ?>
-  </div>
 
-  <?php if (isset($token_data)): ?>
-    <div class="data">
-      <?php var_dump($token_data); ?>
-    </div>
-  <?php endif ?>
+<div class="box">
+<?php if (isset($authUrl)): ?>
+  <div class="request">
+    <a class='login' href='<?= $authUrl ?>'>Connect Me!</a>
+  </div>
+<?php else: ?>
+  <div class="data">
+    <p>Here is the data from your Id Token:</p>
+    <pre><?php var_export($token_data) ?></pre>
+  </div>
+<?php endif ?>
 </div>
-<?php
-echo pageFooter(__FILE__);
+
+<?php echo pageFooter(__FILE__) ?>
