@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-require_once "Google/Cache/Abstract.php";
-require_once "Google/Cache/Exception.php";
+use Google\Auth\CacheInterface;
+use Psr\Log\LoggerInterface;
 
 /*
  * This class implements a basic on disk storage. While that does
@@ -26,39 +26,69 @@ require_once "Google/Cache/Exception.php";
  *
  * @author Chris Chabot <chabotc@google.com>
  */
-class Google_Cache_File extends Google_Cache_Abstract
+class Google_Cache_File implements CacheInterface
 {
   const MAX_LOCK_RETRIES = 10;
   private $path;
   private $fh;
 
-  public function __construct(Google_Client $client)
+  /**
+   * @var use Psr\Log\LoggerInterface logger
+   */
+  private $logger;
+
+  public function __construct($path, LoggerInterface $logger = null)
   {
-    $this->path = $client->getClassConfig($this, 'directory');
+    $this->path = $path;
+    $this->logger = $logger;
   }
-  
+
   public function get($key, $expiration = false)
   {
     $storageFile = $this->getCacheFile($key);
     $data = false;
-    
+
     if (!file_exists($storageFile)) {
+      $this->log(
+          'debug',
+          'File cache miss',
+          array('key' => $key, 'file' => $storageFile)
+      );
       return false;
     }
 
     if ($expiration) {
       $mtime = filemtime($storageFile);
       if ((time() - $mtime) >= $expiration) {
+        $this->log(
+            'debug',
+            'File cache miss (expired)',
+            array('key' => $key, 'file' => $storageFile)
+        );
         $this->delete($key);
         return false;
       }
     }
 
     if ($this->acquireReadLock($storageFile)) {
-      $data = fread($this->fh, filesize($storageFile));
-      $data =  unserialize($data);
-      $this->unlock($storageFile);
+      if (filesize($storageFile) > 0) {
+        $data = fread($this->fh, filesize($storageFile));
+        $data = unserialize($data);
+      } else {
+        $this->log(
+            'debug',
+            'Cache file was empty',
+            array('file' => $storageFile)
+        );
+      }
+      $this->unlock();
     }
+
+    $this->log(
+        'debug',
+        'File cache hit',
+        array('key' => $key, 'file' => $storageFile, 'var' => $data)
+    );
 
     return $data;
   }
@@ -70,8 +100,20 @@ class Google_Cache_File extends Google_Cache_Abstract
       // We serialize the whole request object, since we don't only want the
       // responseContent but also the postBody used, headers, size, etc.
       $data = serialize($value);
-      $result = fwrite($this->fh, $data);
-      $this->unlock($storageFile);
+      fwrite($this->fh, $data);
+      $this->unlock();
+
+      $this->log(
+          'debug',
+          'File cache set',
+          array('key' => $key, 'file' => $storageFile, 'var' => $value)
+      );
+    } else {
+      $this->log(
+          'notice',
+          'File cache set failed',
+          array('key' => $key, 'file' => $storageFile)
+      );
     }
   }
 
@@ -79,10 +121,21 @@ class Google_Cache_File extends Google_Cache_Abstract
   {
     $file = $this->getCacheFile($key);
     if (file_exists($file) && !unlink($file)) {
+      $this->log(
+          'error',
+          'File cache delete failed',
+          array('key' => $key, 'file' => $file)
+      );
       throw new Google_Cache_Exception("Cache file could not be deleted");
     }
+
+    $this->log(
+        'debug',
+        'File cache delete',
+        array('key' => $key, 'file' => $file)
+    );
   }
-  
+
   private function getWriteableCacheFile($file)
   {
     return $this->getCacheFile($file, true);
@@ -92,39 +145,62 @@ class Google_Cache_File extends Google_Cache_Abstract
   {
     return $this->getCacheDir($file, $forWrite) . '/' . md5($file);
   }
-  
+
   private function getCacheDir($file, $forWrite)
   {
     // use the first 2 characters of the hash as a directory prefix
     // this should prevent slowdowns due to huge directory listings
     // and thus give some basic amount of scalability
-    $storageDir = $this->path . '/' . substr(md5($file), 0, 2);
+    $dirHash = substr(md5($file), 0, 2);
+    // trim the directory separator from the path to prevent double separators
+    $storageDir = rtrim($this->path, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR . $dirHash;
     if ($forWrite && ! is_dir($storageDir)) {
-      if (! mkdir($storageDir, 0755, true)) {
+      if (! mkdir($storageDir, 0700, true)) {
+        $this->log(
+            'error',
+            'File cache creation failed',
+            array('dir' => $storageDir)
+        );
         throw new Google_Cache_Exception("Could not create storage directory: $storageDir");
       }
     }
     return $storageDir;
   }
-  
+
   private function acquireReadLock($storageFile)
   {
     return $this->acquireLock(LOCK_SH, $storageFile);
   }
-  
+
   private function acquireWriteLock($storageFile)
   {
     $rc = $this->acquireLock(LOCK_EX, $storageFile);
     if (!$rc) {
+      $this->log(
+          'notice',
+          'File cache write lock failed',
+          array('file' => $storageFile)
+      );
       $this->delete($storageFile);
     }
     return $rc;
   }
-  
+
   private function acquireLock($type, $storageFile)
   {
     $mode = $type == LOCK_EX ? "w" : "r";
     $this->fh = fopen($storageFile, $mode);
+    if (!$this->fh) {
+      $this->log(
+          'error',
+          'Failed to open file during lock acquisition',
+          array('file' => $storageFile)
+      );
+      return false;
+    }
+    if ($type == LOCK_EX) {
+      chmod($storageFile, 0600);
+    }
     $count = 0;
     while (!flock($this->fh, $type | LOCK_NB)) {
       // Sleep for 10ms.
@@ -135,11 +211,18 @@ class Google_Cache_File extends Google_Cache_Abstract
     }
     return true;
   }
-  
-  public function unlock($storageFile)
+
+  public function unlock()
   {
     if ($this->fh) {
       flock($this->fh, LOCK_UN);
+    }
+  }
+
+  private function log($level, $message, $context = array())
+  {
+    if ($this->logger) {
+      $this->logger->log($level, $message, $context);
     }
   }
 }
