@@ -15,15 +15,24 @@
  * limitations under the License.
  */
 
-require_once 'Google/Client.php';
-require_once 'Google/Http/Request.php';
-require_once 'Google/Http/REST.php';
+use GuzzleHttp\Psr7;
+use GuzzleHttp\Psr7\Request;
+use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
 /**
- * @author Chirag Shah <chirags@google.com>
+ * Class to handle batched requests to the Google API service.
  */
 class Google_Http_Batch
 {
+  const BATCH_PATH = 'batch';
+
+  private static $CONNECTION_ESTABLISHED_HEADERS = array(
+    "HTTP/1.0 200 Connection established\r\n\r\n",
+    "HTTP/1.1 200 Connection established\r\n\r\n",
+  );
+
   /** @var string Multipart Boundary. */
   private $boundary;
 
@@ -33,20 +42,13 @@ class Google_Http_Batch
   /** @var Google_Client */
   private $client;
 
-  private $expected_classes = array();
-
-  private $base_path;
-
-  public function __construct(Google_Client $client, $boundary = false)
+  public function __construct(Google_Client $client)
   {
     $this->client = $client;
-    $this->base_path = $this->client->getBasePath();
-    $this->expected_classes = array();
-    $boundary = (false == $boundary) ? mt_rand() : $boundary;
-    $this->boundary = str_replace('"', '', $boundary);
+    $this->boundary = mt_rand();
   }
 
-  public function add(Google_Http_Request $request, $key = false)
+  public function add(RequestInterface $request, $key = false)
   {
     if (false == $key) {
       $key = mt_rand();
@@ -58,32 +60,71 @@ class Google_Http_Batch
   public function execute()
   {
     $body = '';
+    $classes = array();
+    $batchHttpTemplate = <<<EOF
+--%s
+Content-Type: application/http
+Content-Transfer-Encoding: binary
+MIME-Version: 1.0
+Content-ID: %s
+
+%s%s
+%s
+
+
+EOF;
 
     /** @var Google_Http_Request $req */
-    foreach ($this->requests as $key => $req) {
-      $body .= "--{$this->boundary}\n";
-      $body .= $req->toBatchString($key) . "\n";
-      $this->expected_classes["response-" . $key] = $req->getExpectedClass();
+    foreach ($this->requests as $key => $request) {
+      $firstLine = sprintf(
+          '%s %s HTTP/%s',
+          $request->getMethod(),
+          $request->getRequestTarget(),
+          $request->getProtocolVersion()
+      );
+
+      $content = (string) $request->getBody();
+
+      $headers = '';
+      foreach ($request->getHeaders() as $name => $values) {
+          $headers .= sprintf("%s:%s\r\n", $name, implode(', ', $values));
+      }
+
+      $body .= sprintf(
+          $batchHttpTemplate,
+          $this->boundary,
+          $key,
+          $firstLine,
+          $headers,
+          $content ? "\n".$content : ''
+      );
+
+      $classes['response-' . $key] = $request->getHeaderLine('X-Php-Expected-Class');
     }
 
-    $body = rtrim($body);
-    $body .= "\n--{$this->boundary}--";
-
-    $url = $this->base_path . '/batch';
-    $httpRequest = new Google_Http_Request($url, 'POST');
-    $httpRequest->setRequestHeaders(
-        array('Content-Type' => 'multipart/mixed; boundary=' . $this->boundary)
+    $body .= "--{$this->boundary}--";
+    $body = trim($body);
+    $url = Google_Client::API_BASE_PATH . '/' . self::BATCH_PATH;
+    $headers = array(
+      'Content-Type' => sprintf('multipart/mixed; boundary=%s', $this->boundary),
+      'Content-Length' => strlen($body),
     );
 
-    $httpRequest->setPostBody($body);
-    $response = $this->client->getIo()->makeRequest($httpRequest);
+    $request = new Request(
+        'POST',
+        $url,
+        $headers,
+        $body
+    );
 
-    return $this->parseResponse($response);
+    $response = $this->client->execute($request);
+
+    return $this->parseResponse($response, $classes);
   }
 
-  public function parseResponse(Google_Http_Request $response)
+  public function parseResponse(ResponseInterface $response, $classes = array())
   {
-    $contentType = $response->getResponseHeader('content-type');
+    $contentType = $response->getHeaderLine('content-type');
     $contentType = explode(';', $contentType);
     $boundary = false;
     foreach ($contentType as $part) {
@@ -93,45 +134,46 @@ class Google_Http_Batch
       }
     }
 
-    $body = $response->getResponseBody();
-    if ($body) {
+    $body = (string) $response->getBody();
+    if (!empty($body)) {
       $body = str_replace("--$boundary--", "--$boundary", $body);
       $parts = explode("--$boundary", $body);
       $responses = array();
+      $requests = array_values($this->requests);
 
-      foreach ($parts as $part) {
+      foreach ($parts as $i => $part) {
         $part = trim($part);
         if (!empty($part)) {
-          list($metaHeaders, $part) = explode("\r\n\r\n", $part, 2);
-          $metaHeaders = $this->client->getIo()->getHttpResponseHeaders($metaHeaders);
+          list($rawHeaders, $part) = explode("\r\n\r\n", $part, 2);
+          $headers = $this->parseRawHeaders($rawHeaders);
 
           $status = substr($part, 0, strpos($part, "\n"));
           $status = explode(" ", $status);
           $status = $status[1];
 
-          list($partHeaders, $partBody) = $this->client->getIo()->ParseHttpResponse($part, false);
-          $response = new Google_Http_Request("");
-          $response->setResponseHttpCode($status);
-          $response->setResponseHeaders($partHeaders);
-          $response->setResponseBody($partBody);
+          list($partHeaders, $partBody) = $this->parseHttpResponse($part, false);
+          $response = new Response(
+              $status,
+              $partHeaders,
+              Psr7\stream_for($partBody)
+          );
 
           // Need content id.
-          $key = $metaHeaders['content-id'];
-
-          if (isset($this->expected_classes[$key]) &&
-              strlen($this->expected_classes[$key]) > 0) {
+          $key = $headers['content-id'];
+          $class = '';
+          if (!empty($this->expected_classes[$key])) {
             $class = $this->expected_classes[$key];
-            $response->setExpectedClass($class);
           }
 
           try {
-            $response = Google_Http_REST::decodeHttpResponse($response);
-            $responses[$key] = $response;
+            $response = Google_Http_REST::decodeHttpResponse($response, $requests[$i-1]);
           } catch (Google_Service_Exception $e) {
-            // Store the exception as the response, so succesful responses
+            // Store the exception as the response, so successful responses
             // can be processed.
-            $responses[$key] = $e;
+            $response = $e;
           }
+
+          $responses[$key] = $response;
         }
       }
 
@@ -139,5 +181,63 @@ class Google_Http_Batch
     }
 
     return null;
+  }
+
+  private function parseRawHeaders($rawHeaders)
+  {
+    $headers = array();
+    $responseHeaderLines = explode("\r\n", $rawHeaders);
+    foreach ($responseHeaderLines as $headerLine) {
+      if ($headerLine && strpos($headerLine, ':') !== false) {
+        list($header, $value) = explode(': ', $headerLine, 2);
+        $header = strtolower($header);
+        if (isset($headers[$header])) {
+          $headers[$header] .= "\n" . $value;
+        } else {
+          $headers[$header] = $value;
+        }
+      }
+    }
+    return $headers;
+  }
+
+  /**
+   * Used by the IO lib and also the batch processing.
+   *
+   * @param $respData
+   * @param $headerSize
+   * @return array
+   */
+  private function parseHttpResponse($respData, $headerSize)
+  {
+    // check proxy header
+    foreach (self::$CONNECTION_ESTABLISHED_HEADERS as $established_header) {
+      if (stripos($respData, $established_header) !== false) {
+        // existed, remove it
+        $respData = str_ireplace($established_header, '', $respData);
+        // Subtract the proxy header size unless the cURL bug prior to 7.30.0
+        // is present which prevented the proxy header size from being taken into
+        // account.
+        // @TODO look into this
+        // if (!$this->needsQuirk()) {
+        //   $headerSize -= strlen($established_header);
+        // }
+        break;
+      }
+    }
+
+    if ($headerSize) {
+      $responseBody = substr($respData, $headerSize);
+      $responseHeaders = substr($respData, 0, $headerSize);
+    } else {
+      $responseSegments = explode("\r\n\r\n", $respData, 2);
+      $responseHeaders = $responseSegments[0];
+      $responseBody = isset($responseSegments[1]) ? $responseSegments[1] :
+                                                    null;
+    }
+
+    $responseHeaders = $this->parseRawHeaders($responseHeaders);
+
+    return array($responseHeaders, $responseBody);
   }
 }
