@@ -20,11 +20,15 @@
  * from a given json schema.
  * http://tools.ietf.org/html/draft-zyp-json-schema-03#section-5
  *
- * @author Chirag Shah <chirags@google.com>
- *
  */
 class Google_Model implements ArrayAccess
 {
+  /**
+   * If you need to specify a NULL JSON value, use Google_Model::NULL_VALUE
+   * instead - it will be replaced when converting to JSON with a real null.
+   */
+  const NULL_VALUE = "{}gapi-php-null";
+  protected $internal_gapi_mappings = array();
   protected $modelData = array();
   protected $processed = array();
 
@@ -32,15 +36,21 @@ class Google_Model implements ArrayAccess
    * Polymorphic - accepts a variable number of arguments dependent
    * on the type of the model subclass.
    */
-  public function __construct()
+  final public function __construct()
   {
     if (func_num_args() == 1 && is_array(func_get_arg(0))) {
       // Initialize the model with the array's contents.
       $array = func_get_arg(0);
       $this->mapTypes($array);
     }
+    $this->gapiInit();
   }
 
+  /**
+   * Getter that handles passthrough access to the data array, and lazy object creation.
+   * @param string $key Property name.
+   * @return mixed The value if any, or null.
+   */
   public function __get($key)
   {
     $keyTypeName = $this->keyType($key);
@@ -75,7 +85,7 @@ class Google_Model implements ArrayAccess
       $this->processed[$key] = true;
     }
 
-    return $this->modelData[$key];
+    return isset($this->modelData[$key]) ? $this->modelData[$key] : null;
   }
 
   /**
@@ -86,19 +96,29 @@ class Google_Model implements ArrayAccess
    */
   protected function mapTypes($array)
   {
-    // Hard initilise simple types, lazy load more complex ones.
+    // Hard initialise simple types, lazy load more complex ones.
     foreach ($array as $key => $val) {
       if ( !property_exists($this, $this->keyType($key)) &&
         property_exists($this, $key)) {
           $this->$key = $val;
           unset($array[$key]);
-      } elseif (property_exists($this, $camelKey = Google_Utils::camelCase($key))) {
+      } elseif (property_exists($this, $camelKey = $this->camelCase($key))) {
           // This checks if property exists as camelCase, leaving it in array as snake_case
           // in case of backwards compatibility issues.
           $this->$camelKey = $val;
       }
     }
     $this->modelData = $array;
+  }
+
+  /**
+   * Blank initialiser to be used in subclasses to do  post-construction initialisation - this
+   * avoids the need for subclasses to have to implement the variadics handling in their
+   * constructors.
+   */
+  protected function gapiInit()
+  {
+    return;
   }
 
   /**
@@ -115,7 +135,7 @@ class Google_Model implements ArrayAccess
     foreach ($this->modelData as $key => $val) {
       $result = $this->getSimpleValue($val);
       if ($result !== null) {
-        $object->$key = $result;
+        $object->$key = $this->nullPlaceholderCheck($result);
       }
     }
 
@@ -126,7 +146,8 @@ class Google_Model implements ArrayAccess
       $name = $member->getName();
       $result = $this->getSimpleValue($this->$name);
       if ($result !== null) {
-        $object->$name = $result;
+        $name = $this->getMappedName($name);
+        $object->$name = $this->nullPlaceholderCheck($result);
       }
     }
 
@@ -146,12 +167,36 @@ class Google_Model implements ArrayAccess
       foreach ($value as $key => $a_value) {
         $a_value = $this->getSimpleValue($a_value);
         if ($a_value !== null) {
-          $return[$key] = $a_value;
+          $key = $this->getMappedName($key);
+          $return[$key] = $this->nullPlaceholderCheck($a_value);
         }
       }
       return $return;
     }
     return $value;
+  }
+
+  /**
+   * Check whether the value is the null placeholder and return true null.
+   */
+  private function nullPlaceholderCheck($value)
+  {
+    if ($value === self::NULL_VALUE) {
+      return null;
+    }
+    return $value;
+  }
+
+  /**
+   * If there is an internal name mapping, use that.
+   */
+  private function getMappedName($key)
+  {
+    if (isset($this->internal_gapi_mappings) &&
+        isset($this->internal_gapi_mappings[$key])) {
+      $key = $this->internal_gapi_mappings[$key];
+    }
+    return $key;
   }
 
   /**
@@ -246,5 +291,18 @@ class Google_Model implements ArrayAccess
   public function __unset($key)
   {
     unset($this->modelData[$key]);
+  }
+
+  /**
+   * Convert a string to camelCase
+   * @param  string $value
+   * @return string
+   */
+  private function camelCase($value)
+  {
+    $value = ucwords(str_replace(array('-', '_'), ' ', $value));
+    $value = str_replace(' ', '', $value);
+    $value[0] = strtolower($value[0]);
+    return $value;
   }
 }
