@@ -19,8 +19,8 @@
  * Service definition for AdSense (v1.4).
  *
  * <p>
- * Gives AdSense publishers access to their inventory and the ability to generate reports
- * </p>
+ * Gives AdSense publishers access to their inventory and the ability to
+ * generate reports</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,9 +32,11 @@
 class Google_Service_AdSense extends Google_Service
 {
   /** View and manage your AdSense data. */
-  const ADSENSE = "https://www.googleapis.com/auth/adsense";
+  const ADSENSE =
+      "https://www.googleapis.com/auth/adsense";
   /** View your AdSense data. */
-  const ADSENSE_READONLY = "https://www.googleapis.com/auth/adsense.readonly";
+  const ADSENSE_READONLY =
+      "https://www.googleapis.com/auth/adsense.readonly";
 
   public $accounts;
   public $accounts_adclients;
@@ -71,6 +73,7 @@ class Google_Service_AdSense extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'adsense/v1.4/';
     $this->version = 'v1.4';
     $this->serviceName = 'adsense';
@@ -1073,12 +1076,10 @@ class Google_Service_AdSense_Accounts_Resource extends Google_Service_Resource
   /**
    * Get information about the selected AdSense account. (accounts.get)
    *
-   * @param string $accountId
-   * Account to get information about.
+   * @param string $accountId Account to get information about.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool tree
-   * Whether the tree of sub accounts should be returned.
+   * @opt_param bool tree Whether the tree of sub accounts should be returned.
    * @return Google_Service_AdSense_Account
    */
   public function get($accountId, $optParams = array())
@@ -1087,16 +1088,17 @@ class Google_Service_AdSense_Accounts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_Account");
   }
+
   /**
    * List all accounts available to this AdSense account. (accounts.listAccounts)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through accounts. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of accounts to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through
+   * accounts. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of accounts to include in the
+   * response, used for paging.
    * @return Google_Service_AdSense_Accounts
    */
   public function listAccounts($optParams = array())
@@ -1122,15 +1124,14 @@ class Google_Service_AdSense_AccountsAdclients_Resource extends Google_Service_R
    * List all ad clients in the specified account.
    * (adclients.listAccountsAdclients)
    *
-   * @param string $accountId
-   * Account for which to list ad clients.
+   * @param string $accountId Account for which to list ad clients.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad clients. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of ad clients to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * clients. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of ad clients to include in the
+   * response, used for paging.
    * @return Google_Service_AdSense_AdClients
    */
   public function listAccountsAdclients($accountId, $optParams = array())
@@ -1155,12 +1156,9 @@ class Google_Service_AdSense_AccountsAdunits_Resource extends Google_Service_Res
    * Gets the specified ad unit in the specified ad client for the specified
    * account. (adunits.get)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client for which to get the ad unit.
-   * @param string $adUnitId
-   * Ad unit to retrieve.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client for which to get the ad unit.
+   * @param string $adUnitId Ad unit to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_AdUnit
    */
@@ -1170,15 +1168,13 @@ class Google_Service_AdSense_AccountsAdunits_Resource extends Google_Service_Res
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_AdUnit");
   }
+
   /**
    * Get ad code for the specified ad unit. (adunits.getAdCode)
    *
-   * @param string $accountId
-   * Account which contains the ad client.
-   * @param string $adClientId
-   * Ad client with contains the ad unit.
-   * @param string $adUnitId
-   * Ad unit to get the code for.
+   * @param string $accountId Account which contains the ad client.
+   * @param string $adClientId Ad client with contains the ad unit.
+   * @param string $adUnitId Ad unit to get the code for.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_AdCode
    */
@@ -1188,23 +1184,22 @@ class Google_Service_AdSense_AccountsAdunits_Resource extends Google_Service_Res
     $params = array_merge($params, $optParams);
     return $this->call('getAdCode', array($params), "Google_Service_AdSense_AdCode");
   }
+
   /**
    * List all ad units in the specified ad client for the specified account.
    * (adunits.listAccountsAdunits)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client for which to list ad units.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client for which to list ad units.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool includeInactive
-   * Whether to include inactive ad units. Default: true.
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad units. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of ad units to include in the response, used for paging.
+   * @opt_param bool includeInactive Whether to include inactive ad units.
+   * Default: true.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * units. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of ad units to include in the
+   * response, used for paging.
    * @return Google_Service_AdSense_AdUnits
    */
   public function listAccountsAdunits($accountId, $adClientId, $optParams = array())
@@ -1230,19 +1225,16 @@ class Google_Service_AdSense_AccountsAdunitsCustomchannels_Resource extends Goog
    * List all custom channels which the specified ad unit belongs to.
    * (customchannels.listAccountsAdunitsCustomchannels)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client which contains the ad unit.
-   * @param string $adUnitId
-   * Ad unit for which to list custom channels.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client which contains the ad unit.
+   * @param string $adUnitId Ad unit for which to list custom channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through custom channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of custom channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through custom
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of custom channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_CustomChannels
    */
   public function listAccountsAdunitsCustomchannels($accountId, $adClientId, $adUnitId, $optParams = array())
@@ -1267,10 +1259,8 @@ class Google_Service_AdSense_AccountsAlerts_Resource extends Google_Service_Reso
    * Dismiss (delete) the specified alert from the specified publisher AdSense
    * account. (alerts.delete)
    *
-   * @param string $accountId
-   * Account which contains the ad unit.
-   * @param string $alertId
-   * Alert to delete.
+   * @param string $accountId Account which contains the ad unit.
+   * @param string $alertId Alert to delete.
    * @param array $optParams Optional parameters.
    */
   public function delete($accountId, $alertId, $optParams = array())
@@ -1279,18 +1269,17 @@ class Google_Service_AdSense_AccountsAlerts_Resource extends Google_Service_Reso
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * List the alerts for the specified AdSense account.
    * (alerts.listAccountsAlerts)
    *
-   * @param string $accountId
-   * Account for which to retrieve the alerts.
+   * @param string $accountId Account for which to retrieve the alerts.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string locale
-   * The locale to use for translating alert messages. The account locale will be used if this is not
-    * supplied. The AdSense default (English) will be used if the supplied locale is invalid or
-    * unsupported.
+   * @opt_param string locale The locale to use for translating alert messages.
+   * The account locale will be used if this is not supplied. The AdSense default
+   * (English) will be used if the supplied locale is invalid or unsupported.
    * @return Google_Service_AdSense_Alerts
    */
   public function listAccountsAlerts($accountId, $optParams = array())
@@ -1315,12 +1304,9 @@ class Google_Service_AdSense_AccountsCustomchannels_Resource extends Google_Serv
    * Get the specified custom channel from the specified ad client for the
    * specified account. (customchannels.get)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client which contains the custom channel.
-   * @param string $customChannelId
-   * Custom channel to retrieve.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client which contains the custom channel.
+   * @param string $customChannelId Custom channel to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_CustomChannel
    */
@@ -1330,21 +1316,20 @@ class Google_Service_AdSense_AccountsCustomchannels_Resource extends Google_Serv
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_CustomChannel");
   }
+
   /**
    * List all custom channels in the specified ad client for the specified
    * account. (customchannels.listAccountsCustomchannels)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client for which to list custom channels.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client for which to list custom channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through custom channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of custom channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through custom
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of custom channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_CustomChannels
    */
   public function listAccountsCustomchannels($accountId, $adClientId, $optParams = array())
@@ -1370,21 +1355,18 @@ class Google_Service_AdSense_AccountsCustomchannelsAdunits_Resource extends Goog
    * List all ad units in the specified custom channel.
    * (adunits.listAccountsCustomchannelsAdunits)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client which contains the custom channel.
-   * @param string $customChannelId
-   * Custom channel for which to list ad units.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client which contains the custom channel.
+   * @param string $customChannelId Custom channel for which to list ad units.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool includeInactive
-   * Whether to include inactive ad units. Default: true.
-   * @opt_param int maxResults
-   * The maximum number of ad units to include in the response, used for paging.
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad units. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param bool includeInactive Whether to include inactive ad units.
+   * Default: true.
+   * @opt_param int maxResults The maximum number of ad units to include in the
+   * response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * units. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
    * @return Google_Service_AdSense_AdUnits
    */
   public function listAccountsCustomchannelsAdunits($accountId, $adClientId, $customChannelId, $optParams = array())
@@ -1409,8 +1391,7 @@ class Google_Service_AdSense_AccountsPayments_Resource extends Google_Service_Re
    * List the payments for the specified AdSense account.
    * (payments.listAccountsPayments)
    *
-   * @param string $accountId
-   * Account for which to retrieve the payments.
+   * @param string $accountId Account for which to retrieve the payments.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_Payments
    */
@@ -1437,37 +1418,29 @@ class Google_Service_AdSense_AccountsReports_Resource extends Google_Service_Res
    * parameters. Returns the result as JSON; to retrieve output in CSV format
    * specify "alt=csv" as a query parameter. (reports.generate)
    *
-   * @param string $accountId
-   * Account upon which to report.
-   * @param string $startDate
-   * Start of the date range to report on in "YYYY-MM-DD" format, inclusive.
-   * @param string $endDate
-   * End of the date range to report on in "YYYY-MM-DD" format, inclusive.
+   * @param string $accountId Account upon which to report.
+   * @param string $startDate Start of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
+   * @param string $endDate End of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string sort
-   * The name of a dimension or metric to sort the resulting report on, optionally prefixed with "+"
-    * to sort ascending or "-" to sort descending. If no prefix is specified, the column is sorted
-    * ascending.
-   * @opt_param string locale
-   * Optional locale to use for translating report output to a local language. Defaults to "en_US" if
-    * not specified.
-   * @opt_param string metric
-   * Numeric columns to include in the report.
-   * @opt_param int maxResults
-   * The maximum number of rows of report data to return.
-   * @opt_param string filter
-   * Filters to be run on the report.
-   * @opt_param string currency
-   * Optional currency to use when reporting on monetary metrics. Defaults to the account's currency
-    * if not set.
-   * @opt_param int startIndex
-   * Index of the first row of report data to return.
-   * @opt_param bool useTimezoneReporting
-   * Whether the report should be generated in the AdSense account's local timezone. If false default
-    * PST/PDT timezone will be used.
-   * @opt_param string dimension
-   * Dimensions to base the report on.
+   * @opt_param string sort The name of a dimension or metric to sort the
+   * resulting report on, optionally prefixed with "+" to sort ascending or "-" to
+   * sort descending. If no prefix is specified, the column is sorted ascending.
+   * @opt_param string locale Optional locale to use for translating report output
+   * to a local language. Defaults to "en_US" if not specified.
+   * @opt_param string metric Numeric columns to include in the report.
+   * @opt_param int maxResults The maximum number of rows of report data to
+   * return.
+   * @opt_param string filter Filters to be run on the report.
+   * @opt_param string currency Optional currency to use when reporting on
+   * monetary metrics. Defaults to the account's currency if not set.
+   * @opt_param int startIndex Index of the first row of report data to return.
+   * @opt_param bool useTimezoneReporting Whether the report should be generated
+   * in the AdSense account's local timezone. If false default PST/PDT timezone
+   * will be used.
+   * @opt_param string dimension Dimensions to base the report on.
    * @return Google_Service_AdSense_AdsenseReportsGenerateResponse
    */
   public function generate($accountId, $startDate, $endDate, $optParams = array())
@@ -1493,19 +1466,15 @@ class Google_Service_AdSense_AccountsReportsSaved_Resource extends Google_Servic
    * Generate an AdSense report based on the saved report ID sent in the query
    * parameters. (saved.generate)
    *
-   * @param string $accountId
-   * Account to which the saved reports belong.
-   * @param string $savedReportId
-   * The saved report to retrieve.
+   * @param string $accountId Account to which the saved reports belong.
+   * @param string $savedReportId The saved report to retrieve.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string locale
-   * Optional locale to use for translating report output to a local language. Defaults to "en_US" if
-    * not specified.
-   * @opt_param int startIndex
-   * Index of the first row of report data to return.
-   * @opt_param int maxResults
-   * The maximum number of rows of report data to return.
+   * @opt_param string locale Optional locale to use for translating report output
+   * to a local language. Defaults to "en_US" if not specified.
+   * @opt_param int startIndex Index of the first row of report data to return.
+   * @opt_param int maxResults The maximum number of rows of report data to
+   * return.
    * @return Google_Service_AdSense_AdsenseReportsGenerateResponse
    */
   public function generate($accountId, $savedReportId, $optParams = array())
@@ -1514,19 +1483,19 @@ class Google_Service_AdSense_AccountsReportsSaved_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('generate', array($params), "Google_Service_AdSense_AdsenseReportsGenerateResponse");
   }
+
   /**
    * List all saved reports in the specified AdSense account.
    * (saved.listAccountsReportsSaved)
    *
-   * @param string $accountId
-   * Account to which the saved reports belong.
+   * @param string $accountId Account to which the saved reports belong.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through saved reports. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of saved reports to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through saved
+   * reports. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of saved reports to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_SavedReports
    */
   public function listAccountsReportsSaved($accountId, $optParams = array())
@@ -1550,10 +1519,8 @@ class Google_Service_AdSense_AccountsSavedadstyles_Resource extends Google_Servi
   /**
    * List a specific saved ad style for the specified account. (savedadstyles.get)
    *
-   * @param string $accountId
-   * Account for which to get the saved ad style.
-   * @param string $savedAdStyleId
-   * Saved ad style to retrieve.
+   * @param string $accountId Account for which to get the saved ad style.
+   * @param string $savedAdStyleId Saved ad style to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_SavedAdStyle
    */
@@ -1563,19 +1530,19 @@ class Google_Service_AdSense_AccountsSavedadstyles_Resource extends Google_Servi
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_SavedAdStyle");
   }
+
   /**
    * List all saved ad styles in the specified account.
    * (savedadstyles.listAccountsSavedadstyles)
    *
-   * @param string $accountId
-   * Account for which to list saved ad styles.
+   * @param string $accountId Account for which to list saved ad styles.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through saved ad styles. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of saved ad styles to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through saved
+   * ad styles. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of saved ad styles to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_SavedAdStyles
    */
   public function listAccountsSavedadstyles($accountId, $optParams = array())
@@ -1600,17 +1567,15 @@ class Google_Service_AdSense_AccountsUrlchannels_Resource extends Google_Service
    * List all URL channels in the specified ad client for the specified account.
    * (urlchannels.listAccountsUrlchannels)
    *
-   * @param string $accountId
-   * Account to which the ad client belongs.
-   * @param string $adClientId
-   * Ad client for which to list URL channels.
+   * @param string $accountId Account to which the ad client belongs.
+   * @param string $adClientId Ad client for which to list URL channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through URL channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of URL channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through URL
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of URL channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_UrlChannels
    */
   public function listAccountsUrlchannels($accountId, $adClientId, $optParams = array())
@@ -1637,11 +1602,11 @@ class Google_Service_AdSense_Adclients_Resource extends Google_Service_Resource
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad clients. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of ad clients to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * clients. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of ad clients to include in the
+   * response, used for paging.
    * @return Google_Service_AdSense_AdClients
    */
   public function listAdclients($optParams = array())
@@ -1666,10 +1631,8 @@ class Google_Service_AdSense_Adunits_Resource extends Google_Service_Resource
   /**
    * Gets the specified ad unit in the specified ad client. (adunits.get)
    *
-   * @param string $adClientId
-   * Ad client for which to get the ad unit.
-   * @param string $adUnitId
-   * Ad unit to retrieve.
+   * @param string $adClientId Ad client for which to get the ad unit.
+   * @param string $adUnitId Ad unit to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_AdUnit
    */
@@ -1679,13 +1642,12 @@ class Google_Service_AdSense_Adunits_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_AdUnit");
   }
+
   /**
    * Get ad code for the specified ad unit. (adunits.getAdCode)
    *
-   * @param string $adClientId
-   * Ad client with contains the ad unit.
-   * @param string $adUnitId
-   * Ad unit to get the code for.
+   * @param string $adClientId Ad client with contains the ad unit.
+   * @param string $adUnitId Ad unit to get the code for.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_AdCode
    */
@@ -1695,21 +1657,21 @@ class Google_Service_AdSense_Adunits_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('getAdCode', array($params), "Google_Service_AdSense_AdCode");
   }
+
   /**
    * List all ad units in the specified ad client for this AdSense account.
    * (adunits.listAdunits)
    *
-   * @param string $adClientId
-   * Ad client for which to list ad units.
+   * @param string $adClientId Ad client for which to list ad units.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool includeInactive
-   * Whether to include inactive ad units. Default: true.
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad units. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of ad units to include in the response, used for paging.
+   * @opt_param bool includeInactive Whether to include inactive ad units.
+   * Default: true.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * units. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of ad units to include in the
+   * response, used for paging.
    * @return Google_Service_AdSense_AdUnits
    */
   public function listAdunits($adClientId, $optParams = array())
@@ -1735,17 +1697,15 @@ class Google_Service_AdSense_AdunitsCustomchannels_Resource extends Google_Servi
    * List all custom channels which the specified ad unit belongs to.
    * (customchannels.listAdunitsCustomchannels)
    *
-   * @param string $adClientId
-   * Ad client which contains the ad unit.
-   * @param string $adUnitId
-   * Ad unit for which to list custom channels.
+   * @param string $adClientId Ad client which contains the ad unit.
+   * @param string $adUnitId Ad unit for which to list custom channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through custom channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of custom channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through custom
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of custom channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_CustomChannels
    */
   public function listAdunitsCustomchannels($adClientId, $adUnitId, $optParams = array())
@@ -1771,8 +1731,7 @@ class Google_Service_AdSense_Alerts_Resource extends Google_Service_Resource
    * Dismiss (delete) the specified alert from the publisher's AdSense account.
    * (alerts.delete)
    *
-   * @param string $alertId
-   * Alert to delete.
+   * @param string $alertId Alert to delete.
    * @param array $optParams Optional parameters.
    */
   public function delete($alertId, $optParams = array())
@@ -1781,15 +1740,15 @@ class Google_Service_AdSense_Alerts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * List the alerts for this AdSense account. (alerts.listAlerts)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string locale
-   * The locale to use for translating alert messages. The account locale will be used if this is not
-    * supplied. The AdSense default (English) will be used if the supplied locale is invalid or
-    * unsupported.
+   * @opt_param string locale The locale to use for translating alert messages.
+   * The account locale will be used if this is not supplied. The AdSense default
+   * (English) will be used if the supplied locale is invalid or unsupported.
    * @return Google_Service_AdSense_Alerts
    */
   public function listAlerts($optParams = array())
@@ -1815,10 +1774,8 @@ class Google_Service_AdSense_Customchannels_Resource extends Google_Service_Reso
    * Get the specified custom channel from the specified ad client.
    * (customchannels.get)
    *
-   * @param string $adClientId
-   * Ad client which contains the custom channel.
-   * @param string $customChannelId
-   * Custom channel to retrieve.
+   * @param string $adClientId Ad client which contains the custom channel.
+   * @param string $customChannelId Custom channel to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_CustomChannel
    */
@@ -1828,19 +1785,19 @@ class Google_Service_AdSense_Customchannels_Resource extends Google_Service_Reso
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_CustomChannel");
   }
+
   /**
    * List all custom channels in the specified ad client for this AdSense account.
    * (customchannels.listCustomchannels)
    *
-   * @param string $adClientId
-   * Ad client for which to list custom channels.
+   * @param string $adClientId Ad client for which to list custom channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through custom channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of custom channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through custom
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of custom channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_CustomChannels
    */
   public function listCustomchannels($adClientId, $optParams = array())
@@ -1866,19 +1823,17 @@ class Google_Service_AdSense_CustomchannelsAdunits_Resource extends Google_Servi
    * List all ad units in the specified custom channel.
    * (adunits.listCustomchannelsAdunits)
    *
-   * @param string $adClientId
-   * Ad client which contains the custom channel.
-   * @param string $customChannelId
-   * Custom channel for which to list ad units.
+   * @param string $adClientId Ad client which contains the custom channel.
+   * @param string $customChannelId Custom channel for which to list ad units.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool includeInactive
-   * Whether to include inactive ad units. Default: true.
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad units. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of ad units to include in the response, used for paging.
+   * @opt_param bool includeInactive Whether to include inactive ad units.
+   * Default: true.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * units. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of ad units to include in the
+   * response, used for paging.
    * @return Google_Service_AdSense_AdUnits
    */
   public function listCustomchannelsAdunits($adClientId, $customChannelId, $optParams = array())
@@ -1899,7 +1854,6 @@ class Google_Service_AdSense_CustomchannelsAdunits_Resource extends Google_Servi
  */
 class Google_Service_AdSense_Metadata_Resource extends Google_Service_Resource
 {
-
 }
 
 /**
@@ -1994,37 +1948,29 @@ class Google_Service_AdSense_Reports_Resource extends Google_Service_Resource
    * parameters. Returns the result as JSON; to retrieve output in CSV format
    * specify "alt=csv" as a query parameter. (reports.generate)
    *
-   * @param string $startDate
-   * Start of the date range to report on in "YYYY-MM-DD" format, inclusive.
-   * @param string $endDate
-   * End of the date range to report on in "YYYY-MM-DD" format, inclusive.
+   * @param string $startDate Start of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
+   * @param string $endDate End of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string sort
-   * The name of a dimension or metric to sort the resulting report on, optionally prefixed with "+"
-    * to sort ascending or "-" to sort descending. If no prefix is specified, the column is sorted
-    * ascending.
-   * @opt_param string locale
-   * Optional locale to use for translating report output to a local language. Defaults to "en_US" if
-    * not specified.
-   * @opt_param string metric
-   * Numeric columns to include in the report.
-   * @opt_param int maxResults
-   * The maximum number of rows of report data to return.
-   * @opt_param string filter
-   * Filters to be run on the report.
-   * @opt_param string currency
-   * Optional currency to use when reporting on monetary metrics. Defaults to the account's currency
-    * if not set.
-   * @opt_param int startIndex
-   * Index of the first row of report data to return.
-   * @opt_param bool useTimezoneReporting
-   * Whether the report should be generated in the AdSense account's local timezone. If false default
-    * PST/PDT timezone will be used.
-   * @opt_param string dimension
-   * Dimensions to base the report on.
-   * @opt_param string accountId
-   * Accounts upon which to report.
+   * @opt_param string sort The name of a dimension or metric to sort the
+   * resulting report on, optionally prefixed with "+" to sort ascending or "-" to
+   * sort descending. If no prefix is specified, the column is sorted ascending.
+   * @opt_param string locale Optional locale to use for translating report output
+   * to a local language. Defaults to "en_US" if not specified.
+   * @opt_param string metric Numeric columns to include in the report.
+   * @opt_param int maxResults The maximum number of rows of report data to
+   * return.
+   * @opt_param string filter Filters to be run on the report.
+   * @opt_param string currency Optional currency to use when reporting on
+   * monetary metrics. Defaults to the account's currency if not set.
+   * @opt_param int startIndex Index of the first row of report data to return.
+   * @opt_param bool useTimezoneReporting Whether the report should be generated
+   * in the AdSense account's local timezone. If false default PST/PDT timezone
+   * will be used.
+   * @opt_param string dimension Dimensions to base the report on.
+   * @opt_param string accountId Accounts upon which to report.
    * @return Google_Service_AdSense_AdsenseReportsGenerateResponse
    */
   public function generate($startDate, $endDate, $optParams = array())
@@ -2050,17 +1996,14 @@ class Google_Service_AdSense_ReportsSaved_Resource extends Google_Service_Resour
    * Generate an AdSense report based on the saved report ID sent in the query
    * parameters. (saved.generate)
    *
-   * @param string $savedReportId
-   * The saved report to retrieve.
+   * @param string $savedReportId The saved report to retrieve.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string locale
-   * Optional locale to use for translating report output to a local language. Defaults to "en_US" if
-    * not specified.
-   * @opt_param int startIndex
-   * Index of the first row of report data to return.
-   * @opt_param int maxResults
-   * The maximum number of rows of report data to return.
+   * @opt_param string locale Optional locale to use for translating report output
+   * to a local language. Defaults to "en_US" if not specified.
+   * @opt_param int startIndex Index of the first row of report data to return.
+   * @opt_param int maxResults The maximum number of rows of report data to
+   * return.
    * @return Google_Service_AdSense_AdsenseReportsGenerateResponse
    */
   public function generate($savedReportId, $optParams = array())
@@ -2069,16 +2012,17 @@ class Google_Service_AdSense_ReportsSaved_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('generate', array($params), "Google_Service_AdSense_AdsenseReportsGenerateResponse");
   }
+
   /**
    * List all saved reports in this AdSense account. (saved.listReportsSaved)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through saved reports. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of saved reports to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through saved
+   * reports. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of saved reports to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_SavedReports
    */
   public function listReportsSaved($optParams = array())
@@ -2103,8 +2047,7 @@ class Google_Service_AdSense_Savedadstyles_Resource extends Google_Service_Resou
   /**
    * Get a specific saved ad style from the user's account. (savedadstyles.get)
    *
-   * @param string $savedAdStyleId
-   * Saved ad style to retrieve.
+   * @param string $savedAdStyleId Saved ad style to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSense_SavedAdStyle
    */
@@ -2114,17 +2057,18 @@ class Google_Service_AdSense_Savedadstyles_Resource extends Google_Service_Resou
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSense_SavedAdStyle");
   }
+
   /**
    * List all saved ad styles in the user's account.
    * (savedadstyles.listSavedadstyles)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through saved ad styles. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of saved ad styles to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through saved
+   * ad styles. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of saved ad styles to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_SavedAdStyles
    */
   public function listSavedadstyles($optParams = array())
@@ -2150,15 +2094,14 @@ class Google_Service_AdSense_Urlchannels_Resource extends Google_Service_Resourc
    * List all URL channels in the specified ad client for this AdSense account.
    * (urlchannels.listUrlchannels)
    *
-   * @param string $adClientId
-   * Ad client for which to list URL channels.
+   * @param string $adClientId Ad client for which to list URL channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through URL channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param int maxResults
-   * The maximum number of URL channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through URL
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param int maxResults The maximum number of URL channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSense_UrlChannels
    */
   public function listUrlchannels($adClientId, $optParams = array())
@@ -2174,6 +2117,11 @@ class Google_Service_AdSense_Urlchannels_Resource extends Google_Service_Resourc
 
 class Google_Service_AdSense_Account extends Google_Collection
 {
+  protected $collection_key = 'subAccounts';
+  protected $internal_gapi_mappings = array(
+        "creationTime" => "creation_time",
+  );
+  public $creationTime;
   public $id;
   public $kind;
   public $name;
@@ -2182,61 +2130,59 @@ class Google_Service_AdSense_Account extends Google_Collection
   protected $subAccountsDataType = 'array';
   public $timezone;
 
+
+  public function setCreationTime($creationTime)
+  {
+    $this->creationTime = $creationTime;
+  }
+  public function getCreationTime()
+  {
+    return $this->creationTime;
+  }
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setPremium($premium)
   {
     $this->premium = $premium;
   }
-
   public function getPremium()
   {
     return $this->premium;
   }
-
   public function setSubAccounts($subAccounts)
   {
     $this->subAccounts = $subAccounts;
   }
-
   public function getSubAccounts()
   {
     return $this->subAccounts;
   }
-
   public function setTimezone($timezone)
   {
     $this->timezone = $timezone;
   }
-
   public function getTimezone()
   {
     return $this->timezone;
@@ -2245,47 +2191,44 @@ class Google_Service_AdSense_Account extends Google_Collection
 
 class Google_Service_AdSense_Accounts extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_Account';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -2294,6 +2237,8 @@ class Google_Service_AdSense_Accounts extends Google_Collection
 
 class Google_Service_AdSense_AdClient extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $arcOptIn;
   public $arcReviewMode;
   public $id;
@@ -2301,61 +2246,51 @@ class Google_Service_AdSense_AdClient extends Google_Model
   public $productCode;
   public $supportsReporting;
 
+
   public function setArcOptIn($arcOptIn)
   {
     $this->arcOptIn = $arcOptIn;
   }
-
   public function getArcOptIn()
   {
     return $this->arcOptIn;
   }
-
   public function setArcReviewMode($arcReviewMode)
   {
     $this->arcReviewMode = $arcReviewMode;
   }
-
   public function getArcReviewMode()
   {
     return $this->arcReviewMode;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setProductCode($productCode)
   {
     $this->productCode = $productCode;
   }
-
   public function getProductCode()
   {
     return $this->productCode;
   }
-
   public function setSupportsReporting($supportsReporting)
   {
     $this->supportsReporting = $supportsReporting;
   }
-
   public function getSupportsReporting()
   {
     return $this->supportsReporting;
@@ -2364,47 +2299,44 @@ class Google_Service_AdSense_AdClient extends Google_Model
 
 class Google_Service_AdSense_AdClients extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_AdClient';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -2413,24 +2345,24 @@ class Google_Service_AdSense_AdClients extends Google_Collection
 
 class Google_Service_AdSense_AdCode extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $adCode;
   public $kind;
+
 
   public function setAdCode($adCode)
   {
     $this->adCode = $adCode;
   }
-
   public function getAdCode()
   {
     return $this->adCode;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -2439,6 +2371,8 @@ class Google_Service_AdSense_AdCode extends Google_Model
 
 class Google_Service_AdSense_AdStyle extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $colorsType = 'Google_Service_AdSense_AdStyleColors';
   protected $colorsDataType = '';
   public $corners;
@@ -2446,41 +2380,35 @@ class Google_Service_AdSense_AdStyle extends Google_Model
   protected $fontDataType = '';
   public $kind;
 
+
   public function setColors(Google_Service_AdSense_AdStyleColors $colors)
   {
     $this->colors = $colors;
   }
-
   public function getColors()
   {
     return $this->colors;
   }
-
   public function setCorners($corners)
   {
     $this->corners = $corners;
   }
-
   public function getCorners()
   {
     return $this->corners;
   }
-
   public function setFont(Google_Service_AdSense_AdStyleFont $font)
   {
     $this->font = $font;
   }
-
   public function getFont()
   {
     return $this->font;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -2489,57 +2417,51 @@ class Google_Service_AdSense_AdStyle extends Google_Model
 
 class Google_Service_AdSense_AdStyleColors extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $background;
   public $border;
   public $text;
   public $title;
   public $url;
 
+
   public function setBackground($background)
   {
     $this->background = $background;
   }
-
   public function getBackground()
   {
     return $this->background;
   }
-
   public function setBorder($border)
   {
     $this->border = $border;
   }
-
   public function getBorder()
   {
     return $this->border;
   }
-
   public function setText($text)
   {
     $this->text = $text;
   }
-
   public function getText()
   {
     return $this->text;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -2548,24 +2470,24 @@ class Google_Service_AdSense_AdStyleColors extends Google_Model
 
 class Google_Service_AdSense_AdStyleFont extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $family;
   public $size;
+
 
   public function setFamily($family)
   {
     $this->family = $family;
   }
-
   public function getFamily()
   {
     return $this->family;
   }
-
   public function setSize($size)
   {
     $this->size = $size;
   }
-
   public function getSize()
   {
     return $this->size;
@@ -2574,6 +2496,8 @@ class Google_Service_AdSense_AdStyleFont extends Google_Model
 
 class Google_Service_AdSense_AdUnit extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $code;
   protected $contentAdsSettingsType = 'Google_Service_AdSense_AdUnitContentAdsSettings';
   protected $contentAdsSettingsDataType = '';
@@ -2589,101 +2513,83 @@ class Google_Service_AdSense_AdUnit extends Google_Model
   public $savedStyleId;
   public $status;
 
+
   public function setCode($code)
   {
     $this->code = $code;
   }
-
   public function getCode()
   {
     return $this->code;
   }
-
   public function setContentAdsSettings(Google_Service_AdSense_AdUnitContentAdsSettings $contentAdsSettings)
   {
     $this->contentAdsSettings = $contentAdsSettings;
   }
-
   public function getContentAdsSettings()
   {
     return $this->contentAdsSettings;
   }
-
   public function setCustomStyle(Google_Service_AdSense_AdStyle $customStyle)
   {
     $this->customStyle = $customStyle;
   }
-
   public function getCustomStyle()
   {
     return $this->customStyle;
   }
-
   public function setFeedAdsSettings(Google_Service_AdSense_AdUnitFeedAdsSettings $feedAdsSettings)
   {
     $this->feedAdsSettings = $feedAdsSettings;
   }
-
   public function getFeedAdsSettings()
   {
     return $this->feedAdsSettings;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setMobileContentAdsSettings(Google_Service_AdSense_AdUnitMobileContentAdsSettings $mobileContentAdsSettings)
   {
     $this->mobileContentAdsSettings = $mobileContentAdsSettings;
   }
-
   public function getMobileContentAdsSettings()
   {
     return $this->mobileContentAdsSettings;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setSavedStyleId($savedStyleId)
   {
     $this->savedStyleId = $savedStyleId;
   }
-
   public function getSavedStyleId()
   {
     return $this->savedStyleId;
   }
-
   public function setStatus($status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
@@ -2692,36 +2598,34 @@ class Google_Service_AdSense_AdUnit extends Google_Model
 
 class Google_Service_AdSense_AdUnitContentAdsSettings extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $backupOptionType = 'Google_Service_AdSense_AdUnitContentAdsSettingsBackupOption';
   protected $backupOptionDataType = '';
   public $size;
   public $type;
 
+
   public function setBackupOption(Google_Service_AdSense_AdUnitContentAdsSettingsBackupOption $backupOption)
   {
     $this->backupOption = $backupOption;
   }
-
   public function getBackupOption()
   {
     return $this->backupOption;
   }
-
   public function setSize($size)
   {
     $this->size = $size;
   }
-
   public function getSize()
   {
     return $this->size;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -2730,35 +2634,33 @@ class Google_Service_AdSense_AdUnitContentAdsSettings extends Google_Model
 
 class Google_Service_AdSense_AdUnitContentAdsSettingsBackupOption extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $color;
   public $type;
   public $url;
+
 
   public function setColor($color)
   {
     $this->color = $color;
   }
-
   public function getColor()
   {
     return $this->color;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -2767,46 +2669,42 @@ class Google_Service_AdSense_AdUnitContentAdsSettingsBackupOption extends Google
 
 class Google_Service_AdSense_AdUnitFeedAdsSettings extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $adPosition;
   public $frequency;
   public $minimumWordCount;
   public $type;
 
+
   public function setAdPosition($adPosition)
   {
     $this->adPosition = $adPosition;
   }
-
   public function getAdPosition()
   {
     return $this->adPosition;
   }
-
   public function setFrequency($frequency)
   {
     $this->frequency = $frequency;
   }
-
   public function getFrequency()
   {
     return $this->frequency;
   }
-
   public function setMinimumWordCount($minimumWordCount)
   {
     $this->minimumWordCount = $minimumWordCount;
   }
-
   public function getMinimumWordCount()
   {
     return $this->minimumWordCount;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -2815,46 +2713,42 @@ class Google_Service_AdSense_AdUnitFeedAdsSettings extends Google_Model
 
 class Google_Service_AdSense_AdUnitMobileContentAdsSettings extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $markupLanguage;
   public $scriptingLanguage;
   public $size;
   public $type;
 
+
   public function setMarkupLanguage($markupLanguage)
   {
     $this->markupLanguage = $markupLanguage;
   }
-
   public function getMarkupLanguage()
   {
     return $this->markupLanguage;
   }
-
   public function setScriptingLanguage($scriptingLanguage)
   {
     $this->scriptingLanguage = $scriptingLanguage;
   }
-
   public function getScriptingLanguage()
   {
     return $this->scriptingLanguage;
   }
-
   public function setSize($size)
   {
     $this->size = $size;
   }
-
   public function getSize()
   {
     return $this->size;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -2863,47 +2757,44 @@ class Google_Service_AdSense_AdUnitMobileContentAdsSettings extends Google_Model
 
 class Google_Service_AdSense_AdUnits extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_AdUnit';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -2912,6 +2803,9 @@ class Google_Service_AdSense_AdUnits extends Google_Collection
 
 class Google_Service_AdSense_AdsenseReportsGenerateResponse extends Google_Collection
 {
+  protected $collection_key = 'warnings';
+  protected $internal_gapi_mappings = array(
+  );
   public $averages;
   public $endDate;
   protected $headersType = 'Google_Service_AdSense_AdsenseReportsGenerateResponseHeaders';
@@ -2923,91 +2817,75 @@ class Google_Service_AdSense_AdsenseReportsGenerateResponse extends Google_Colle
   public $totals;
   public $warnings;
 
+
   public function setAverages($averages)
   {
     $this->averages = $averages;
   }
-
   public function getAverages()
   {
     return $this->averages;
   }
-
   public function setEndDate($endDate)
   {
     $this->endDate = $endDate;
   }
-
   public function getEndDate()
   {
     return $this->endDate;
   }
-
   public function setHeaders($headers)
   {
     $this->headers = $headers;
   }
-
   public function getHeaders()
   {
     return $this->headers;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setRows($rows)
   {
     $this->rows = $rows;
   }
-
   public function getRows()
   {
     return $this->rows;
   }
-
   public function setStartDate($startDate)
   {
     $this->startDate = $startDate;
   }
-
   public function getStartDate()
   {
     return $this->startDate;
   }
-
   public function setTotalMatchedRows($totalMatchedRows)
   {
     $this->totalMatchedRows = $totalMatchedRows;
   }
-
   public function getTotalMatchedRows()
   {
     return $this->totalMatchedRows;
   }
-
   public function setTotals($totals)
   {
     $this->totals = $totals;
   }
-
   public function getTotals()
   {
     return $this->totals;
   }
-
   public function setWarnings($warnings)
   {
     $this->warnings = $warnings;
   }
-
   public function getWarnings()
   {
     return $this->warnings;
@@ -3016,35 +2894,33 @@ class Google_Service_AdSense_AdsenseReportsGenerateResponse extends Google_Colle
 
 class Google_Service_AdSense_AdsenseReportsGenerateResponseHeaders extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $currency;
   public $name;
   public $type;
+
 
   public function setCurrency($currency)
   {
     $this->currency = $currency;
   }
-
   public function getCurrency()
   {
     return $this->currency;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -3053,6 +2929,8 @@ class Google_Service_AdSense_AdsenseReportsGenerateResponseHeaders extends Googl
 
 class Google_Service_AdSense_Alert extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $isDismissible;
   public $kind;
@@ -3060,61 +2938,51 @@ class Google_Service_AdSense_Alert extends Google_Model
   public $severity;
   public $type;
 
+
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setIsDismissible($isDismissible)
   {
     $this->isDismissible = $isDismissible;
   }
-
   public function getIsDismissible()
   {
     return $this->isDismissible;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setMessage($message)
   {
     $this->message = $message;
   }
-
   public function getMessage()
   {
     return $this->message;
   }
-
   public function setSeverity($severity)
   {
     $this->severity = $severity;
   }
-
   public function getSeverity()
   {
     return $this->severity;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -3123,25 +2991,26 @@ class Google_Service_AdSense_Alert extends Google_Model
 
 class Google_Service_AdSense_Alerts extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_AdSense_Alert';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -3150,6 +3019,8 @@ class Google_Service_AdSense_Alerts extends Google_Collection
 
 class Google_Service_AdSense_CustomChannel extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $code;
   public $id;
   public $kind;
@@ -3157,51 +3028,43 @@ class Google_Service_AdSense_CustomChannel extends Google_Model
   protected $targetingInfoType = 'Google_Service_AdSense_CustomChannelTargetingInfo';
   protected $targetingInfoDataType = '';
 
+
   public function setCode($code)
   {
     $this->code = $code;
   }
-
   public function getCode()
   {
     return $this->code;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setTargetingInfo(Google_Service_AdSense_CustomChannelTargetingInfo $targetingInfo)
   {
     $this->targetingInfo = $targetingInfo;
   }
-
   public function getTargetingInfo()
   {
     return $this->targetingInfo;
@@ -3210,46 +3073,42 @@ class Google_Service_AdSense_CustomChannel extends Google_Model
 
 class Google_Service_AdSense_CustomChannelTargetingInfo extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $adsAppearOn;
   public $description;
   public $location;
   public $siteLanguage;
 
+
   public function setAdsAppearOn($adsAppearOn)
   {
     $this->adsAppearOn = $adsAppearOn;
   }
-
   public function getAdsAppearOn()
   {
     return $this->adsAppearOn;
   }
-
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setLocation($location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setSiteLanguage($siteLanguage)
   {
     $this->siteLanguage = $siteLanguage;
   }
-
   public function getSiteLanguage()
   {
     return $this->siteLanguage;
@@ -3258,47 +3117,44 @@ class Google_Service_AdSense_CustomChannelTargetingInfo extends Google_Model
 
 class Google_Service_AdSense_CustomChannels extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_CustomChannel';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -3307,25 +3163,26 @@ class Google_Service_AdSense_CustomChannels extends Google_Collection
 
 class Google_Service_AdSense_Metadata extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_AdSense_ReportingMetadataEntry';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -3334,57 +3191,51 @@ class Google_Service_AdSense_Metadata extends Google_Collection
 
 class Google_Service_AdSense_Payment extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
   public $paymentAmount;
   public $paymentAmountCurrencyCode;
   public $paymentDate;
 
+
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setPaymentAmount($paymentAmount)
   {
     $this->paymentAmount = $paymentAmount;
   }
-
   public function getPaymentAmount()
   {
     return $this->paymentAmount;
   }
-
   public function setPaymentAmountCurrencyCode($paymentAmountCurrencyCode)
   {
     $this->paymentAmountCurrencyCode = $paymentAmountCurrencyCode;
   }
-
   public function getPaymentAmountCurrencyCode()
   {
     return $this->paymentAmountCurrencyCode;
   }
-
   public function setPaymentDate($paymentDate)
   {
     $this->paymentDate = $paymentDate;
   }
-
   public function getPaymentDate()
   {
     return $this->paymentDate;
@@ -3393,25 +3244,26 @@ class Google_Service_AdSense_Payment extends Google_Model
 
 class Google_Service_AdSense_Payments extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_AdSense_Payment';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -3420,6 +3272,9 @@ class Google_Service_AdSense_Payments extends Google_Collection
 
 class Google_Service_AdSense_ReportingMetadataEntry extends Google_Collection
 {
+  protected $collection_key = 'supportedProducts';
+  protected $internal_gapi_mappings = array(
+  );
   public $compatibleDimensions;
   public $compatibleMetrics;
   public $id;
@@ -3428,71 +3283,59 @@ class Google_Service_AdSense_ReportingMetadataEntry extends Google_Collection
   public $requiredMetrics;
   public $supportedProducts;
 
+
   public function setCompatibleDimensions($compatibleDimensions)
   {
     $this->compatibleDimensions = $compatibleDimensions;
   }
-
   public function getCompatibleDimensions()
   {
     return $this->compatibleDimensions;
   }
-
   public function setCompatibleMetrics($compatibleMetrics)
   {
     $this->compatibleMetrics = $compatibleMetrics;
   }
-
   public function getCompatibleMetrics()
   {
     return $this->compatibleMetrics;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setRequiredDimensions($requiredDimensions)
   {
     $this->requiredDimensions = $requiredDimensions;
   }
-
   public function getRequiredDimensions()
   {
     return $this->requiredDimensions;
   }
-
   public function setRequiredMetrics($requiredMetrics)
   {
     $this->requiredMetrics = $requiredMetrics;
   }
-
   public function getRequiredMetrics()
   {
     return $this->requiredMetrics;
   }
-
   public function setSupportedProducts($supportedProducts)
   {
     $this->supportedProducts = $supportedProducts;
   }
-
   public function getSupportedProducts()
   {
     return $this->supportedProducts;
@@ -3501,47 +3344,43 @@ class Google_Service_AdSense_ReportingMetadataEntry extends Google_Collection
 
 class Google_Service_AdSense_SavedAdStyle extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $adStyleType = 'Google_Service_AdSense_AdStyle';
   protected $adStyleDataType = '';
   public $id;
   public $kind;
   public $name;
 
+
   public function setAdStyle(Google_Service_AdSense_AdStyle $adStyle)
   {
     $this->adStyle = $adStyle;
   }
-
   public function getAdStyle()
   {
     return $this->adStyle;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
@@ -3550,47 +3389,44 @@ class Google_Service_AdSense_SavedAdStyle extends Google_Model
 
 class Google_Service_AdSense_SavedAdStyles extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_SavedAdStyle';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -3599,35 +3435,33 @@ class Google_Service_AdSense_SavedAdStyles extends Google_Collection
 
 class Google_Service_AdSense_SavedReport extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
   public $name;
+
 
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
@@ -3636,47 +3470,44 @@ class Google_Service_AdSense_SavedReport extends Google_Model
 
 class Google_Service_AdSense_SavedReports extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_SavedReport';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -3685,35 +3516,33 @@ class Google_Service_AdSense_SavedReports extends Google_Collection
 
 class Google_Service_AdSense_UrlChannel extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
   public $urlPattern;
+
 
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setUrlPattern($urlPattern)
   {
     $this->urlPattern = $urlPattern;
   }
-
   public function getUrlPattern()
   {
     return $this->urlPattern;
@@ -3722,47 +3551,44 @@ class Google_Service_AdSense_UrlChannel extends Google_Model
 
 class Google_Service_AdSense_UrlChannels extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSense_UrlChannel';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
