@@ -19,8 +19,8 @@
  * Service definition for AdSenseHost (v4.1).
  *
  * <p>
- * Gives AdSense Hosts access to report generation, ad code generation, and publisher management capabilities.
- * </p>
+ * Gives AdSense Hosts access to report generation, ad code generation, and
+ * publisher management capabilities.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,7 +32,8 @@
 class Google_Service_AdSenseHost extends Google_Service
 {
   /** View and manage your AdSense host data and associated accounts. */
-  const ADSENSEHOST = "https://www.googleapis.com/auth/adsensehost";
+  const ADSENSEHOST =
+      "https://www.googleapis.com/auth/adsensehost";
 
   public $accounts;
   public $accounts_adclients;
@@ -53,6 +54,7 @@ class Google_Service_AdSenseHost extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'adsensehost/v4.1/';
     $this->version = 'v4.1';
     $this->serviceName = 'adsensehost';
@@ -643,8 +645,7 @@ class Google_Service_AdSenseHost_Accounts_Resource extends Google_Service_Resour
   /**
    * Get information about the selected associated AdSense account. (accounts.get)
    *
-   * @param string $accountId
-   * Account to get information about.
+   * @param string $accountId Account to get information about.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_Account
    */
@@ -654,12 +655,12 @@ class Google_Service_AdSenseHost_Accounts_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSenseHost_Account");
   }
+
   /**
    * List hosted accounts associated with this AdSense account by ad client id.
    * (accounts.listAccounts)
    *
-   * @param string $filterAdClientId
-   * Ad clients to list accounts for.
+   * @param string $filterAdClientId Ad clients to list accounts for.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_Accounts
    */
@@ -686,10 +687,8 @@ class Google_Service_AdSenseHost_AccountsAdclients_Resource extends Google_Servi
    * Get information about one of the ad clients in the specified publisher's
    * AdSense account. (adclients.get)
    *
-   * @param string $accountId
-   * Account which contains the ad client.
-   * @param string $adClientId
-   * Ad client to get.
+   * @param string $accountId Account which contains the ad client.
+   * @param string $adClientId Ad client to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdClient
    */
@@ -699,19 +698,19 @@ class Google_Service_AdSenseHost_AccountsAdclients_Resource extends Google_Servi
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSenseHost_AdClient");
   }
+
   /**
    * List all hosted ad clients in the specified hosted account.
    * (adclients.listAccountsAdclients)
    *
-   * @param string $accountId
-   * Account for which to list ad clients.
+   * @param string $accountId Account for which to list ad clients.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad clients. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of ad clients to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * clients. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of ad clients to include in
+   * the response, used for paging.
    * @return Google_Service_AdSenseHost_AdClients
    */
   public function listAccountsAdclients($accountId, $optParams = array())
@@ -736,12 +735,9 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
    * Delete the specified ad unit from the specified publisher AdSense account.
    * (adunits.delete)
    *
-   * @param string $accountId
-   * Account which contains the ad unit.
-   * @param string $adClientId
-   * Ad client for which to get ad unit.
-   * @param string $adUnitId
-   * Ad unit to delete.
+   * @param string $accountId Account which contains the ad unit.
+   * @param string $adClientId Ad client for which to get ad unit.
+   * @param string $adUnitId Ad unit to delete.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdUnit
    */
@@ -751,15 +747,13 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params), "Google_Service_AdSenseHost_AdUnit");
   }
+
   /**
    * Get the specified host ad unit in this AdSense account. (adunits.get)
    *
-   * @param string $accountId
-   * Account which contains the ad unit.
-   * @param string $adClientId
-   * Ad client for which to get ad unit.
-   * @param string $adUnitId
-   * Ad unit to get.
+   * @param string $accountId Account which contains the ad unit.
+   * @param string $adClientId Ad client for which to get ad unit.
+   * @param string $adUnitId Ad unit to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdUnit
    */
@@ -769,20 +763,18 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSenseHost_AdUnit");
   }
+
   /**
    * Get ad code for the specified ad unit, attaching the specified host custom
    * channels. (adunits.getAdCode)
    *
-   * @param string $accountId
-   * Account which contains the ad client.
-   * @param string $adClientId
-   * Ad client with contains the ad unit.
-   * @param string $adUnitId
-   * Ad unit to get the code for.
+   * @param string $accountId Account which contains the ad client.
+   * @param string $adClientId Ad client with contains the ad unit.
+   * @param string $adUnitId Ad unit to get the code for.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string hostCustomChannelId
-   * Host custom channel to attach to the ad code.
+   * @opt_param string hostCustomChannelId Host custom channel to attach to the ad
+   * code.
    * @return Google_Service_AdSenseHost_AdCode
    */
   public function getAdCode($accountId, $adClientId, $adUnitId, $optParams = array())
@@ -791,14 +783,13 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
     $params = array_merge($params, $optParams);
     return $this->call('getAdCode', array($params), "Google_Service_AdSenseHost_AdCode");
   }
+
   /**
    * Insert the supplied ad unit into the specified publisher AdSense account.
    * (adunits.insert)
    *
-   * @param string $accountId
-   * Account which will contain the ad unit.
-   * @param string $adClientId
-   * Ad client into which to insert the ad unit.
+   * @param string $accountId Account which will contain the ad unit.
+   * @param string $adClientId Ad client into which to insert the ad unit.
    * @param Google_AdUnit $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdUnit
@@ -809,23 +800,22 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_AdSenseHost_AdUnit");
   }
+
   /**
    * List all ad units in the specified publisher's AdSense account.
    * (adunits.listAccountsAdunits)
    *
-   * @param string $accountId
-   * Account which contains the ad client.
-   * @param string $adClientId
-   * Ad client for which to list ad units.
+   * @param string $accountId Account which contains the ad client.
+   * @param string $adClientId Ad client for which to list ad units.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool includeInactive
-   * Whether to include inactive ad units. Default: true.
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad units. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of ad units to include in the response, used for paging.
+   * @opt_param bool includeInactive Whether to include inactive ad units.
+   * Default: true.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * units. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of ad units to include in the
+   * response, used for paging.
    * @return Google_Service_AdSenseHost_AdUnits
    */
   public function listAccountsAdunits($accountId, $adClientId, $optParams = array())
@@ -834,16 +824,14 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_AdSenseHost_AdUnits");
   }
+
   /**
    * Update the supplied ad unit in the specified publisher AdSense account. This
    * method supports patch semantics. (adunits.patch)
    *
-   * @param string $accountId
-   * Account which contains the ad client.
-   * @param string $adClientId
-   * Ad client which contains the ad unit.
-   * @param string $adUnitId
-   * Ad unit to get.
+   * @param string $accountId Account which contains the ad client.
+   * @param string $adClientId Ad client which contains the ad unit.
+   * @param string $adUnitId Ad unit to get.
    * @param Google_AdUnit $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdUnit
@@ -854,14 +842,13 @@ class Google_Service_AdSenseHost_AccountsAdunits_Resource extends Google_Service
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_AdSenseHost_AdUnit");
   }
+
   /**
    * Update the supplied ad unit in the specified publisher AdSense account.
    * (adunits.update)
    *
-   * @param string $accountId
-   * Account which contains the ad client.
-   * @param string $adClientId
-   * Ad client which contains the ad unit.
+   * @param string $accountId Account which contains the ad client.
+   * @param string $adClientId Ad client which contains the ad unit.
    * @param Google_AdUnit $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdUnit
@@ -889,31 +876,24 @@ class Google_Service_AdSenseHost_AccountsReports_Resource extends Google_Service
    * parameters. Returns the result as JSON; to retrieve output in CSV format
    * specify "alt=csv" as a query parameter. (reports.generate)
    *
-   * @param string $accountId
-   * Hosted account upon which to report.
-   * @param string $startDate
-   * Start of the date range to report on in "YYYY-MM-DD" format, inclusive.
-   * @param string $endDate
-   * End of the date range to report on in "YYYY-MM-DD" format, inclusive.
+   * @param string $accountId Hosted account upon which to report.
+   * @param string $startDate Start of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
+   * @param string $endDate End of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string sort
-   * The name of a dimension or metric to sort the resulting report on, optionally prefixed with "+"
-    * to sort ascending or "-" to sort descending. If no prefix is specified, the column is sorted
-    * ascending.
-   * @opt_param string locale
-   * Optional locale to use for translating report output to a local language. Defaults to "en_US" if
-    * not specified.
-   * @opt_param string metric
-   * Numeric columns to include in the report.
-   * @opt_param string maxResults
-   * The maximum number of rows of report data to return.
-   * @opt_param string filter
-   * Filters to be run on the report.
-   * @opt_param string startIndex
-   * Index of the first row of report data to return.
-   * @opt_param string dimension
-   * Dimensions to base the report on.
+   * @opt_param string sort The name of a dimension or metric to sort the
+   * resulting report on, optionally prefixed with "+" to sort ascending or "-" to
+   * sort descending. If no prefix is specified, the column is sorted ascending.
+   * @opt_param string locale Optional locale to use for translating report output
+   * to a local language. Defaults to "en_US" if not specified.
+   * @opt_param string metric Numeric columns to include in the report.
+   * @opt_param string maxResults The maximum number of rows of report data to
+   * return.
+   * @opt_param string filter Filters to be run on the report.
+   * @opt_param string startIndex Index of the first row of report data to return.
+   * @opt_param string dimension Dimensions to base the report on.
    * @return Google_Service_AdSenseHost_Report
    */
   public function generate($accountId, $startDate, $endDate, $optParams = array())
@@ -939,8 +919,7 @@ class Google_Service_AdSenseHost_Adclients_Resource extends Google_Service_Resou
    * Get information about one of the ad clients in the Host AdSense account.
    * (adclients.get)
    *
-   * @param string $adClientId
-   * Ad client to get.
+   * @param string $adClientId Ad client to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AdClient
    */
@@ -950,16 +929,17 @@ class Google_Service_AdSenseHost_Adclients_Resource extends Google_Service_Resou
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSenseHost_AdClient");
   }
+
   /**
    * List all host ad clients in this AdSense account. (adclients.listAdclients)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through ad clients. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of ad clients to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through ad
+   * clients. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of ad clients to include in
+   * the response, used for paging.
    * @return Google_Service_AdSenseHost_AdClients
    */
   public function listAdclients($optParams = array())
@@ -985,16 +965,12 @@ class Google_Service_AdSenseHost_Associationsessions_Resource extends Google_Ser
    * Create an association session for initiating an association with an AdSense
    * user. (associationsessions.start)
    *
-   * @param string $productCode
-   * Products to associate with the user.
-   * @param string $websiteUrl
-   * The URL of the user's hosted website.
+   * @param string $productCode Products to associate with the user.
+   * @param string $websiteUrl The URL of the user's hosted website.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string websiteLocale
-   * The locale of the user's hosted website.
-   * @opt_param string userLocale
-   * The preferred locale of the user.
+   * @opt_param string websiteLocale The locale of the user's hosted website.
+   * @opt_param string userLocale The preferred locale of the user.
    * @return Google_Service_AdSenseHost_AssociationSession
    */
   public function start($productCode, $websiteUrl, $optParams = array())
@@ -1003,12 +979,12 @@ class Google_Service_AdSenseHost_Associationsessions_Resource extends Google_Ser
     $params = array_merge($params, $optParams);
     return $this->call('start', array($params), "Google_Service_AdSenseHost_AssociationSession");
   }
+
   /**
    * Verify an association session after the association callback returns from
    * AdSense signup. (associationsessions.verify)
    *
-   * @param string $token
-   * The token returned to the association callback URL.
+   * @param string $token The token returned to the association callback URL.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_AssociationSession
    */
@@ -1035,10 +1011,8 @@ class Google_Service_AdSenseHost_Customchannels_Resource extends Google_Service_
    * Delete a specific custom channel from the host AdSense account.
    * (customchannels.delete)
    *
-   * @param string $adClientId
-   * Ad client from which to delete the custom channel.
-   * @param string $customChannelId
-   * Custom channel to delete.
+   * @param string $adClientId Ad client from which to delete the custom channel.
+   * @param string $customChannelId Custom channel to delete.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_CustomChannel
    */
@@ -1048,14 +1022,13 @@ class Google_Service_AdSenseHost_Customchannels_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params), "Google_Service_AdSenseHost_CustomChannel");
   }
+
   /**
    * Get a specific custom channel from the host AdSense account.
    * (customchannels.get)
    *
-   * @param string $adClientId
-   * Ad client from which to get the custom channel.
-   * @param string $customChannelId
-   * Custom channel to get.
+   * @param string $adClientId Ad client from which to get the custom channel.
+   * @param string $customChannelId Custom channel to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_CustomChannel
    */
@@ -1065,11 +1038,12 @@ class Google_Service_AdSenseHost_Customchannels_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_AdSenseHost_CustomChannel");
   }
+
   /**
    * Add a new custom channel to the host AdSense account. (customchannels.insert)
    *
-   * @param string $adClientId
-   * Ad client to which the new custom channel will be added.
+   * @param string $adClientId Ad client to which the new custom channel will be
+   * added.
    * @param Google_CustomChannel $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_CustomChannel
@@ -1080,19 +1054,19 @@ class Google_Service_AdSenseHost_Customchannels_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_AdSenseHost_CustomChannel");
   }
+
   /**
    * List all host custom channels in this AdSense account.
    * (customchannels.listCustomchannels)
    *
-   * @param string $adClientId
-   * Ad client for which to list custom channels.
+   * @param string $adClientId Ad client for which to list custom channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through custom channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of custom channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through custom
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of custom channels to include
+   * in the response, used for paging.
    * @return Google_Service_AdSenseHost_CustomChannels
    */
   public function listCustomchannels($adClientId, $optParams = array())
@@ -1101,14 +1075,14 @@ class Google_Service_AdSenseHost_Customchannels_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_AdSenseHost_CustomChannels");
   }
+
   /**
    * Update a custom channel in the host AdSense account. This method supports
    * patch semantics. (customchannels.patch)
    *
-   * @param string $adClientId
-   * Ad client in which the custom channel will be updated.
-   * @param string $customChannelId
-   * Custom channel to get.
+   * @param string $adClientId Ad client in which the custom channel will be
+   * updated.
+   * @param string $customChannelId Custom channel to get.
    * @param Google_CustomChannel $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_CustomChannel
@@ -1119,11 +1093,12 @@ class Google_Service_AdSenseHost_Customchannels_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_AdSenseHost_CustomChannel");
   }
+
   /**
    * Update a custom channel in the host AdSense account. (customchannels.update)
    *
-   * @param string $adClientId
-   * Ad client in which the custom channel will be updated.
+   * @param string $adClientId Ad client in which the custom channel will be
+   * updated.
    * @param Google_CustomChannel $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_CustomChannel
@@ -1152,29 +1127,23 @@ class Google_Service_AdSenseHost_Reports_Resource extends Google_Service_Resourc
    * parameters. Returns the result as JSON; to retrieve output in CSV format
    * specify "alt=csv" as a query parameter. (reports.generate)
    *
-   * @param string $startDate
-   * Start of the date range to report on in "YYYY-MM-DD" format, inclusive.
-   * @param string $endDate
-   * End of the date range to report on in "YYYY-MM-DD" format, inclusive.
+   * @param string $startDate Start of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
+   * @param string $endDate End of the date range to report on in "YYYY-MM-DD"
+   * format, inclusive.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string sort
-   * The name of a dimension or metric to sort the resulting report on, optionally prefixed with "+"
-    * to sort ascending or "-" to sort descending. If no prefix is specified, the column is sorted
-    * ascending.
-   * @opt_param string locale
-   * Optional locale to use for translating report output to a local language. Defaults to "en_US" if
-    * not specified.
-   * @opt_param string metric
-   * Numeric columns to include in the report.
-   * @opt_param string maxResults
-   * The maximum number of rows of report data to return.
-   * @opt_param string filter
-   * Filters to be run on the report.
-   * @opt_param string startIndex
-   * Index of the first row of report data to return.
-   * @opt_param string dimension
-   * Dimensions to base the report on.
+   * @opt_param string sort The name of a dimension or metric to sort the
+   * resulting report on, optionally prefixed with "+" to sort ascending or "-" to
+   * sort descending. If no prefix is specified, the column is sorted ascending.
+   * @opt_param string locale Optional locale to use for translating report output
+   * to a local language. Defaults to "en_US" if not specified.
+   * @opt_param string metric Numeric columns to include in the report.
+   * @opt_param string maxResults The maximum number of rows of report data to
+   * return.
+   * @opt_param string filter Filters to be run on the report.
+   * @opt_param string startIndex Index of the first row of report data to return.
+   * @opt_param string dimension Dimensions to base the report on.
    * @return Google_Service_AdSenseHost_Report
    */
   public function generate($startDate, $endDate, $optParams = array())
@@ -1199,10 +1168,8 @@ class Google_Service_AdSenseHost_Urlchannels_Resource extends Google_Service_Res
   /**
    * Delete a URL channel from the host AdSense account. (urlchannels.delete)
    *
-   * @param string $adClientId
-   * Ad client from which to delete the URL channel.
-   * @param string $urlChannelId
-   * URL channel to delete.
+   * @param string $adClientId Ad client from which to delete the URL channel.
+   * @param string $urlChannelId URL channel to delete.
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_UrlChannel
    */
@@ -1212,11 +1179,12 @@ class Google_Service_AdSenseHost_Urlchannels_Resource extends Google_Service_Res
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params), "Google_Service_AdSenseHost_UrlChannel");
   }
+
   /**
    * Add a new URL channel to the host AdSense account. (urlchannels.insert)
    *
-   * @param string $adClientId
-   * Ad client to which the new URL channel will be added.
+   * @param string $adClientId Ad client to which the new URL channel will be
+   * added.
    * @param Google_UrlChannel $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_AdSenseHost_UrlChannel
@@ -1227,19 +1195,19 @@ class Google_Service_AdSenseHost_Urlchannels_Resource extends Google_Service_Res
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_AdSenseHost_UrlChannel");
   }
+
   /**
    * List all host URL channels in the host AdSense account.
    * (urlchannels.listUrlchannels)
    *
-   * @param string $adClientId
-   * Ad client for which to list URL channels.
+   * @param string $adClientId Ad client for which to list URL channels.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * A continuation token, used to page through URL channels. To retrieve the next page, set this
-    * parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of URL channels to include in the response, used for paging.
+   * @opt_param string pageToken A continuation token, used to page through URL
+   * channels. To retrieve the next page, set this parameter to the value of
+   * "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of URL channels to include in
+   * the response, used for paging.
    * @return Google_Service_AdSenseHost_UrlChannels
    */
   public function listUrlchannels($adClientId, $optParams = array())
@@ -1255,46 +1223,42 @@ class Google_Service_AdSenseHost_Urlchannels_Resource extends Google_Service_Res
 
 class Google_Service_AdSenseHost_Account extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
   public $name;
   public $status;
 
+
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setStatus($status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
@@ -1303,36 +1267,35 @@ class Google_Service_AdSenseHost_Account extends Google_Model
 
 class Google_Service_AdSenseHost_Accounts extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSenseHost_Account';
   protected $itemsDataType = 'array';
   public $kind;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -1341,57 +1304,51 @@ class Google_Service_AdSenseHost_Accounts extends Google_Collection
 
 class Google_Service_AdSenseHost_AdClient extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $arcOptIn;
   public $id;
   public $kind;
   public $productCode;
   public $supportsReporting;
 
+
   public function setArcOptIn($arcOptIn)
   {
     $this->arcOptIn = $arcOptIn;
   }
-
   public function getArcOptIn()
   {
     return $this->arcOptIn;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setProductCode($productCode)
   {
     $this->productCode = $productCode;
   }
-
   public function getProductCode()
   {
     return $this->productCode;
   }
-
   public function setSupportsReporting($supportsReporting)
   {
     $this->supportsReporting = $supportsReporting;
   }
-
   public function getSupportsReporting()
   {
     return $this->supportsReporting;
@@ -1400,47 +1357,44 @@ class Google_Service_AdSenseHost_AdClient extends Google_Model
 
 class Google_Service_AdSenseHost_AdClients extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSenseHost_AdClient';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -1449,24 +1403,24 @@ class Google_Service_AdSenseHost_AdClients extends Google_Collection
 
 class Google_Service_AdSenseHost_AdCode extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $adCode;
   public $kind;
+
 
   public function setAdCode($adCode)
   {
     $this->adCode = $adCode;
   }
-
   public function getAdCode()
   {
     return $this->adCode;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -1475,6 +1429,8 @@ class Google_Service_AdSenseHost_AdCode extends Google_Model
 
 class Google_Service_AdSenseHost_AdStyle extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $colorsType = 'Google_Service_AdSenseHost_AdStyleColors';
   protected $colorsDataType = '';
   public $corners;
@@ -1482,41 +1438,35 @@ class Google_Service_AdSenseHost_AdStyle extends Google_Model
   protected $fontDataType = '';
   public $kind;
 
+
   public function setColors(Google_Service_AdSenseHost_AdStyleColors $colors)
   {
     $this->colors = $colors;
   }
-
   public function getColors()
   {
     return $this->colors;
   }
-
   public function setCorners($corners)
   {
     $this->corners = $corners;
   }
-
   public function getCorners()
   {
     return $this->corners;
   }
-
   public function setFont(Google_Service_AdSenseHost_AdStyleFont $font)
   {
     $this->font = $font;
   }
-
   public function getFont()
   {
     return $this->font;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -1525,57 +1475,51 @@ class Google_Service_AdSenseHost_AdStyle extends Google_Model
 
 class Google_Service_AdSenseHost_AdStyleColors extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $background;
   public $border;
   public $text;
   public $title;
   public $url;
 
+
   public function setBackground($background)
   {
     $this->background = $background;
   }
-
   public function getBackground()
   {
     return $this->background;
   }
-
   public function setBorder($border)
   {
     $this->border = $border;
   }
-
   public function getBorder()
   {
     return $this->border;
   }
-
   public function setText($text)
   {
     $this->text = $text;
   }
-
   public function getText()
   {
     return $this->text;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1584,24 +1528,24 @@ class Google_Service_AdSenseHost_AdStyleColors extends Google_Model
 
 class Google_Service_AdSenseHost_AdStyleFont extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $family;
   public $size;
+
 
   public function setFamily($family)
   {
     $this->family = $family;
   }
-
   public function getFamily()
   {
     return $this->family;
   }
-
   public function setSize($size)
   {
     $this->size = $size;
   }
-
   public function getSize()
   {
     return $this->size;
@@ -1610,6 +1554,8 @@ class Google_Service_AdSenseHost_AdStyleFont extends Google_Model
 
 class Google_Service_AdSenseHost_AdUnit extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $code;
   protected $contentAdsSettingsType = 'Google_Service_AdSenseHost_AdUnitContentAdsSettings';
   protected $contentAdsSettingsDataType = '';
@@ -1622,81 +1568,67 @@ class Google_Service_AdSenseHost_AdUnit extends Google_Model
   public $name;
   public $status;
 
+
   public function setCode($code)
   {
     $this->code = $code;
   }
-
   public function getCode()
   {
     return $this->code;
   }
-
   public function setContentAdsSettings(Google_Service_AdSenseHost_AdUnitContentAdsSettings $contentAdsSettings)
   {
     $this->contentAdsSettings = $contentAdsSettings;
   }
-
   public function getContentAdsSettings()
   {
     return $this->contentAdsSettings;
   }
-
   public function setCustomStyle(Google_Service_AdSenseHost_AdStyle $customStyle)
   {
     $this->customStyle = $customStyle;
   }
-
   public function getCustomStyle()
   {
     return $this->customStyle;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setMobileContentAdsSettings(Google_Service_AdSenseHost_AdUnitMobileContentAdsSettings $mobileContentAdsSettings)
   {
     $this->mobileContentAdsSettings = $mobileContentAdsSettings;
   }
-
   public function getMobileContentAdsSettings()
   {
     return $this->mobileContentAdsSettings;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setStatus($status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
@@ -1705,36 +1637,34 @@ class Google_Service_AdSenseHost_AdUnit extends Google_Model
 
 class Google_Service_AdSenseHost_AdUnitContentAdsSettings extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $backupOptionType = 'Google_Service_AdSenseHost_AdUnitContentAdsSettingsBackupOption';
   protected $backupOptionDataType = '';
   public $size;
   public $type;
 
+
   public function setBackupOption(Google_Service_AdSenseHost_AdUnitContentAdsSettingsBackupOption $backupOption)
   {
     $this->backupOption = $backupOption;
   }
-
   public function getBackupOption()
   {
     return $this->backupOption;
   }
-
   public function setSize($size)
   {
     $this->size = $size;
   }
-
   public function getSize()
   {
     return $this->size;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -1743,35 +1673,33 @@ class Google_Service_AdSenseHost_AdUnitContentAdsSettings extends Google_Model
 
 class Google_Service_AdSenseHost_AdUnitContentAdsSettingsBackupOption extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $color;
   public $type;
   public $url;
+
 
   public function setColor($color)
   {
     $this->color = $color;
   }
-
   public function getColor()
   {
     return $this->color;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1780,46 +1708,42 @@ class Google_Service_AdSenseHost_AdUnitContentAdsSettingsBackupOption extends Go
 
 class Google_Service_AdSenseHost_AdUnitMobileContentAdsSettings extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $markupLanguage;
   public $scriptingLanguage;
   public $size;
   public $type;
 
+
   public function setMarkupLanguage($markupLanguage)
   {
     $this->markupLanguage = $markupLanguage;
   }
-
   public function getMarkupLanguage()
   {
     return $this->markupLanguage;
   }
-
   public function setScriptingLanguage($scriptingLanguage)
   {
     $this->scriptingLanguage = $scriptingLanguage;
   }
-
   public function getScriptingLanguage()
   {
     return $this->scriptingLanguage;
   }
-
   public function setSize($size)
   {
     $this->size = $size;
   }
-
   public function getSize()
   {
     return $this->size;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -1828,47 +1752,44 @@ class Google_Service_AdSenseHost_AdUnitMobileContentAdsSettings extends Google_M
 
 class Google_Service_AdSenseHost_AdUnits extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSenseHost_AdUnit';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -1877,6 +1798,9 @@ class Google_Service_AdSenseHost_AdUnits extends Google_Collection
 
 class Google_Service_AdSenseHost_AssociationSession extends Google_Collection
 {
+  protected $collection_key = 'productCodes';
+  protected $internal_gapi_mappings = array(
+  );
   public $accountId;
   public $id;
   public $kind;
@@ -1887,91 +1811,75 @@ class Google_Service_AdSenseHost_AssociationSession extends Google_Collection
   public $websiteLocale;
   public $websiteUrl;
 
+
   public function setAccountId($accountId)
   {
     $this->accountId = $accountId;
   }
-
   public function getAccountId()
   {
     return $this->accountId;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setProductCodes($productCodes)
   {
     $this->productCodes = $productCodes;
   }
-
   public function getProductCodes()
   {
     return $this->productCodes;
   }
-
   public function setRedirectUrl($redirectUrl)
   {
     $this->redirectUrl = $redirectUrl;
   }
-
   public function getRedirectUrl()
   {
     return $this->redirectUrl;
   }
-
   public function setStatus($status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
   }
-
   public function setUserLocale($userLocale)
   {
     $this->userLocale = $userLocale;
   }
-
   public function getUserLocale()
   {
     return $this->userLocale;
   }
-
   public function setWebsiteLocale($websiteLocale)
   {
     $this->websiteLocale = $websiteLocale;
   }
-
   public function getWebsiteLocale()
   {
     return $this->websiteLocale;
   }
-
   public function setWebsiteUrl($websiteUrl)
   {
     $this->websiteUrl = $websiteUrl;
   }
-
   public function getWebsiteUrl()
   {
     return $this->websiteUrl;
@@ -1980,46 +1888,42 @@ class Google_Service_AdSenseHost_AssociationSession extends Google_Collection
 
 class Google_Service_AdSenseHost_CustomChannel extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $code;
   public $id;
   public $kind;
   public $name;
 
+
   public function setCode($code)
   {
     $this->code = $code;
   }
-
   public function getCode()
   {
     return $this->code;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
@@ -2028,47 +1932,44 @@ class Google_Service_AdSenseHost_CustomChannel extends Google_Model
 
 class Google_Service_AdSenseHost_CustomChannels extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSenseHost_CustomChannel';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -2077,6 +1978,9 @@ class Google_Service_AdSenseHost_CustomChannels extends Google_Collection
 
 class Google_Service_AdSenseHost_Report extends Google_Collection
 {
+  protected $collection_key = 'warnings';
+  protected $internal_gapi_mappings = array(
+  );
   public $averages;
   protected $headersType = 'Google_Service_AdSenseHost_ReportHeaders';
   protected $headersDataType = 'array';
@@ -2086,71 +1990,59 @@ class Google_Service_AdSenseHost_Report extends Google_Collection
   public $totals;
   public $warnings;
 
+
   public function setAverages($averages)
   {
     $this->averages = $averages;
   }
-
   public function getAverages()
   {
     return $this->averages;
   }
-
   public function setHeaders($headers)
   {
     $this->headers = $headers;
   }
-
   public function getHeaders()
   {
     return $this->headers;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setRows($rows)
   {
     $this->rows = $rows;
   }
-
   public function getRows()
   {
     return $this->rows;
   }
-
   public function setTotalMatchedRows($totalMatchedRows)
   {
     $this->totalMatchedRows = $totalMatchedRows;
   }
-
   public function getTotalMatchedRows()
   {
     return $this->totalMatchedRows;
   }
-
   public function setTotals($totals)
   {
     $this->totals = $totals;
   }
-
   public function getTotals()
   {
     return $this->totals;
   }
-
   public function setWarnings($warnings)
   {
     $this->warnings = $warnings;
   }
-
   public function getWarnings()
   {
     return $this->warnings;
@@ -2159,35 +2051,33 @@ class Google_Service_AdSenseHost_Report extends Google_Collection
 
 class Google_Service_AdSenseHost_ReportHeaders extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $currency;
   public $name;
   public $type;
+
 
   public function setCurrency($currency)
   {
     $this->currency = $currency;
   }
-
   public function getCurrency()
   {
     return $this->currency;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -2196,35 +2086,33 @@ class Google_Service_AdSenseHost_ReportHeaders extends Google_Model
 
 class Google_Service_AdSenseHost_UrlChannel extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
   public $urlPattern;
+
 
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setUrlPattern($urlPattern)
   {
     $this->urlPattern = $urlPattern;
   }
-
   public function getUrlPattern()
   {
     return $this->urlPattern;
@@ -2233,47 +2121,44 @@ class Google_Service_AdSenseHost_UrlChannel extends Google_Model
 
 class Google_Service_AdSenseHost_UrlChannels extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_AdSenseHost_UrlChannel';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
