@@ -19,8 +19,7 @@
  * Service definition for Admin (email_migration_v2).
  *
  * <p>
- * Email Migration API lets you migrate emails of users to Google backends.
- * </p>
+ * Email Migration API lets you migrate emails of users to Google backends.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,7 +31,8 @@
 class Google_Service_Admin extends Google_Service
 {
   /** Manage email messages of users on your domain. */
-  const EMAIL_MIGRATION = "https://www.googleapis.com/auth/email.migration";
+  const EMAIL_MIGRATION =
+      "https://www.googleapis.com/auth/email.migration";
 
   public $mail;
   
@@ -45,6 +45,7 @@ class Google_Service_Admin extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'email/v2/users/';
     $this->version = 'email_migration_v2';
     $this->serviceName = 'admin';
@@ -87,8 +88,7 @@ class Google_Service_Admin_Mail_Resource extends Google_Service_Resource
   /**
    * Insert Mail into Google's Gmail backends (mail.insert)
    *
-   * @param string $userKey
-   * The email or immutable id of the user
+   * @param string $userKey The email or immutable id of the user
    * @param Google_MailItem $postBody
    * @param array $optParams Optional parameters.
    */
@@ -105,6 +105,9 @@ class Google_Service_Admin_Mail_Resource extends Google_Service_Resource
 
 class Google_Service_Admin_MailItem extends Google_Collection
 {
+  protected $collection_key = 'labels';
+  protected $internal_gapi_mappings = array(
+  );
   public $isDeleted;
   public $isDraft;
   public $isInbox;
@@ -115,91 +118,75 @@ class Google_Service_Admin_MailItem extends Google_Collection
   public $kind;
   public $labels;
 
+
   public function setIsDeleted($isDeleted)
   {
     $this->isDeleted = $isDeleted;
   }
-
   public function getIsDeleted()
   {
     return $this->isDeleted;
   }
-
   public function setIsDraft($isDraft)
   {
     $this->isDraft = $isDraft;
   }
-
   public function getIsDraft()
   {
     return $this->isDraft;
   }
-
   public function setIsInbox($isInbox)
   {
     $this->isInbox = $isInbox;
   }
-
   public function getIsInbox()
   {
     return $this->isInbox;
   }
-
   public function setIsSent($isSent)
   {
     $this->isSent = $isSent;
   }
-
   public function getIsSent()
   {
     return $this->isSent;
   }
-
   public function setIsStarred($isStarred)
   {
     $this->isStarred = $isStarred;
   }
-
   public function getIsStarred()
   {
     return $this->isStarred;
   }
-
   public function setIsTrash($isTrash)
   {
     $this->isTrash = $isTrash;
   }
-
   public function getIsTrash()
   {
     return $this->isTrash;
   }
-
   public function setIsUnread($isUnread)
   {
     $this->isUnread = $isUnread;
   }
-
   public function getIsUnread()
   {
     return $this->isUnread;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLabels($labels)
   {
     $this->labels = $labels;
   }
-
   public function getLabels()
   {
     return $this->labels;
