@@ -19,8 +19,7 @@
  * Service definition for DoubleClickBidManager (v1).
  *
  * <p>
- * API for viewing and managing your reports in DoubleClick Bid Manager.
- * </p>
+ * API for viewing and managing your reports in DoubleClick Bid Manager.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -47,6 +46,7 @@ class Google_Service_DoubleClickBidManager extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'doubleclickbidmanager/v1/';
     $this->version = 'v1';
     $this->serviceName = 'doubleclickbidmanager';
@@ -165,6 +165,7 @@ class Google_Service_DoubleClickBidManager_Lineitems_Resource extends Google_Ser
     $params = array_merge($params, $optParams);
     return $this->call('downloadlineitems', array($params), "Google_Service_DoubleClickBidManager_DownloadLineItemsResponse");
   }
+
   /**
    * Uploads line items in CSV format. (lineitems.uploadlineitems)
    *
@@ -204,12 +205,12 @@ class Google_Service_DoubleClickBidManager_Queries_Resource extends Google_Servi
     $params = array_merge($params, $optParams);
     return $this->call('createquery', array($params), "Google_Service_DoubleClickBidManager_Query");
   }
+
   /**
    * Deletes a stored query as well as the associated stored reports.
    * (queries.deletequery)
    *
-   * @param string $queryId
-   * Query ID to delete.
+   * @param string $queryId Query ID to delete.
    * @param array $optParams Optional parameters.
    */
   public function deletequery($queryId, $optParams = array())
@@ -218,11 +219,11 @@ class Google_Service_DoubleClickBidManager_Queries_Resource extends Google_Servi
     $params = array_merge($params, $optParams);
     return $this->call('deletequery', array($params));
   }
+
   /**
    * Retrieves a stored query. (queries.getquery)
    *
-   * @param string $queryId
-   * Query ID to retrieve.
+   * @param string $queryId Query ID to retrieve.
    * @param array $optParams Optional parameters.
    * @return Google_Service_DoubleClickBidManager_Query
    */
@@ -232,6 +233,7 @@ class Google_Service_DoubleClickBidManager_Queries_Resource extends Google_Servi
     $params = array_merge($params, $optParams);
     return $this->call('getquery', array($params), "Google_Service_DoubleClickBidManager_Query");
   }
+
   /**
    * Retrieves stored queries. (queries.listqueries)
    *
@@ -244,11 +246,11 @@ class Google_Service_DoubleClickBidManager_Queries_Resource extends Google_Servi
     $params = array_merge($params, $optParams);
     return $this->call('listqueries', array($params), "Google_Service_DoubleClickBidManager_ListQueriesResponse");
   }
+
   /**
    * Runs a stored query to generate a report. (queries.runquery)
    *
-   * @param string $queryId
-   * Query ID to run.
+   * @param string $queryId Query ID to run.
    * @param Google_RunQueryRequest $postBody
    * @param array $optParams Optional parameters.
    */
@@ -274,8 +276,7 @@ class Google_Service_DoubleClickBidManager_Reports_Resource extends Google_Servi
   /**
    * Retrieves stored reports. (reports.listreports)
    *
-   * @param string $queryId
-   * Query ID with which the reports are associated.
+   * @param string $queryId Query ID with which the reports are associated.
    * @param array $optParams Optional parameters.
    * @return Google_Service_DoubleClickBidManager_ListReportsResponse
    */
@@ -292,35 +293,43 @@ class Google_Service_DoubleClickBidManager_Reports_Resource extends Google_Servi
 
 class Google_Service_DoubleClickBidManager_DownloadLineItemsRequest extends Google_Collection
 {
+  protected $collection_key = 'filterIds';
+  protected $internal_gapi_mappings = array(
+  );
+  public $fileSpec;
   public $filterIds;
   public $filterType;
   public $format;
 
+
+  public function setFileSpec($fileSpec)
+  {
+    $this->fileSpec = $fileSpec;
+  }
+  public function getFileSpec()
+  {
+    return $this->fileSpec;
+  }
   public function setFilterIds($filterIds)
   {
     $this->filterIds = $filterIds;
   }
-
   public function getFilterIds()
   {
     return $this->filterIds;
   }
-
   public function setFilterType($filterType)
   {
     $this->filterType = $filterType;
   }
-
   public function getFilterType()
   {
     return $this->filterType;
   }
-
   public function setFormat($format)
   {
     $this->format = $format;
   }
-
   public function getFormat()
   {
     return $this->format;
@@ -329,13 +338,15 @@ class Google_Service_DoubleClickBidManager_DownloadLineItemsRequest extends Goog
 
 class Google_Service_DoubleClickBidManager_DownloadLineItemsResponse extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $lineItems;
+
 
   public function setLineItems($lineItems)
   {
     $this->lineItems = $lineItems;
   }
-
   public function getLineItems()
   {
     return $this->lineItems;
@@ -344,24 +355,24 @@ class Google_Service_DoubleClickBidManager_DownloadLineItemsResponse extends Goo
 
 class Google_Service_DoubleClickBidManager_FilterPair extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $type;
   public $value;
+
 
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -370,25 +381,26 @@ class Google_Service_DoubleClickBidManager_FilterPair extends Google_Model
 
 class Google_Service_DoubleClickBidManager_ListQueriesResponse extends Google_Collection
 {
+  protected $collection_key = 'queries';
+  protected $internal_gapi_mappings = array(
+  );
   public $kind;
   protected $queriesType = 'Google_Service_DoubleClickBidManager_Query';
   protected $queriesDataType = 'array';
+
 
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setQueries($queries)
   {
     $this->queries = $queries;
   }
-
   public function getQueries()
   {
     return $this->queries;
@@ -397,25 +409,26 @@ class Google_Service_DoubleClickBidManager_ListQueriesResponse extends Google_Co
 
 class Google_Service_DoubleClickBidManager_ListReportsResponse extends Google_Collection
 {
+  protected $collection_key = 'reports';
+  protected $internal_gapi_mappings = array(
+  );
   public $kind;
   protected $reportsType = 'Google_Service_DoubleClickBidManager_Report';
   protected $reportsDataType = 'array';
+
 
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setReports($reports)
   {
     $this->reports = $reports;
   }
-
   public function getReports()
   {
     return $this->reports;
@@ -424,6 +437,9 @@ class Google_Service_DoubleClickBidManager_ListReportsResponse extends Google_Co
 
 class Google_Service_DoubleClickBidManager_Parameters extends Google_Collection
 {
+  protected $collection_key = 'metrics';
+  protected $internal_gapi_mappings = array(
+  );
   protected $filtersType = 'Google_Service_DoubleClickBidManager_FilterPair';
   protected $filtersDataType = 'array';
   public $groupBys;
@@ -431,51 +447,43 @@ class Google_Service_DoubleClickBidManager_Parameters extends Google_Collection
   public $metrics;
   public $type;
 
+
   public function setFilters($filters)
   {
     $this->filters = $filters;
   }
-
   public function getFilters()
   {
     return $this->filters;
   }
-
   public function setGroupBys($groupBys)
   {
     $this->groupBys = $groupBys;
   }
-
   public function getGroupBys()
   {
     return $this->groupBys;
   }
-
   public function setIncludeInviteData($includeInviteData)
   {
     $this->includeInviteData = $includeInviteData;
   }
-
   public function getIncludeInviteData()
   {
     return $this->includeInviteData;
   }
-
   public function setMetrics($metrics)
   {
     $this->metrics = $metrics;
   }
-
   public function getMetrics()
   {
     return $this->metrics;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -484,6 +492,8 @@ class Google_Service_DoubleClickBidManager_Parameters extends Google_Collection
 
 class Google_Service_DoubleClickBidManager_Query extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $kind;
   protected $metadataType = 'Google_Service_DoubleClickBidManager_QueryMetadata';
   protected $metadataDataType = '';
@@ -496,81 +506,67 @@ class Google_Service_DoubleClickBidManager_Query extends Google_Model
   protected $scheduleDataType = '';
   public $timezoneCode;
 
+
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setMetadata(Google_Service_DoubleClickBidManager_QueryMetadata $metadata)
   {
     $this->metadata = $metadata;
   }
-
   public function getMetadata()
   {
     return $this->metadata;
   }
-
   public function setParams(Google_Service_DoubleClickBidManager_Parameters $params)
   {
     $this->params = $params;
   }
-
   public function getParams()
   {
     return $this->params;
   }
-
   public function setQueryId($queryId)
   {
     $this->queryId = $queryId;
   }
-
   public function getQueryId()
   {
     return $this->queryId;
   }
-
   public function setReportDataEndTimeMs($reportDataEndTimeMs)
   {
     $this->reportDataEndTimeMs = $reportDataEndTimeMs;
   }
-
   public function getReportDataEndTimeMs()
   {
     return $this->reportDataEndTimeMs;
   }
-
   public function setReportDataStartTimeMs($reportDataStartTimeMs)
   {
     $this->reportDataStartTimeMs = $reportDataStartTimeMs;
   }
-
   public function getReportDataStartTimeMs()
   {
     return $this->reportDataStartTimeMs;
   }
-
   public function setSchedule(Google_Service_DoubleClickBidManager_QuerySchedule $schedule)
   {
     $this->schedule = $schedule;
   }
-
   public function getSchedule()
   {
     return $this->schedule;
   }
-
   public function setTimezoneCode($timezoneCode)
   {
     $this->timezoneCode = $timezoneCode;
   }
-
   public function getTimezoneCode()
   {
     return $this->timezoneCode;
@@ -579,112 +575,106 @@ class Google_Service_DoubleClickBidManager_Query extends Google_Model
 
 class Google_Service_DoubleClickBidManager_QueryMetadata extends Google_Collection
 {
+  protected $collection_key = 'shareEmailAddress';
+  protected $internal_gapi_mappings = array(
+  );
   public $dataRange;
   public $format;
   public $googleCloudStoragePathForLatestReport;
   public $googleDrivePathForLatestReport;
   public $latestReportRunTimeMs;
+  public $locale;
   public $reportCount;
   public $running;
   public $sendNotification;
   public $shareEmailAddress;
   public $title;
 
+
   public function setDataRange($dataRange)
   {
     $this->dataRange = $dataRange;
   }
-
   public function getDataRange()
   {
     return $this->dataRange;
   }
-
   public function setFormat($format)
   {
     $this->format = $format;
   }
-
   public function getFormat()
   {
     return $this->format;
   }
-
   public function setGoogleCloudStoragePathForLatestReport($googleCloudStoragePathForLatestReport)
   {
     $this->googleCloudStoragePathForLatestReport = $googleCloudStoragePathForLatestReport;
   }
-
   public function getGoogleCloudStoragePathForLatestReport()
   {
     return $this->googleCloudStoragePathForLatestReport;
   }
-
   public function setGoogleDrivePathForLatestReport($googleDrivePathForLatestReport)
   {
     $this->googleDrivePathForLatestReport = $googleDrivePathForLatestReport;
   }
-
   public function getGoogleDrivePathForLatestReport()
   {
     return $this->googleDrivePathForLatestReport;
   }
-
   public function setLatestReportRunTimeMs($latestReportRunTimeMs)
   {
     $this->latestReportRunTimeMs = $latestReportRunTimeMs;
   }
-
   public function getLatestReportRunTimeMs()
   {
     return $this->latestReportRunTimeMs;
   }
-
+  public function setLocale($locale)
+  {
+    $this->locale = $locale;
+  }
+  public function getLocale()
+  {
+    return $this->locale;
+  }
   public function setReportCount($reportCount)
   {
     $this->reportCount = $reportCount;
   }
-
   public function getReportCount()
   {
     return $this->reportCount;
   }
-
   public function setRunning($running)
   {
     $this->running = $running;
   }
-
   public function getRunning()
   {
     return $this->running;
   }
-
   public function setSendNotification($sendNotification)
   {
     $this->sendNotification = $sendNotification;
   }
-
   public function getSendNotification()
   {
     return $this->sendNotification;
   }
-
   public function setShareEmailAddress($shareEmailAddress)
   {
     $this->shareEmailAddress = $shareEmailAddress;
   }
-
   public function getShareEmailAddress()
   {
     return $this->shareEmailAddress;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
@@ -693,46 +683,42 @@ class Google_Service_DoubleClickBidManager_QueryMetadata extends Google_Collecti
 
 class Google_Service_DoubleClickBidManager_QuerySchedule extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $endTimeMs;
   public $frequency;
   public $nextRunMinuteOfDay;
   public $nextRunTimezoneCode;
 
+
   public function setEndTimeMs($endTimeMs)
   {
     $this->endTimeMs = $endTimeMs;
   }
-
   public function getEndTimeMs()
   {
     return $this->endTimeMs;
   }
-
   public function setFrequency($frequency)
   {
     $this->frequency = $frequency;
   }
-
   public function getFrequency()
   {
     return $this->frequency;
   }
-
   public function setNextRunMinuteOfDay($nextRunMinuteOfDay)
   {
     $this->nextRunMinuteOfDay = $nextRunMinuteOfDay;
   }
-
   public function getNextRunMinuteOfDay()
   {
     return $this->nextRunMinuteOfDay;
   }
-
   public function setNextRunTimezoneCode($nextRunTimezoneCode)
   {
     $this->nextRunTimezoneCode = $nextRunTimezoneCode;
   }
-
   public function getNextRunTimezoneCode()
   {
     return $this->nextRunTimezoneCode;
@@ -741,6 +727,8 @@ class Google_Service_DoubleClickBidManager_QuerySchedule extends Google_Model
 
 class Google_Service_DoubleClickBidManager_Report extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $keyType = 'Google_Service_DoubleClickBidManager_ReportKey';
   protected $keyDataType = '';
   protected $metadataType = 'Google_Service_DoubleClickBidManager_ReportMetadata';
@@ -748,31 +736,27 @@ class Google_Service_DoubleClickBidManager_Report extends Google_Model
   protected $paramsType = 'Google_Service_DoubleClickBidManager_Parameters';
   protected $paramsDataType = '';
 
+
   public function setKey(Google_Service_DoubleClickBidManager_ReportKey $key)
   {
     $this->key = $key;
   }
-
   public function getKey()
   {
     return $this->key;
   }
-
   public function setMetadata(Google_Service_DoubleClickBidManager_ReportMetadata $metadata)
   {
     $this->metadata = $metadata;
   }
-
   public function getMetadata()
   {
     return $this->metadata;
   }
-
   public function setParams(Google_Service_DoubleClickBidManager_Parameters $params)
   {
     $this->params = $params;
   }
-
   public function getParams()
   {
     return $this->params;
@@ -781,13 +765,15 @@ class Google_Service_DoubleClickBidManager_Report extends Google_Model
 
 class Google_Service_DoubleClickBidManager_ReportFailure extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $errorCode;
+
 
   public function setErrorCode($errorCode)
   {
     $this->errorCode = $errorCode;
   }
-
   public function getErrorCode()
   {
     return $this->errorCode;
@@ -796,24 +782,24 @@ class Google_Service_DoubleClickBidManager_ReportFailure extends Google_Model
 
 class Google_Service_DoubleClickBidManager_ReportKey extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $queryId;
   public $reportId;
+
 
   public function setQueryId($queryId)
   {
     $this->queryId = $queryId;
   }
-
   public function getQueryId()
   {
     return $this->queryId;
   }
-
   public function setReportId($reportId)
   {
     $this->reportId = $reportId;
   }
-
   public function getReportId()
   {
     return $this->reportId;
@@ -822,47 +808,43 @@ class Google_Service_DoubleClickBidManager_ReportKey extends Google_Model
 
 class Google_Service_DoubleClickBidManager_ReportMetadata extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $googleCloudStoragePath;
   public $reportDataEndTimeMs;
   public $reportDataStartTimeMs;
   protected $statusType = 'Google_Service_DoubleClickBidManager_ReportStatus';
   protected $statusDataType = '';
 
+
   public function setGoogleCloudStoragePath($googleCloudStoragePath)
   {
     $this->googleCloudStoragePath = $googleCloudStoragePath;
   }
-
   public function getGoogleCloudStoragePath()
   {
     return $this->googleCloudStoragePath;
   }
-
   public function setReportDataEndTimeMs($reportDataEndTimeMs)
   {
     $this->reportDataEndTimeMs = $reportDataEndTimeMs;
   }
-
   public function getReportDataEndTimeMs()
   {
     return $this->reportDataEndTimeMs;
   }
-
   public function setReportDataStartTimeMs($reportDataStartTimeMs)
   {
     $this->reportDataStartTimeMs = $reportDataStartTimeMs;
   }
-
   public function getReportDataStartTimeMs()
   {
     return $this->reportDataStartTimeMs;
   }
-
   public function setStatus(Google_Service_DoubleClickBidManager_ReportStatus $status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
@@ -871,47 +853,43 @@ class Google_Service_DoubleClickBidManager_ReportMetadata extends Google_Model
 
 class Google_Service_DoubleClickBidManager_ReportStatus extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $failureType = 'Google_Service_DoubleClickBidManager_ReportFailure';
   protected $failureDataType = '';
   public $finishTimeMs;
   public $format;
   public $state;
 
+
   public function setFailure(Google_Service_DoubleClickBidManager_ReportFailure $failure)
   {
     $this->failure = $failure;
   }
-
   public function getFailure()
   {
     return $this->failure;
   }
-
   public function setFinishTimeMs($finishTimeMs)
   {
     $this->finishTimeMs = $finishTimeMs;
   }
-
   public function getFinishTimeMs()
   {
     return $this->finishTimeMs;
   }
-
   public function setFormat($format)
   {
     $this->format = $format;
   }
-
   public function getFormat()
   {
     return $this->format;
   }
-
   public function setState($state)
   {
     $this->state = $state;
   }
-
   public function getState()
   {
     return $this->state;
@@ -920,6 +898,9 @@ class Google_Service_DoubleClickBidManager_ReportStatus extends Google_Model
 
 class Google_Service_DoubleClickBidManager_RowStatus extends Google_Collection
 {
+  protected $collection_key = 'errors';
+  protected $internal_gapi_mappings = array(
+  );
   public $changed;
   public $entityId;
   public $entityName;
@@ -927,61 +908,51 @@ class Google_Service_DoubleClickBidManager_RowStatus extends Google_Collection
   public $persisted;
   public $rowNumber;
 
+
   public function setChanged($changed)
   {
     $this->changed = $changed;
   }
-
   public function getChanged()
   {
     return $this->changed;
   }
-
   public function setEntityId($entityId)
   {
     $this->entityId = $entityId;
   }
-
   public function getEntityId()
   {
     return $this->entityId;
   }
-
   public function setEntityName($entityName)
   {
     $this->entityName = $entityName;
   }
-
   public function getEntityName()
   {
     return $this->entityName;
   }
-
   public function setErrors($errors)
   {
     $this->errors = $errors;
   }
-
   public function getErrors()
   {
     return $this->errors;
   }
-
   public function setPersisted($persisted)
   {
     $this->persisted = $persisted;
   }
-
   public function getPersisted()
   {
     return $this->persisted;
   }
-
   public function setRowNumber($rowNumber)
   {
     $this->rowNumber = $rowNumber;
   }
-
   public function getRowNumber()
   {
     return $this->rowNumber;
@@ -990,46 +961,42 @@ class Google_Service_DoubleClickBidManager_RowStatus extends Google_Collection
 
 class Google_Service_DoubleClickBidManager_RunQueryRequest extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $dataRange;
   public $reportDataEndTimeMs;
   public $reportDataStartTimeMs;
   public $timezoneCode;
 
+
   public function setDataRange($dataRange)
   {
     $this->dataRange = $dataRange;
   }
-
   public function getDataRange()
   {
     return $this->dataRange;
   }
-
   public function setReportDataEndTimeMs($reportDataEndTimeMs)
   {
     $this->reportDataEndTimeMs = $reportDataEndTimeMs;
   }
-
   public function getReportDataEndTimeMs()
   {
     return $this->reportDataEndTimeMs;
   }
-
   public function setReportDataStartTimeMs($reportDataStartTimeMs)
   {
     $this->reportDataStartTimeMs = $reportDataStartTimeMs;
   }
-
   public function getReportDataStartTimeMs()
   {
     return $this->reportDataStartTimeMs;
   }
-
   public function setTimezoneCode($timezoneCode)
   {
     $this->timezoneCode = $timezoneCode;
   }
-
   public function getTimezoneCode()
   {
     return $this->timezoneCode;
@@ -1038,35 +1005,33 @@ class Google_Service_DoubleClickBidManager_RunQueryRequest extends Google_Model
 
 class Google_Service_DoubleClickBidManager_UploadLineItemsRequest extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $dryRun;
   public $format;
   public $lineItems;
+
 
   public function setDryRun($dryRun)
   {
     $this->dryRun = $dryRun;
   }
-
   public function getDryRun()
   {
     return $this->dryRun;
   }
-
   public function setFormat($format)
   {
     $this->format = $format;
   }
-
   public function getFormat()
   {
     return $this->format;
   }
-
   public function setLineItems($lineItems)
   {
     $this->lineItems = $lineItems;
   }
-
   public function getLineItems()
   {
     return $this->lineItems;
@@ -1075,14 +1040,16 @@ class Google_Service_DoubleClickBidManager_UploadLineItemsRequest extends Google
 
 class Google_Service_DoubleClickBidManager_UploadLineItemsResponse extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $uploadStatusType = 'Google_Service_DoubleClickBidManager_UploadStatus';
   protected $uploadStatusDataType = '';
+
 
   public function setUploadStatus(Google_Service_DoubleClickBidManager_UploadStatus $uploadStatus)
   {
     $this->uploadStatus = $uploadStatus;
   }
-
   public function getUploadStatus()
   {
     return $this->uploadStatus;
@@ -1091,25 +1058,26 @@ class Google_Service_DoubleClickBidManager_UploadLineItemsResponse extends Googl
 
 class Google_Service_DoubleClickBidManager_UploadStatus extends Google_Collection
 {
+  protected $collection_key = 'rowStatus';
+  protected $internal_gapi_mappings = array(
+  );
   public $errors;
   protected $rowStatusType = 'Google_Service_DoubleClickBidManager_RowStatus';
   protected $rowStatusDataType = 'array';
+
 
   public function setErrors($errors)
   {
     $this->errors = $errors;
   }
-
   public function getErrors()
   {
     return $this->errors;
   }
-
   public function setRowStatus($rowStatus)
   {
     $this->rowStatus = $rowStatus;
   }
-
   public function getRowStatus()
   {
     return $this->rowStatus;
