@@ -1,6 +1,19 @@
 <?php
-
-require_once 'Google/Exception.php';
+/*
+ * Copyright 2014 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 class Google_Service_Exception extends Google_Exception
 {
@@ -10,13 +23,15 @@ class Google_Service_Exception extends Google_Exception
   protected $errors = array();
 
   /**
-   * Override default constructor to add ability to set $errors.
+   * Override default constructor to add the ability to set $errors and a retry
+   * map.
    *
    * @param string $message
    * @param int $code
    * @param Exception|null $previous
    * @param [{string, string}] errors List of errors returned in an HTTP
    * response.  Defaults to [].
+   * @param array|null $retryMap Map of errors with retry counts.
    */
   public function __construct(
       $message,
