@@ -19,8 +19,7 @@
  * Service definition for Licensing (v1).
  *
  * <p>
- * Licensing API to view and manage license for your domain.
- * </p>
+ * Licensing API to view and manage license for your domain.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -31,7 +30,9 @@
  */
 class Google_Service_Licensing extends Google_Service
 {
-
+  /** View and manage Google Apps licenses for your domain. */
+  const APPS_LICENSING =
+      "https://www.googleapis.com/auth/apps.licensing";
 
   public $licenseAssignments;
   
@@ -44,6 +45,7 @@ class Google_Service_Licensing extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'apps/licensing/v1/product/';
     $this->version = 'v1';
     $this->serviceName = 'licensing';
@@ -222,12 +224,9 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
   /**
    * Revoke License. (licenseAssignments.delete)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $skuId
-   * Name for sku
-   * @param string $userId
-   * email id or unique Id of the user
+   * @param string $productId Name for product
+   * @param string $skuId Name for sku
+   * @param string $userId email id or unique Id of the user
    * @param array $optParams Optional parameters.
    */
   public function delete($productId, $skuId, $userId, $optParams = array())
@@ -236,16 +235,14 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Get license assignment of a particular product and sku for a user
    * (licenseAssignments.get)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $skuId
-   * Name for sku
-   * @param string $userId
-   * email id or unique Id of the user
+   * @param string $productId Name for product
+   * @param string $skuId Name for sku
+   * @param string $userId email id or unique Id of the user
    * @param array $optParams Optional parameters.
    * @return Google_Service_Licensing_LicenseAssignment
    */
@@ -255,13 +252,12 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Licensing_LicenseAssignment");
   }
+
   /**
    * Assign License. (licenseAssignments.insert)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $skuId
-   * Name for sku
+   * @param string $productId Name for product
+   * @param string $skuId Name for sku
    * @param Google_LicenseAssignmentInsert $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Licensing_LicenseAssignment
@@ -272,21 +268,20 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Licensing_LicenseAssignment");
   }
+
   /**
    * List license assignments for given product of the customer.
    * (licenseAssignments.listForProduct)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $customerId
-   * CustomerId represents the customer for whom licenseassignments are queried
+   * @param string $productId Name for product
+   * @param string $customerId CustomerId represents the customer for whom
+   * licenseassignments are queried
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * Token to fetch the next page.Optional. By default server will return first page
-   * @opt_param string maxResults
-   * Maximum number of campaigns to return at one time. Must be positive. Optional. Default value is
-    * 100.
+   * @opt_param string pageToken Token to fetch the next page.Optional. By default
+   * server will return first page
+   * @opt_param string maxResults Maximum number of campaigns to return at one
+   * time. Must be positive. Optional. Default value is 100.
    * @return Google_Service_Licensing_LicenseAssignmentList
    */
   public function listForProduct($productId, $customerId, $optParams = array())
@@ -295,23 +290,21 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('listForProduct', array($params), "Google_Service_Licensing_LicenseAssignmentList");
   }
+
   /**
    * List license assignments for given product and sku of the customer.
    * (licenseAssignments.listForProductAndSku)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $skuId
-   * Name for sku
-   * @param string $customerId
-   * CustomerId represents the customer for whom licenseassignments are queried
+   * @param string $productId Name for product
+   * @param string $skuId Name for sku
+   * @param string $customerId CustomerId represents the customer for whom
+   * licenseassignments are queried
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * Token to fetch the next page.Optional. By default server will return first page
-   * @opt_param string maxResults
-   * Maximum number of campaigns to return at one time. Must be positive. Optional. Default value is
-    * 100.
+   * @opt_param string pageToken Token to fetch the next page.Optional. By default
+   * server will return first page
+   * @opt_param string maxResults Maximum number of campaigns to return at one
+   * time. Must be positive. Optional. Default value is 100.
    * @return Google_Service_Licensing_LicenseAssignmentList
    */
   public function listForProductAndSku($productId, $skuId, $customerId, $optParams = array())
@@ -320,16 +313,14 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('listForProductAndSku', array($params), "Google_Service_Licensing_LicenseAssignmentList");
   }
+
   /**
    * Assign License. This method supports patch semantics.
    * (licenseAssignments.patch)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $skuId
-   * Name for sku for which license would be revoked
-   * @param string $userId
-   * email id or unique Id of the user
+   * @param string $productId Name for product
+   * @param string $skuId Name for sku for which license would be revoked
+   * @param string $userId email id or unique Id of the user
    * @param Google_LicenseAssignment $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Licensing_LicenseAssignment
@@ -340,15 +331,13 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_Licensing_LicenseAssignment");
   }
+
   /**
    * Assign License. (licenseAssignments.update)
    *
-   * @param string $productId
-   * Name for product
-   * @param string $skuId
-   * Name for sku for which license would be revoked
-   * @param string $userId
-   * email id or unique Id of the user
+   * @param string $productId Name for product
+   * @param string $skuId Name for sku for which license would be revoked
+   * @param string $userId email id or unique Id of the user
    * @param Google_LicenseAssignment $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Licensing_LicenseAssignment
@@ -366,6 +355,8 @@ class Google_Service_Licensing_LicenseAssignments_Resource extends Google_Servic
 
 class Google_Service_Licensing_LicenseAssignment extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $etags;
   public $kind;
   public $productId;
@@ -373,61 +364,51 @@ class Google_Service_Licensing_LicenseAssignment extends Google_Model
   public $skuId;
   public $userId;
 
+
   public function setEtags($etags)
   {
     $this->etags = $etags;
   }
-
   public function getEtags()
   {
     return $this->etags;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setProductId($productId)
   {
     $this->productId = $productId;
   }
-
   public function getProductId()
   {
     return $this->productId;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setSkuId($skuId)
   {
     $this->skuId = $skuId;
   }
-
   public function getSkuId()
   {
     return $this->skuId;
   }
-
   public function setUserId($userId)
   {
     $this->userId = $userId;
   }
-
   public function getUserId()
   {
     return $this->userId;
@@ -436,13 +417,15 @@ class Google_Service_Licensing_LicenseAssignment extends Google_Model
 
 class Google_Service_Licensing_LicenseAssignmentInsert extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $userId;
+
 
   public function setUserId($userId)
   {
     $this->userId = $userId;
   }
-
   public function getUserId()
   {
     return $this->userId;
@@ -451,47 +434,44 @@ class Google_Service_Licensing_LicenseAssignmentInsert extends Google_Model
 
 class Google_Service_Licensing_LicenseAssignmentList extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_Licensing_LicenseAssignment';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
