@@ -19,8 +19,7 @@
  * Service definition for Mirror (v1).
  *
  * <p>
- * API for interacting with Glass users via the timeline.
- * </p>
+ * API for interacting with Glass users via the timeline.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,9 +31,11 @@
 class Google_Service_Mirror extends Google_Service
 {
   /** View your location. */
-  const GLASS_LOCATION = "https://www.googleapis.com/auth/glass.location";
+  const GLASS_LOCATION =
+      "https://www.googleapis.com/auth/glass.location";
   /** View and manage your Glass timeline. */
-  const GLASS_TIMELINE = "https://www.googleapis.com/auth/glass.timeline";
+  const GLASS_TIMELINE =
+      "https://www.googleapis.com/auth/glass.timeline";
 
   public $accounts;
   public $contacts;
@@ -53,6 +54,7 @@ class Google_Service_Mirror extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'mirror/v1/';
     $this->version = 'v1';
     $this->serviceName = 'mirror';
@@ -392,12 +394,11 @@ class Google_Service_Mirror_Accounts_Resource extends Google_Service_Resource
   /**
    * Inserts a new account for a user (accounts.insert)
    *
-   * @param string $userToken
-   * The ID for the user.
-   * @param string $accountType
-   * Account type to be passed to Android Account Manager.
-   * @param string $accountName
-   * The name of the account to be passed to the Android Account Manager.
+   * @param string $userToken The ID for the user.
+   * @param string $accountType Account type to be passed to Android Account
+   * Manager.
+   * @param string $accountName The name of the account to be passed to the
+   * Android Account Manager.
    * @param Google_Account $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Account
@@ -424,8 +425,7 @@ class Google_Service_Mirror_Contacts_Resource extends Google_Service_Resource
   /**
    * Deletes a contact. (contacts.delete)
    *
-   * @param string $id
-   * The ID of the contact.
+   * @param string $id The ID of the contact.
    * @param array $optParams Optional parameters.
    */
   public function delete($id, $optParams = array())
@@ -434,11 +434,11 @@ class Google_Service_Mirror_Contacts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Gets a single contact by ID. (contacts.get)
    *
-   * @param string $id
-   * The ID of the contact.
+   * @param string $id The ID of the contact.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Contact
    */
@@ -448,6 +448,7 @@ class Google_Service_Mirror_Contacts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Mirror_Contact");
   }
+
   /**
    * Inserts a new contact. (contacts.insert)
    *
@@ -461,6 +462,7 @@ class Google_Service_Mirror_Contacts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Mirror_Contact");
   }
+
   /**
    * Retrieves a list of contacts for the authenticated user.
    * (contacts.listContacts)
@@ -474,12 +476,12 @@ class Google_Service_Mirror_Contacts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Mirror_ContactsListResponse");
   }
+
   /**
    * Updates a contact in place. This method supports patch semantics.
    * (contacts.patch)
    *
-   * @param string $id
-   * The ID of the contact.
+   * @param string $id The ID of the contact.
    * @param Google_Contact $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Contact
@@ -490,11 +492,11 @@ class Google_Service_Mirror_Contacts_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_Mirror_Contact");
   }
+
   /**
    * Updates a contact in place. (contacts.update)
    *
-   * @param string $id
-   * The ID of the contact.
+   * @param string $id The ID of the contact.
    * @param Google_Contact $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Contact
@@ -521,8 +523,8 @@ class Google_Service_Mirror_Locations_Resource extends Google_Service_Resource
   /**
    * Gets a single location by ID. (locations.get)
    *
-   * @param string $id
-   * The ID of the location or latest for the last known location.
+   * @param string $id The ID of the location or latest for the last known
+   * location.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Location
    */
@@ -532,6 +534,7 @@ class Google_Service_Mirror_Locations_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Mirror_Location");
   }
+
   /**
    * Retrieves a list of locations for the user. (locations.listLocations)
    *
@@ -560,10 +563,11 @@ class Google_Service_Mirror_Settings_Resource extends Google_Service_Resource
   /**
    * Gets a single setting by ID. (settings.get)
    *
-   * @param string $id
-   * The ID of the setting. The following IDs are valid:
-  - locale - The key to the user’s
-    * language/locale (BCP 47 identifier) that Glassware should use to render localized content.
+   * @param string $id The ID of the setting. The following IDs are valid: -
+   * locale - The key to the user’s language/locale (BCP 47 identifier) that
+   * Glassware should use to render localized content.  - timezone - The key to
+   * the user’s current time zone region as defined in the tz database. Example:
+   * America/Los_Angeles.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Setting
    */
@@ -589,8 +593,7 @@ class Google_Service_Mirror_Subscriptions_Resource extends Google_Service_Resour
   /**
    * Deletes a subscription. (subscriptions.delete)
    *
-   * @param string $id
-   * The ID of the subscription.
+   * @param string $id The ID of the subscription.
    * @param array $optParams Optional parameters.
    */
   public function delete($id, $optParams = array())
@@ -599,6 +602,7 @@ class Google_Service_Mirror_Subscriptions_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Creates a new subscription. (subscriptions.insert)
    *
@@ -612,6 +616,7 @@ class Google_Service_Mirror_Subscriptions_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Mirror_Subscription");
   }
+
   /**
    * Retrieves a list of subscriptions for the authenticated user and service.
    * (subscriptions.listSubscriptions)
@@ -625,11 +630,11 @@ class Google_Service_Mirror_Subscriptions_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Mirror_SubscriptionsListResponse");
   }
+
   /**
    * Updates an existing subscription in place. (subscriptions.update)
    *
-   * @param string $id
-   * The ID of the subscription.
+   * @param string $id The ID of the subscription.
    * @param Google_Subscription $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Subscription
@@ -656,8 +661,7 @@ class Google_Service_Mirror_Timeline_Resource extends Google_Service_Resource
   /**
    * Deletes a timeline item. (timeline.delete)
    *
-   * @param string $id
-   * The ID of the timeline item.
+   * @param string $id The ID of the timeline item.
    * @param array $optParams Optional parameters.
    */
   public function delete($id, $optParams = array())
@@ -666,11 +670,11 @@ class Google_Service_Mirror_Timeline_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Gets a single timeline item by ID. (timeline.get)
    *
-   * @param string $id
-   * The ID of the timeline item.
+   * @param string $id The ID of the timeline item.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_TimelineItem
    */
@@ -680,6 +684,7 @@ class Google_Service_Mirror_Timeline_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Mirror_TimelineItem");
   }
+
   /**
    * Inserts a new item into the timeline. (timeline.insert)
    *
@@ -693,26 +698,25 @@ class Google_Service_Mirror_Timeline_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Mirror_TimelineItem");
   }
+
   /**
    * Retrieves a list of timeline items for the authenticated user.
    * (timeline.listTimeline)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string orderBy
-   * Controls the order in which timeline items are returned.
-   * @opt_param bool includeDeleted
-   * If true, tombstone records for deleted items will be returned.
-   * @opt_param string maxResults
-   * The maximum number of items to include in the response, used for paging.
-   * @opt_param string pageToken
-   * Token for the page of results to return.
-   * @opt_param string sourceItemId
-   * If provided, only items with the given sourceItemId will be returned.
-   * @opt_param bool pinnedOnly
-   * If true, only pinned items will be returned.
-   * @opt_param string bundleId
-   * If provided, only items with the given bundleId will be returned.
+   * @opt_param string orderBy Controls the order in which timeline items are
+   * returned.
+   * @opt_param bool includeDeleted If true, tombstone records for deleted items
+   * will be returned.
+   * @opt_param string maxResults The maximum number of items to include in the
+   * response, used for paging.
+   * @opt_param string pageToken Token for the page of results to return.
+   * @opt_param string sourceItemId If provided, only items with the given
+   * sourceItemId will be returned.
+   * @opt_param bool pinnedOnly If true, only pinned items will be returned.
+   * @opt_param string bundleId If provided, only items with the given bundleId
+   * will be returned.
    * @return Google_Service_Mirror_TimelineListResponse
    */
   public function listTimeline($optParams = array())
@@ -721,12 +725,12 @@ class Google_Service_Mirror_Timeline_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Mirror_TimelineListResponse");
   }
+
   /**
    * Updates a timeline item in place. This method supports patch semantics.
    * (timeline.patch)
    *
-   * @param string $id
-   * The ID of the timeline item.
+   * @param string $id The ID of the timeline item.
    * @param Google_TimelineItem $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_TimelineItem
@@ -737,11 +741,11 @@ class Google_Service_Mirror_Timeline_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_Mirror_TimelineItem");
   }
+
   /**
    * Updates a timeline item in place. (timeline.update)
    *
-   * @param string $id
-   * The ID of the timeline item.
+   * @param string $id The ID of the timeline item.
    * @param Google_TimelineItem $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_TimelineItem
@@ -768,10 +772,8 @@ class Google_Service_Mirror_TimelineAttachments_Resource extends Google_Service_
   /**
    * Deletes an attachment from a timeline item. (attachments.delete)
    *
-   * @param string $itemId
-   * The ID of the timeline item the attachment belongs to.
-   * @param string $attachmentId
-   * The ID of the attachment.
+   * @param string $itemId The ID of the timeline item the attachment belongs to.
+   * @param string $attachmentId The ID of the attachment.
    * @param array $optParams Optional parameters.
    */
   public function delete($itemId, $attachmentId, $optParams = array())
@@ -780,14 +782,13 @@ class Google_Service_Mirror_TimelineAttachments_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Retrieves an attachment on a timeline item by item ID and attachment ID.
    * (attachments.get)
    *
-   * @param string $itemId
-   * The ID of the timeline item the attachment belongs to.
-   * @param string $attachmentId
-   * The ID of the attachment.
+   * @param string $itemId The ID of the timeline item the attachment belongs to.
+   * @param string $attachmentId The ID of the attachment.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Attachment
    */
@@ -797,11 +798,11 @@ class Google_Service_Mirror_TimelineAttachments_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Mirror_Attachment");
   }
+
   /**
    * Adds a new attachment to a timeline item. (attachments.insert)
    *
-   * @param string $itemId
-   * The ID of the timeline item the attachment belongs to.
+   * @param string $itemId The ID of the timeline item the attachment belongs to.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_Attachment
    */
@@ -811,12 +812,13 @@ class Google_Service_Mirror_TimelineAttachments_Resource extends Google_Service_
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Mirror_Attachment");
   }
+
   /**
    * Returns a list of attachments for a timeline item.
    * (attachments.listTimelineAttachments)
    *
-   * @param string $itemId
-   * The ID of the timeline item whose attachments should be listed.
+   * @param string $itemId The ID of the timeline item whose attachments should be
+   * listed.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Mirror_AttachmentsListResponse
    */
@@ -833,6 +835,9 @@ class Google_Service_Mirror_TimelineAttachments_Resource extends Google_Service_
 
 class Google_Service_Mirror_Account extends Google_Collection
 {
+  protected $collection_key = 'userData';
+  protected $internal_gapi_mappings = array(
+  );
   protected $authTokensType = 'Google_Service_Mirror_AuthToken';
   protected $authTokensDataType = 'array';
   public $features;
@@ -840,41 +845,35 @@ class Google_Service_Mirror_Account extends Google_Collection
   protected $userDataType = 'Google_Service_Mirror_UserData';
   protected $userDataDataType = 'array';
 
+
   public function setAuthTokens($authTokens)
   {
     $this->authTokens = $authTokens;
   }
-
   public function getAuthTokens()
   {
     return $this->authTokens;
   }
-
   public function setFeatures($features)
   {
     $this->features = $features;
   }
-
   public function getFeatures()
   {
     return $this->features;
   }
-
   public function setPassword($password)
   {
     $this->password = $password;
   }
-
   public function getPassword()
   {
     return $this->password;
   }
-
   public function setUserData($userData)
   {
     $this->userData = $userData;
   }
-
   public function getUserData()
   {
     return $this->userData;
@@ -883,46 +882,42 @@ class Google_Service_Mirror_Account extends Google_Collection
 
 class Google_Service_Mirror_Attachment extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $contentType;
   public $contentUrl;
   public $id;
   public $isProcessingContent;
 
+
   public function setContentType($contentType)
   {
     $this->contentType = $contentType;
   }
-
   public function getContentType()
   {
     return $this->contentType;
   }
-
   public function setContentUrl($contentUrl)
   {
     $this->contentUrl = $contentUrl;
   }
-
   public function getContentUrl()
   {
     return $this->contentUrl;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setIsProcessingContent($isProcessingContent)
   {
     $this->isProcessingContent = $isProcessingContent;
   }
-
   public function getIsProcessingContent()
   {
     return $this->isProcessingContent;
@@ -931,25 +926,26 @@ class Google_Service_Mirror_Attachment extends Google_Model
 
 class Google_Service_Mirror_AttachmentsListResponse extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Mirror_Attachment';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -958,24 +954,24 @@ class Google_Service_Mirror_AttachmentsListResponse extends Google_Collection
 
 class Google_Service_Mirror_AuthToken extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $authToken;
   public $type;
+
 
   public function setAuthToken($authToken)
   {
     $this->authToken = $authToken;
   }
-
   public function getAuthToken()
   {
     return $this->authToken;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -984,13 +980,15 @@ class Google_Service_Mirror_AuthToken extends Google_Model
 
 class Google_Service_Mirror_Command extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $type;
+
 
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -999,6 +997,9 @@ class Google_Service_Mirror_Command extends Google_Model
 
 class Google_Service_Mirror_Contact extends Google_Collection
 {
+  protected $collection_key = 'sharingFeatures';
+  protected $internal_gapi_mappings = array(
+  );
   protected $acceptCommandsType = 'Google_Service_Mirror_Command';
   protected $acceptCommandsDataType = 'array';
   public $acceptTypes;
@@ -1013,121 +1014,99 @@ class Google_Service_Mirror_Contact extends Google_Collection
   public $speakableName;
   public $type;
 
+
   public function setAcceptCommands($acceptCommands)
   {
     $this->acceptCommands = $acceptCommands;
   }
-
   public function getAcceptCommands()
   {
     return $this->acceptCommands;
   }
-
   public function setAcceptTypes($acceptTypes)
   {
     $this->acceptTypes = $acceptTypes;
   }
-
   public function getAcceptTypes()
   {
     return $this->acceptTypes;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImageUrls($imageUrls)
   {
     $this->imageUrls = $imageUrls;
   }
-
   public function getImageUrls()
   {
     return $this->imageUrls;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setPhoneNumber($phoneNumber)
   {
     $this->phoneNumber = $phoneNumber;
   }
-
   public function getPhoneNumber()
   {
     return $this->phoneNumber;
   }
-
   public function setPriority($priority)
   {
     $this->priority = $priority;
   }
-
   public function getPriority()
   {
     return $this->priority;
   }
-
   public function setSharingFeatures($sharingFeatures)
   {
     $this->sharingFeatures = $sharingFeatures;
   }
-
   public function getSharingFeatures()
   {
     return $this->sharingFeatures;
   }
-
   public function setSource($source)
   {
     $this->source = $source;
   }
-
   public function getSource()
   {
     return $this->source;
   }
-
   public function setSpeakableName($speakableName)
   {
     $this->speakableName = $speakableName;
   }
-
   public function getSpeakableName()
   {
     return $this->speakableName;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -1136,25 +1115,26 @@ class Google_Service_Mirror_Contact extends Google_Collection
 
 class Google_Service_Mirror_ContactsListResponse extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Mirror_Contact';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -1163,6 +1143,8 @@ class Google_Service_Mirror_ContactsListResponse extends Google_Collection
 
 class Google_Service_Mirror_Location extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $accuracy;
   public $address;
   public $displayName;
@@ -1172,81 +1154,67 @@ class Google_Service_Mirror_Location extends Google_Model
   public $longitude;
   public $timestamp;
 
+
   public function setAccuracy($accuracy)
   {
     $this->accuracy = $accuracy;
   }
-
   public function getAccuracy()
   {
     return $this->accuracy;
   }
-
   public function setAddress($address)
   {
     $this->address = $address;
   }
-
   public function getAddress()
   {
     return $this->address;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLatitude($latitude)
   {
     $this->latitude = $latitude;
   }
-
   public function getLatitude()
   {
     return $this->latitude;
   }
-
   public function setLongitude($longitude)
   {
     $this->longitude = $longitude;
   }
-
   public function getLongitude()
   {
     return $this->longitude;
   }
-
   public function setTimestamp($timestamp)
   {
     $this->timestamp = $timestamp;
   }
-
   public function getTimestamp()
   {
     return $this->timestamp;
@@ -1255,25 +1223,26 @@ class Google_Service_Mirror_Location extends Google_Model
 
 class Google_Service_Mirror_LocationsListResponse extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Mirror_Location';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -1282,58 +1251,63 @@ class Google_Service_Mirror_LocationsListResponse extends Google_Collection
 
 class Google_Service_Mirror_MenuItem extends Google_Collection
 {
+  protected $collection_key = 'values';
+  protected $internal_gapi_mappings = array(
+        "contextualCommand" => "contextual_command",
+  );
   public $action;
+  public $contextualCommand;
   public $id;
   public $payload;
   public $removeWhenSelected;
   protected $valuesType = 'Google_Service_Mirror_MenuValue';
   protected $valuesDataType = 'array';
 
+
   public function setAction($action)
   {
     $this->action = $action;
   }
-
   public function getAction()
   {
     return $this->action;
   }
-
+  public function setContextualCommand($contextualCommand)
+  {
+    $this->contextualCommand = $contextualCommand;
+  }
+  public function getContextualCommand()
+  {
+    return $this->contextualCommand;
+  }
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setPayload($payload)
   {
     $this->payload = $payload;
   }
-
   public function getPayload()
   {
     return $this->payload;
   }
-
   public function setRemoveWhenSelected($removeWhenSelected)
   {
     $this->removeWhenSelected = $removeWhenSelected;
   }
-
   public function getRemoveWhenSelected()
   {
     return $this->removeWhenSelected;
   }
-
   public function setValues($values)
   {
     $this->values = $values;
   }
-
   public function getValues()
   {
     return $this->values;
@@ -1342,35 +1316,33 @@ class Google_Service_Mirror_MenuItem extends Google_Collection
 
 class Google_Service_Mirror_MenuValue extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $displayName;
   public $iconUrl;
   public $state;
+
 
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setIconUrl($iconUrl)
   {
     $this->iconUrl = $iconUrl;
   }
-
   public function getIconUrl()
   {
     return $this->iconUrl;
   }
-
   public function setState($state)
   {
     $this->state = $state;
   }
-
   public function getState()
   {
     return $this->state;
@@ -1379,6 +1351,9 @@ class Google_Service_Mirror_MenuValue extends Google_Model
 
 class Google_Service_Mirror_Notification extends Google_Collection
 {
+  protected $collection_key = 'userActions';
+  protected $internal_gapi_mappings = array(
+  );
   public $collection;
   public $itemId;
   public $operation;
@@ -1387,61 +1362,51 @@ class Google_Service_Mirror_Notification extends Google_Collection
   public $userToken;
   public $verifyToken;
 
+
   public function setCollection($collection)
   {
     $this->collection = $collection;
   }
-
   public function getCollection()
   {
     return $this->collection;
   }
-
   public function setItemId($itemId)
   {
     $this->itemId = $itemId;
   }
-
   public function getItemId()
   {
     return $this->itemId;
   }
-
   public function setOperation($operation)
   {
     $this->operation = $operation;
   }
-
   public function getOperation()
   {
     return $this->operation;
   }
-
   public function setUserActions($userActions)
   {
     $this->userActions = $userActions;
   }
-
   public function getUserActions()
   {
     return $this->userActions;
   }
-
   public function setUserToken($userToken)
   {
     $this->userToken = $userToken;
   }
-
   public function getUserToken()
   {
     return $this->userToken;
   }
-
   public function setVerifyToken($verifyToken)
   {
     $this->verifyToken = $verifyToken;
   }
-
   public function getVerifyToken()
   {
     return $this->verifyToken;
@@ -1450,24 +1415,24 @@ class Google_Service_Mirror_Notification extends Google_Collection
 
 class Google_Service_Mirror_NotificationConfig extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $deliveryTime;
   public $level;
+
 
   public function setDeliveryTime($deliveryTime)
   {
     $this->deliveryTime = $deliveryTime;
   }
-
   public function getDeliveryTime()
   {
     return $this->deliveryTime;
   }
-
   public function setLevel($level)
   {
     $this->level = $level;
   }
-
   public function getLevel()
   {
     return $this->level;
@@ -1476,35 +1441,33 @@ class Google_Service_Mirror_NotificationConfig extends Google_Model
 
 class Google_Service_Mirror_Setting extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
   public $value;
+
 
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -1513,6 +1476,9 @@ class Google_Service_Mirror_Setting extends Google_Model
 
 class Google_Service_Mirror_Subscription extends Google_Collection
 {
+  protected $collection_key = 'operation';
+  protected $internal_gapi_mappings = array(
+  );
   public $callbackUrl;
   public $collection;
   public $id;
@@ -1524,91 +1490,75 @@ class Google_Service_Mirror_Subscription extends Google_Collection
   public $userToken;
   public $verifyToken;
 
+
   public function setCallbackUrl($callbackUrl)
   {
     $this->callbackUrl = $callbackUrl;
   }
-
   public function getCallbackUrl()
   {
     return $this->callbackUrl;
   }
-
   public function setCollection($collection)
   {
     $this->collection = $collection;
   }
-
   public function getCollection()
   {
     return $this->collection;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNotification(Google_Service_Mirror_Notification $notification)
   {
     $this->notification = $notification;
   }
-
   public function getNotification()
   {
     return $this->notification;
   }
-
   public function setOperation($operation)
   {
     $this->operation = $operation;
   }
-
   public function getOperation()
   {
     return $this->operation;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
   }
-
   public function setUserToken($userToken)
   {
     $this->userToken = $userToken;
   }
-
   public function getUserToken()
   {
     return $this->userToken;
   }
-
   public function setVerifyToken($verifyToken)
   {
     $this->verifyToken = $verifyToken;
   }
-
   public function getVerifyToken()
   {
     return $this->verifyToken;
@@ -1617,25 +1567,26 @@ class Google_Service_Mirror_Subscription extends Google_Collection
 
 class Google_Service_Mirror_SubscriptionsListResponse extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Mirror_Subscription';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -1644,6 +1595,9 @@ class Google_Service_Mirror_SubscriptionsListResponse extends Google_Collection
 
 class Google_Service_Mirror_TimelineItem extends Google_Collection
 {
+  protected $collection_key = 'recipients';
+  protected $internal_gapi_mappings = array(
+  );
   protected $attachmentsType = 'Google_Service_Mirror_Attachment';
   protected $attachmentsDataType = 'array';
   public $bundleId;
@@ -1677,261 +1631,211 @@ class Google_Service_Mirror_TimelineItem extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setAttachments($attachments)
   {
     $this->attachments = $attachments;
   }
-
   public function getAttachments()
   {
     return $this->attachments;
   }
-
   public function setBundleId($bundleId)
   {
     $this->bundleId = $bundleId;
   }
-
   public function getBundleId()
   {
     return $this->bundleId;
   }
-
   public function setCanonicalUrl($canonicalUrl)
   {
     $this->canonicalUrl = $canonicalUrl;
   }
-
   public function getCanonicalUrl()
   {
     return $this->canonicalUrl;
   }
-
   public function setCreated($created)
   {
     $this->created = $created;
   }
-
   public function getCreated()
   {
     return $this->created;
   }
-
   public function setCreator(Google_Service_Mirror_Contact $creator)
   {
     $this->creator = $creator;
   }
-
   public function getCreator()
   {
     return $this->creator;
   }
-
   public function setDisplayTime($displayTime)
   {
     $this->displayTime = $displayTime;
   }
-
   public function getDisplayTime()
   {
     return $this->displayTime;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setHtml($html)
   {
     $this->html = $html;
   }
-
   public function getHtml()
   {
     return $this->html;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setInReplyTo($inReplyTo)
   {
     $this->inReplyTo = $inReplyTo;
   }
-
   public function getInReplyTo()
   {
     return $this->inReplyTo;
   }
-
   public function setIsBundleCover($isBundleCover)
   {
     $this->isBundleCover = $isBundleCover;
   }
-
   public function getIsBundleCover()
   {
     return $this->isBundleCover;
   }
-
   public function setIsDeleted($isDeleted)
   {
     $this->isDeleted = $isDeleted;
   }
-
   public function getIsDeleted()
   {
     return $this->isDeleted;
   }
-
   public function setIsPinned($isPinned)
   {
     $this->isPinned = $isPinned;
   }
-
   public function getIsPinned()
   {
     return $this->isPinned;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLocation(Google_Service_Mirror_Location $location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setMenuItems($menuItems)
   {
     $this->menuItems = $menuItems;
   }
-
   public function getMenuItems()
   {
     return $this->menuItems;
   }
-
   public function setNotification(Google_Service_Mirror_NotificationConfig $notification)
   {
     $this->notification = $notification;
   }
-
   public function getNotification()
   {
     return $this->notification;
   }
-
   public function setPinScore($pinScore)
   {
     $this->pinScore = $pinScore;
   }
-
   public function getPinScore()
   {
     return $this->pinScore;
   }
-
   public function setRecipients($recipients)
   {
     $this->recipients = $recipients;
   }
-
   public function getRecipients()
   {
     return $this->recipients;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setSourceItemId($sourceItemId)
   {
     $this->sourceItemId = $sourceItemId;
   }
-
   public function getSourceItemId()
   {
     return $this->sourceItemId;
   }
-
   public function setSpeakableText($speakableText)
   {
     $this->speakableText = $speakableText;
   }
-
   public function getSpeakableText()
   {
     return $this->speakableText;
   }
-
   public function setSpeakableType($speakableType)
   {
     $this->speakableType = $speakableType;
   }
-
   public function getSpeakableType()
   {
     return $this->speakableType;
   }
-
   public function setText($text)
   {
     $this->text = $text;
   }
-
   public function getText()
   {
     return $this->text;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -1940,36 +1844,35 @@ class Google_Service_Mirror_TimelineItem extends Google_Collection
 
 class Google_Service_Mirror_TimelineListResponse extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Mirror_TimelineItem';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -1978,24 +1881,24 @@ class Google_Service_Mirror_TimelineListResponse extends Google_Collection
 
 class Google_Service_Mirror_UserAction extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $payload;
   public $type;
+
 
   public function setPayload($payload)
   {
     $this->payload = $payload;
   }
-
   public function getPayload()
   {
     return $this->payload;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -2004,24 +1907,24 @@ class Google_Service_Mirror_UserAction extends Google_Model
 
 class Google_Service_Mirror_UserData extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $key;
   public $value;
+
 
   public function setKey($key)
   {
     $this->key = $key;
   }
-
   public function getKey()
   {
     return $this->key;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
