@@ -19,8 +19,7 @@
  * Service definition for Plus (v1).
  *
  * <p>
- * The Google+ API enables developers to build on top of the Google+ platform.
- * </p>
+ * The Google+ API enables developers to build on top of the Google+ platform.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,13 +31,17 @@
 class Google_Service_Plus extends Google_Service
 {
   /** Know your basic profile info and list of people in your circles.. */
-  const PLUS_LOGIN = "https://www.googleapis.com/auth/plus.login";
+  const PLUS_LOGIN =
+      "https://www.googleapis.com/auth/plus.login";
   /** Know who you are on Google. */
-  const PLUS_ME = "https://www.googleapis.com/auth/plus.me";
+  const PLUS_ME =
+      "https://www.googleapis.com/auth/plus.me";
   /** View your email address. */
-  const USERINFO_EMAIL = "https://www.googleapis.com/auth/userinfo.email";
-  /** View basic information about your account. */
-  const USERINFO_PROFILE = "https://www.googleapis.com/auth/userinfo.profile";
+  const USERINFO_EMAIL =
+      "https://www.googleapis.com/auth/userinfo.email";
+  /** View your basic profile info. */
+  const USERINFO_PROFILE =
+      "https://www.googleapis.com/auth/userinfo.profile";
 
   public $activities;
   public $comments;
@@ -54,6 +57,7 @@ class Google_Service_Plus extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'plus/v1/';
     $this->version = 'v1';
     $this->serviceName = 'plus';
@@ -225,16 +229,6 @@ class Google_Service_Plus extends Google_Service
                   'type' => 'string',
                 ),
               ),
-            ),'remove' => array(
-              'path' => 'moments/{id}',
-              'httpMethod' => 'DELETE',
-              'parameters' => array(
-                'id' => array(
-                  'location' => 'path',
-                  'type' => 'string',
-                  'required' => true,
-                ),
-              ),
             ),
           )
         )
@@ -349,8 +343,7 @@ class Google_Service_Plus_Activities_Resource extends Google_Service_Resource
   /**
    * Get an activity. (activities.get)
    *
-   * @param string $activityId
-   * The ID of the activity to get.
+   * @param string $activityId The ID of the activity to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Plus_Activity
    */
@@ -360,23 +353,22 @@ class Google_Service_Plus_Activities_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Plus_Activity");
   }
+
   /**
    * List all of the activities in the specified collection for a particular user.
    * (activities.listActivities)
    *
-   * @param string $userId
-   * The ID of the user to get activities for. The special value "me" can be used to indicate the
-    * authenticated user.
-   * @param string $collection
-   * The collection of activities to list.
+   * @param string $userId The ID of the user to get activities for. The special
+   * value "me" can be used to indicate the authenticated user.
+   * @param string $collection The collection of activities to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of activities to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of activities to include in
+   * the response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_Plus_ActivityFeed
    */
   public function listActivities($userId, $collection, $optParams = array())
@@ -385,24 +377,23 @@ class Google_Service_Plus_Activities_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Plus_ActivityFeed");
   }
+
   /**
    * Search public activities. (activities.search)
    *
-   * @param string $query
-   * Full-text search query string.
+   * @param string $query Full-text search query string.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string orderBy
-   * Specifies how to order search results.
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response. This
-    * token can be of any length.
-   * @opt_param string maxResults
-   * The maximum number of activities to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
-   * @opt_param string language
-   * Specify the preferred language to search with. See search language codes for available values.
+   * @opt_param string orderBy Specifies how to order search results.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response. This
+   * token can be of any length.
+   * @opt_param string maxResults The maximum number of activities to include in
+   * the response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
+   * @opt_param string language Specify the preferred language to search with. See
+   * search language codes for available values.
    * @return Google_Service_Plus_ActivityFeed
    */
   public function search($query, $optParams = array())
@@ -427,8 +418,7 @@ class Google_Service_Plus_Comments_Resource extends Google_Service_Resource
   /**
    * Get a comment. (comments.get)
    *
-   * @param string $commentId
-   * The ID of the comment to get.
+   * @param string $commentId The ID of the comment to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Plus_Comment
    */
@@ -438,21 +428,20 @@ class Google_Service_Plus_Comments_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Plus_Comment");
   }
+
   /**
    * List all of the comments for an activity. (comments.listComments)
    *
-   * @param string $activityId
-   * The ID of the activity to get comments for.
+   * @param string $activityId The ID of the activity to get comments for.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string sortOrder
-   * The order in which to sort the list of comments.
-   * @opt_param string maxResults
-   * The maximum number of comments to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string sortOrder The order in which to sort the list of comments.
+   * @opt_param string maxResults The maximum number of comments to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_Plus_CommentFeed
    */
   public function listComments($activityId, $optParams = array())
@@ -475,19 +464,17 @@ class Google_Service_Plus_Moments_Resource extends Google_Service_Resource
 {
 
   /**
-   * Record a moment representing a user's activity such as making a purchase or
+   * Record a moment representing a user's action such as making a purchase or
    * commenting on a blog. (moments.insert)
    *
-   * @param string $userId
-   * The ID of the user to record activities for. The only valid values are "me" and the ID of the
-    * authenticated user.
-   * @param string $collection
-   * The collection to which to write moments.
+   * @param string $userId The ID of the user to record actions for. The only
+   * valid values are "me" and the ID of the authenticated user.
+   * @param string $collection The collection to which to write moments.
    * @param Google_Moment $postBody
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool debug
-   * Return the moment as written. Should be used only for debugging.
+   * @opt_param bool debug Return the moment as written. Should be used only for
+   * debugging.
    * @return Google_Service_Plus_Moment
    */
   public function insert($userId, $collection, Google_Service_Plus_Moment $postBody, $optParams = array())
@@ -496,26 +483,24 @@ class Google_Service_Plus_Moments_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Plus_Moment");
   }
+
   /**
    * List all of the moments for a particular user. (moments.listMoments)
    *
-   * @param string $userId
-   * The ID of the user to get moments for. The special value "me" can be used to indicate the
-    * authenticated user.
-   * @param string $collection
-   * The collection of moments to list.
+   * @param string $userId The ID of the user to get moments for. The special
+   * value "me" can be used to indicate the authenticated user.
+   * @param string $collection The collection of moments to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string maxResults
-   * The maximum number of moments to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string targetUrl
-   * Only moments containing this targetUrl will be returned.
-   * @opt_param string type
-   * Only moments of this type will be returned.
+   * @opt_param string maxResults The maximum number of moments to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string targetUrl Only moments containing this targetUrl will be
+   * returned.
+   * @opt_param string type Only moments of this type will be returned.
    * @return Google_Service_Plus_MomentsFeed
    */
   public function listMoments($userId, $collection, $optParams = array())
@@ -523,19 +508,6 @@ class Google_Service_Plus_Moments_Resource extends Google_Service_Resource
     $params = array('userId' => $userId, 'collection' => $collection);
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Plus_MomentsFeed");
-  }
-  /**
-   * Delete a moment. (moments.remove)
-   *
-   * @param string $id
-   * The ID of the moment to delete.
-   * @param array $optParams Optional parameters.
-   */
-  public function remove($id, $optParams = array())
-  {
-    $params = array('id' => $id);
-    $params = array_merge($params, $optParams);
-    return $this->call('remove', array($params));
   }
 }
 
@@ -555,9 +527,8 @@ class Google_Service_Plus_People_Resource extends Google_Service_Resource
    * https://www.googleapis.com/auth/plus.login, this method is guaranteed to
    * return ageRange and language. (people.get)
    *
-   * @param string $userId
-   * The ID of the person to get the profile for. The special value "me" can be used to indicate the
-    * authenticated user.
+   * @param string $userId The ID of the person to get the profile for. The
+   * special value "me" can be used to indicate the authenticated user.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Plus_Person
    */
@@ -567,24 +538,22 @@ class Google_Service_Plus_People_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Plus_Person");
   }
+
   /**
    * List all of the people in the specified collection. (people.listPeople)
    *
-   * @param string $userId
-   * Get the collection of people for the person identified. Use "me" to indicate the authenticated
-    * user.
-   * @param string $collection
-   * The collection of people to list.
+   * @param string $userId Get the collection of people for the person identified.
+   * Use "me" to indicate the authenticated user.
+   * @param string $collection The collection of people to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string orderBy
-   * The order to return people in.
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of people to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string orderBy The order to return people in.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of people to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_Plus_PeopleFeed
    */
   public function listPeople($userId, $collection, $optParams = array())
@@ -593,22 +562,22 @@ class Google_Service_Plus_People_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Plus_PeopleFeed");
   }
+
   /**
    * List all of the people in the specified collection for a particular activity.
    * (people.listByActivity)
    *
-   * @param string $activityId
-   * The ID of the activity to get the list of people for.
-   * @param string $collection
-   * The collection of people to list.
+   * @param string $activityId The ID of the activity to get the list of people
+   * for.
+   * @param string $collection The collection of people to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of people to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of people to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_Plus_PeopleFeed
    */
   public function listByActivity($activityId, $collection, $optParams = array())
@@ -617,22 +586,23 @@ class Google_Service_Plus_People_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('listByActivity', array($params), "Google_Service_Plus_PeopleFeed");
   }
+
   /**
    * Search all public profiles. (people.search)
    *
-   * @param string $query
-   * Specify a query string for full text search of public text in all profiles.
+   * @param string $query Specify a query string for full text search of public
+   * text in all profiles.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response. This
-    * token can be of any length.
-   * @opt_param string maxResults
-   * The maximum number of people to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
-   * @opt_param string language
-   * Specify the preferred language to search with. See search language codes for available values.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response. This
+   * token can be of any length.
+   * @opt_param string maxResults The maximum number of people to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
+   * @opt_param string language Specify the preferred language to search with. See
+   * search language codes for available values.
    * @return Google_Service_Plus_PeopleFeed
    */
   public function search($query, $optParams = array())
@@ -648,36 +618,35 @@ class Google_Service_Plus_People_Resource extends Google_Service_Resource
 
 class Google_Service_Plus_Acl extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $description;
   protected $itemsType = 'Google_Service_Plus_PlusAclentryResource';
   protected $itemsDataType = 'array';
   public $kind;
 
+
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -686,6 +655,8 @@ class Google_Service_Plus_Acl extends Google_Collection
 
 class Google_Service_Plus_Activity extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $accessType = 'Google_Service_Plus_Acl';
   protected $accessDataType = '';
   protected $actorType = 'Google_Service_Plus_ActivityActor';
@@ -712,201 +683,163 @@ class Google_Service_Plus_Activity extends Google_Model
   public $url;
   public $verb;
 
+
   public function setAccess(Google_Service_Plus_Acl $access)
   {
     $this->access = $access;
   }
-
   public function getAccess()
   {
     return $this->access;
   }
-
   public function setActor(Google_Service_Plus_ActivityActor $actor)
   {
     $this->actor = $actor;
   }
-
   public function getActor()
   {
     return $this->actor;
   }
-
   public function setAddress($address)
   {
     $this->address = $address;
   }
-
   public function getAddress()
   {
     return $this->address;
   }
-
   public function setAnnotation($annotation)
   {
     $this->annotation = $annotation;
   }
-
   public function getAnnotation()
   {
     return $this->annotation;
   }
-
   public function setCrosspostSource($crosspostSource)
   {
     $this->crosspostSource = $crosspostSource;
   }
-
   public function getCrosspostSource()
   {
     return $this->crosspostSource;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setGeocode($geocode)
   {
     $this->geocode = $geocode;
   }
-
   public function getGeocode()
   {
     return $this->geocode;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLocation(Google_Service_Plus_Place $location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setObject(Google_Service_Plus_ActivityObject $object)
   {
     $this->object = $object;
   }
-
   public function getObject()
   {
     return $this->object;
   }
-
   public function setPlaceId($placeId)
   {
     $this->placeId = $placeId;
   }
-
   public function getPlaceId()
   {
     return $this->placeId;
   }
-
   public function setPlaceName($placeName)
   {
     $this->placeName = $placeName;
   }
-
   public function getPlaceName()
   {
     return $this->placeName;
   }
-
   public function setProvider(Google_Service_Plus_ActivityProvider $provider)
   {
     $this->provider = $provider;
   }
-
   public function getProvider()
   {
     return $this->provider;
   }
-
   public function setPublished($published)
   {
     $this->published = $published;
   }
-
   public function getPublished()
   {
     return $this->published;
   }
-
   public function setRadius($radius)
   {
     $this->radius = $radius;
   }
-
   public function getRadius()
   {
     return $this->radius;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setVerb($verb)
   {
     $this->verb = $verb;
   }
-
   public function getVerb()
   {
     return $this->verb;
@@ -915,6 +848,10 @@ class Google_Service_Plus_Activity extends Google_Model
 
 class Google_Service_Plus_ActivityActor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  protected $clientSpecificActorInfoType = 'Google_Service_Plus_ActivityActorClientSpecificActorInfo';
+  protected $clientSpecificActorInfoDataType = '';
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_Plus_ActivityActorImage';
@@ -922,67 +859,114 @@ class Google_Service_Plus_ActivityActor extends Google_Model
   protected $nameType = 'Google_Service_Plus_ActivityActorName';
   protected $nameDataType = '';
   public $url;
+  protected $verificationType = 'Google_Service_Plus_ActivityActorVerification';
+  protected $verificationDataType = '';
 
+
+  public function setClientSpecificActorInfo(Google_Service_Plus_ActivityActorClientSpecificActorInfo $clientSpecificActorInfo)
+  {
+    $this->clientSpecificActorInfo = $clientSpecificActorInfo;
+  }
+  public function getClientSpecificActorInfo()
+  {
+    return $this->clientSpecificActorInfo;
+  }
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_Plus_ActivityActorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setName(Google_Service_Plus_ActivityActorName $name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
+  }
+  public function setVerification(Google_Service_Plus_ActivityActorVerification $verification)
+  {
+    $this->verification = $verification;
+  }
+  public function getVerification()
+  {
+    return $this->verification;
+  }
+}
+
+class Google_Service_Plus_ActivityActorClientSpecificActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  protected $youtubeActorInfoType = 'Google_Service_Plus_ActivityActorClientSpecificActorInfoYoutubeActorInfo';
+  protected $youtubeActorInfoDataType = '';
+
+
+  public function setYoutubeActorInfo(Google_Service_Plus_ActivityActorClientSpecificActorInfoYoutubeActorInfo $youtubeActorInfo)
+  {
+    $this->youtubeActorInfo = $youtubeActorInfo;
+  }
+  public function getYoutubeActorInfo()
+  {
+    return $this->youtubeActorInfo;
+  }
+}
+
+class Google_Service_Plus_ActivityActorClientSpecificActorInfoYoutubeActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $channelId;
+
+
+  public function setChannelId($channelId)
+  {
+    $this->channelId = $channelId;
+  }
+  public function getChannelId()
+  {
+    return $this->channelId;
   }
 }
 
 class Google_Service_Plus_ActivityActorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -991,32 +975,52 @@ class Google_Service_Plus_ActivityActorImage extends Google_Model
 
 class Google_Service_Plus_ActivityActorName extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $familyName;
   public $givenName;
+
 
   public function setFamilyName($familyName)
   {
     $this->familyName = $familyName;
   }
-
   public function getFamilyName()
   {
     return $this->familyName;
   }
-
   public function setGivenName($givenName)
   {
     $this->givenName = $givenName;
   }
-
   public function getGivenName()
   {
     return $this->givenName;
   }
 }
 
+class Google_Service_Plus_ActivityActorVerification extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $adHocVerified;
+
+
+  public function setAdHocVerified($adHocVerified)
+  {
+    $this->adHocVerified = $adHocVerified;
+  }
+  public function getAdHocVerified()
+  {
+    return $this->adHocVerified;
+  }
+}
+
 class Google_Service_Plus_ActivityFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   public $id;
   protected $itemsType = 'Google_Service_Plus_Activity';
@@ -1028,91 +1032,75 @@ class Google_Service_Plus_ActivityFeed extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextLink($nextLink)
   {
     $this->nextLink = $nextLink;
   }
-
   public function getNextLink()
   {
     return $this->nextLink;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -1121,6 +1109,9 @@ class Google_Service_Plus_ActivityFeed extends Google_Collection
 
 class Google_Service_Plus_ActivityObject extends Google_Collection
 {
+  protected $collection_key = 'attachments';
+  protected $internal_gapi_mappings = array(
+  );
   protected $actorType = 'Google_Service_Plus_ActivityObjectActor';
   protected $actorDataType = '';
   protected $attachmentsType = 'Google_Service_Plus_ActivityObjectAttachments';
@@ -1137,101 +1128,83 @@ class Google_Service_Plus_ActivityObject extends Google_Collection
   protected $resharersDataType = '';
   public $url;
 
+
   public function setActor(Google_Service_Plus_ActivityObjectActor $actor)
   {
     $this->actor = $actor;
   }
-
   public function getActor()
   {
     return $this->actor;
   }
-
   public function setAttachments($attachments)
   {
     $this->attachments = $attachments;
   }
-
   public function getAttachments()
   {
     return $this->attachments;
   }
-
   public function setContent($content)
   {
     $this->content = $content;
   }
-
   public function getContent()
   {
     return $this->content;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setOriginalContent($originalContent)
   {
     $this->originalContent = $originalContent;
   }
-
   public function getOriginalContent()
   {
     return $this->originalContent;
   }
-
   public function setPlusoners(Google_Service_Plus_ActivityObjectPlusoners $plusoners)
   {
     $this->plusoners = $plusoners;
   }
-
   public function getPlusoners()
   {
     return $this->plusoners;
   }
-
   public function setReplies(Google_Service_Plus_ActivityObjectReplies $replies)
   {
     $this->replies = $replies;
   }
-
   public function getReplies()
   {
     return $this->replies;
   }
-
   public function setResharers(Google_Service_Plus_ActivityObjectResharers $resharers)
   {
     $this->resharers = $resharers;
   }
-
   public function getResharers()
   {
     return $this->resharers;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1240,70 +1213,143 @@ class Google_Service_Plus_ActivityObject extends Google_Collection
 
 class Google_Service_Plus_ActivityObjectActor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  protected $clientSpecificActorInfoType = 'Google_Service_Plus_ActivityObjectActorClientSpecificActorInfo';
+  protected $clientSpecificActorInfoDataType = '';
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_Plus_ActivityObjectActorImage';
   protected $imageDataType = '';
   public $url;
+  protected $verificationType = 'Google_Service_Plus_ActivityObjectActorVerification';
+  protected $verificationDataType = '';
 
+
+  public function setClientSpecificActorInfo(Google_Service_Plus_ActivityObjectActorClientSpecificActorInfo $clientSpecificActorInfo)
+  {
+    $this->clientSpecificActorInfo = $clientSpecificActorInfo;
+  }
+  public function getClientSpecificActorInfo()
+  {
+    return $this->clientSpecificActorInfo;
+  }
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_Plus_ActivityObjectActorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
+  }
+  public function setVerification(Google_Service_Plus_ActivityObjectActorVerification $verification)
+  {
+    $this->verification = $verification;
+  }
+  public function getVerification()
+  {
+    return $this->verification;
+  }
+}
+
+class Google_Service_Plus_ActivityObjectActorClientSpecificActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  protected $youtubeActorInfoType = 'Google_Service_Plus_ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo';
+  protected $youtubeActorInfoDataType = '';
+
+
+  public function setYoutubeActorInfo(Google_Service_Plus_ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo $youtubeActorInfo)
+  {
+    $this->youtubeActorInfo = $youtubeActorInfo;
+  }
+  public function getYoutubeActorInfo()
+  {
+    return $this->youtubeActorInfo;
+  }
+}
+
+class Google_Service_Plus_ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $channelId;
+
+
+  public function setChannelId($channelId)
+  {
+    $this->channelId = $channelId;
+  }
+  public function getChannelId()
+  {
+    return $this->channelId;
   }
 }
 
 class Google_Service_Plus_ActivityObjectActorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
 }
 
+class Google_Service_Plus_ActivityObjectActorVerification extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $adHocVerified;
+
+
+  public function setAdHocVerified($adHocVerified)
+  {
+    $this->adHocVerified = $adHocVerified;
+  }
+  public function getAdHocVerified()
+  {
+    return $this->adHocVerified;
+  }
+}
+
 class Google_Service_Plus_ActivityObjectAttachments extends Google_Collection
 {
+  protected $collection_key = 'thumbnails';
+  protected $internal_gapi_mappings = array(
+  );
   public $content;
   public $displayName;
   protected $embedType = 'Google_Service_Plus_ActivityObjectAttachmentsEmbed';
@@ -1318,91 +1364,75 @@ class Google_Service_Plus_ActivityObjectAttachments extends Google_Collection
   protected $thumbnailsDataType = 'array';
   public $url;
 
+
   public function setContent($content)
   {
     $this->content = $content;
   }
-
   public function getContent()
   {
     return $this->content;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setEmbed(Google_Service_Plus_ActivityObjectAttachmentsEmbed $embed)
   {
     $this->embed = $embed;
   }
-
   public function getEmbed()
   {
     return $this->embed;
   }
-
   public function setFullImage(Google_Service_Plus_ActivityObjectAttachmentsFullImage $fullImage)
   {
     $this->fullImage = $fullImage;
   }
-
   public function getFullImage()
   {
     return $this->fullImage;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_Plus_ActivityObjectAttachmentsImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setThumbnails($thumbnails)
   {
     $this->thumbnails = $thumbnails;
   }
-
   public function getThumbnails()
   {
     return $this->thumbnails;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1411,24 +1441,24 @@ class Google_Service_Plus_ActivityObjectAttachments extends Google_Collection
 
 class Google_Service_Plus_ActivityObjectAttachmentsEmbed extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $type;
   public $url;
+
 
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1437,46 +1467,42 @@ class Google_Service_Plus_ActivityObjectAttachmentsEmbed extends Google_Model
 
 class Google_Service_Plus_ActivityObjectAttachmentsFullImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -1485,46 +1511,42 @@ class Google_Service_Plus_ActivityObjectAttachmentsFullImage extends Google_Mode
 
 class Google_Service_Plus_ActivityObjectAttachmentsImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -1533,36 +1555,34 @@ class Google_Service_Plus_ActivityObjectAttachmentsImage extends Google_Model
 
 class Google_Service_Plus_ActivityObjectAttachmentsThumbnails extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $description;
   protected $imageType = 'Google_Service_Plus_ActivityObjectAttachmentsThumbnailsImage';
   protected $imageDataType = '';
   public $url;
 
+
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setImage(Google_Service_Plus_ActivityObjectAttachmentsThumbnailsImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1571,46 +1591,42 @@ class Google_Service_Plus_ActivityObjectAttachmentsThumbnails extends Google_Mod
 
 class Google_Service_Plus_ActivityObjectAttachmentsThumbnailsImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -1619,24 +1635,24 @@ class Google_Service_Plus_ActivityObjectAttachmentsThumbnailsImage extends Googl
 
 class Google_Service_Plus_ActivityObjectPlusoners extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $selfLink;
   public $totalItems;
+
 
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -1645,24 +1661,24 @@ class Google_Service_Plus_ActivityObjectPlusoners extends Google_Model
 
 class Google_Service_Plus_ActivityObjectReplies extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $selfLink;
   public $totalItems;
+
 
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -1671,24 +1687,24 @@ class Google_Service_Plus_ActivityObjectReplies extends Google_Model
 
 class Google_Service_Plus_ActivityObjectResharers extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $selfLink;
   public $totalItems;
+
 
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -1697,13 +1713,15 @@ class Google_Service_Plus_ActivityObjectResharers extends Google_Model
 
 class Google_Service_Plus_ActivityProvider extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $title;
+
 
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
@@ -1712,6 +1730,9 @@ class Google_Service_Plus_ActivityProvider extends Google_Model
 
 class Google_Service_Plus_Comment extends Google_Collection
 {
+  protected $collection_key = 'inReplyTo';
+  protected $internal_gapi_mappings = array(
+  );
   protected $actorType = 'Google_Service_Plus_CommentActor';
   protected $actorDataType = '';
   public $etag;
@@ -1728,111 +1749,91 @@ class Google_Service_Plus_Comment extends Google_Collection
   public $updated;
   public $verb;
 
+
   public function setActor(Google_Service_Plus_CommentActor $actor)
   {
     $this->actor = $actor;
   }
-
   public function getActor()
   {
     return $this->actor;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setInReplyTo($inReplyTo)
   {
     $this->inReplyTo = $inReplyTo;
   }
-
   public function getInReplyTo()
   {
     return $this->inReplyTo;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setObject(Google_Service_Plus_CommentObject $object)
   {
     $this->object = $object;
   }
-
   public function getObject()
   {
     return $this->object;
   }
-
   public function setPlusoners(Google_Service_Plus_CommentPlusoners $plusoners)
   {
     $this->plusoners = $plusoners;
   }
-
   public function getPlusoners()
   {
     return $this->plusoners;
   }
-
   public function setPublished($published)
   {
     $this->published = $published;
   }
-
   public function getPublished()
   {
     return $this->published;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
   }
-
   public function setVerb($verb)
   {
     $this->verb = $verb;
   }
-
   public function getVerb()
   {
     return $this->verb;
@@ -1841,70 +1842,143 @@ class Google_Service_Plus_Comment extends Google_Collection
 
 class Google_Service_Plus_CommentActor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  protected $clientSpecificActorInfoType = 'Google_Service_Plus_CommentActorClientSpecificActorInfo';
+  protected $clientSpecificActorInfoDataType = '';
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_Plus_CommentActorImage';
   protected $imageDataType = '';
   public $url;
+  protected $verificationType = 'Google_Service_Plus_CommentActorVerification';
+  protected $verificationDataType = '';
 
+
+  public function setClientSpecificActorInfo(Google_Service_Plus_CommentActorClientSpecificActorInfo $clientSpecificActorInfo)
+  {
+    $this->clientSpecificActorInfo = $clientSpecificActorInfo;
+  }
+  public function getClientSpecificActorInfo()
+  {
+    return $this->clientSpecificActorInfo;
+  }
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_Plus_CommentActorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
+  }
+  public function setVerification(Google_Service_Plus_CommentActorVerification $verification)
+  {
+    $this->verification = $verification;
+  }
+  public function getVerification()
+  {
+    return $this->verification;
+  }
+}
+
+class Google_Service_Plus_CommentActorClientSpecificActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  protected $youtubeActorInfoType = 'Google_Service_Plus_CommentActorClientSpecificActorInfoYoutubeActorInfo';
+  protected $youtubeActorInfoDataType = '';
+
+
+  public function setYoutubeActorInfo(Google_Service_Plus_CommentActorClientSpecificActorInfoYoutubeActorInfo $youtubeActorInfo)
+  {
+    $this->youtubeActorInfo = $youtubeActorInfo;
+  }
+  public function getYoutubeActorInfo()
+  {
+    return $this->youtubeActorInfo;
+  }
+}
+
+class Google_Service_Plus_CommentActorClientSpecificActorInfoYoutubeActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $channelId;
+
+
+  public function setChannelId($channelId)
+  {
+    $this->channelId = $channelId;
+  }
+  public function getChannelId()
+  {
+    return $this->channelId;
   }
 }
 
 class Google_Service_Plus_CommentActorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
 }
 
+class Google_Service_Plus_CommentActorVerification extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $adHocVerified;
+
+
+  public function setAdHocVerified($adHocVerified)
+  {
+    $this->adHocVerified = $adHocVerified;
+  }
+  public function getAdHocVerified()
+  {
+    return $this->adHocVerified;
+  }
+}
+
 class Google_Service_Plus_CommentFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   public $id;
   protected $itemsType = 'Google_Service_Plus_Comment';
@@ -1915,81 +1989,67 @@ class Google_Service_Plus_CommentFeed extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextLink($nextLink)
   {
     $this->nextLink = $nextLink;
   }
-
   public function getNextLink()
   {
     return $this->nextLink;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -1998,24 +2058,24 @@ class Google_Service_Plus_CommentFeed extends Google_Collection
 
 class Google_Service_Plus_CommentInReplyTo extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $url;
+
 
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -2024,35 +2084,33 @@ class Google_Service_Plus_CommentInReplyTo extends Google_Model
 
 class Google_Service_Plus_CommentObject extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $content;
   public $objectType;
   public $originalContent;
+
 
   public function setContent($content)
   {
     $this->content = $content;
   }
-
   public function getContent()
   {
     return $this->content;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setOriginalContent($originalContent)
   {
     $this->originalContent = $originalContent;
   }
-
   public function getOriginalContent()
   {
     return $this->originalContent;
@@ -2061,13 +2119,15 @@ class Google_Service_Plus_CommentObject extends Google_Model
 
 class Google_Service_Plus_CommentPlusoners extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $totalItems;
+
 
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2076,6 +2136,10 @@ class Google_Service_Plus_CommentPlusoners extends Google_Model
 
 class Google_Service_Plus_ItemScope extends Google_Collection
 {
+  protected $collection_key = 'performers';
+  protected $internal_gapi_mappings = array(
+        "associatedMedia" => "associated_media",
+  );
   protected $aboutType = 'Google_Service_Plus_ItemScope';
   protected $aboutDataType = '';
   public $additionalName;
@@ -2147,551 +2211,443 @@ class Google_Service_Plus_ItemScope extends Google_Collection
   public $width;
   public $worstRating;
 
+
   public function setAbout(Google_Service_Plus_ItemScope $about)
   {
     $this->about = $about;
   }
-
   public function getAbout()
   {
     return $this->about;
   }
-
   public function setAdditionalName($additionalName)
   {
     $this->additionalName = $additionalName;
   }
-
   public function getAdditionalName()
   {
     return $this->additionalName;
   }
-
   public function setAddress(Google_Service_Plus_ItemScope $address)
   {
     $this->address = $address;
   }
-
   public function getAddress()
   {
     return $this->address;
   }
-
   public function setAddressCountry($addressCountry)
   {
     $this->addressCountry = $addressCountry;
   }
-
   public function getAddressCountry()
   {
     return $this->addressCountry;
   }
-
   public function setAddressLocality($addressLocality)
   {
     $this->addressLocality = $addressLocality;
   }
-
   public function getAddressLocality()
   {
     return $this->addressLocality;
   }
-
   public function setAddressRegion($addressRegion)
   {
     $this->addressRegion = $addressRegion;
   }
-
   public function getAddressRegion()
   {
     return $this->addressRegion;
   }
-
   public function setAssociatedMedia($associatedMedia)
   {
     $this->associatedMedia = $associatedMedia;
   }
-
   public function getAssociatedMedia()
   {
     return $this->associatedMedia;
   }
-
   public function setAttendeeCount($attendeeCount)
   {
     $this->attendeeCount = $attendeeCount;
   }
-
   public function getAttendeeCount()
   {
     return $this->attendeeCount;
   }
-
   public function setAttendees($attendees)
   {
     $this->attendees = $attendees;
   }
-
   public function getAttendees()
   {
     return $this->attendees;
   }
-
   public function setAudio(Google_Service_Plus_ItemScope $audio)
   {
     $this->audio = $audio;
   }
-
   public function getAudio()
   {
     return $this->audio;
   }
-
   public function setAuthor($author)
   {
     $this->author = $author;
   }
-
   public function getAuthor()
   {
     return $this->author;
   }
-
   public function setBestRating($bestRating)
   {
     $this->bestRating = $bestRating;
   }
-
   public function getBestRating()
   {
     return $this->bestRating;
   }
-
   public function setBirthDate($birthDate)
   {
     $this->birthDate = $birthDate;
   }
-
   public function getBirthDate()
   {
     return $this->birthDate;
   }
-
   public function setByArtist(Google_Service_Plus_ItemScope $byArtist)
   {
     $this->byArtist = $byArtist;
   }
-
   public function getByArtist()
   {
     return $this->byArtist;
   }
-
   public function setCaption($caption)
   {
     $this->caption = $caption;
   }
-
   public function getCaption()
   {
     return $this->caption;
   }
-
   public function setContentSize($contentSize)
   {
     $this->contentSize = $contentSize;
   }
-
   public function getContentSize()
   {
     return $this->contentSize;
   }
-
   public function setContentUrl($contentUrl)
   {
     $this->contentUrl = $contentUrl;
   }
-
   public function getContentUrl()
   {
     return $this->contentUrl;
   }
-
   public function setContributor($contributor)
   {
     $this->contributor = $contributor;
   }
-
   public function getContributor()
   {
     return $this->contributor;
   }
-
   public function setDateCreated($dateCreated)
   {
     $this->dateCreated = $dateCreated;
   }
-
   public function getDateCreated()
   {
     return $this->dateCreated;
   }
-
   public function setDateModified($dateModified)
   {
     $this->dateModified = $dateModified;
   }
-
   public function getDateModified()
   {
     return $this->dateModified;
   }
-
   public function setDatePublished($datePublished)
   {
     $this->datePublished = $datePublished;
   }
-
   public function getDatePublished()
   {
     return $this->datePublished;
   }
-
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setDuration($duration)
   {
     $this->duration = $duration;
   }
-
   public function getDuration()
   {
     return $this->duration;
   }
-
   public function setEmbedUrl($embedUrl)
   {
     $this->embedUrl = $embedUrl;
   }
-
   public function getEmbedUrl()
   {
     return $this->embedUrl;
   }
-
   public function setEndDate($endDate)
   {
     $this->endDate = $endDate;
   }
-
   public function getEndDate()
   {
     return $this->endDate;
   }
-
   public function setFamilyName($familyName)
   {
     $this->familyName = $familyName;
   }
-
   public function getFamilyName()
   {
     return $this->familyName;
   }
-
   public function setGender($gender)
   {
     $this->gender = $gender;
   }
-
   public function getGender()
   {
     return $this->gender;
   }
-
   public function setGeo(Google_Service_Plus_ItemScope $geo)
   {
     $this->geo = $geo;
   }
-
   public function getGeo()
   {
     return $this->geo;
   }
-
   public function setGivenName($givenName)
   {
     $this->givenName = $givenName;
   }
-
   public function getGivenName()
   {
     return $this->givenName;
   }
-
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage($image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setInAlbum(Google_Service_Plus_ItemScope $inAlbum)
   {
     $this->inAlbum = $inAlbum;
   }
-
   public function getInAlbum()
   {
     return $this->inAlbum;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLatitude($latitude)
   {
     $this->latitude = $latitude;
   }
-
   public function getLatitude()
   {
     return $this->latitude;
   }
-
   public function setLocation(Google_Service_Plus_ItemScope $location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setLongitude($longitude)
   {
     $this->longitude = $longitude;
   }
-
   public function getLongitude()
   {
     return $this->longitude;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setPartOfTVSeries(Google_Service_Plus_ItemScope $partOfTVSeries)
   {
     $this->partOfTVSeries = $partOfTVSeries;
   }
-
   public function getPartOfTVSeries()
   {
     return $this->partOfTVSeries;
   }
-
   public function setPerformers($performers)
   {
     $this->performers = $performers;
   }
-
   public function getPerformers()
   {
     return $this->performers;
   }
-
   public function setPlayerType($playerType)
   {
     $this->playerType = $playerType;
   }
-
   public function getPlayerType()
   {
     return $this->playerType;
   }
-
   public function setPostOfficeBoxNumber($postOfficeBoxNumber)
   {
     $this->postOfficeBoxNumber = $postOfficeBoxNumber;
   }
-
   public function getPostOfficeBoxNumber()
   {
     return $this->postOfficeBoxNumber;
   }
-
   public function setPostalCode($postalCode)
   {
     $this->postalCode = $postalCode;
   }
-
   public function getPostalCode()
   {
     return $this->postalCode;
   }
-
   public function setRatingValue($ratingValue)
   {
     $this->ratingValue = $ratingValue;
   }
-
   public function getRatingValue()
   {
     return $this->ratingValue;
   }
-
   public function setReviewRating(Google_Service_Plus_ItemScope $reviewRating)
   {
     $this->reviewRating = $reviewRating;
   }
-
   public function getReviewRating()
   {
     return $this->reviewRating;
   }
-
   public function setStartDate($startDate)
   {
     $this->startDate = $startDate;
   }
-
   public function getStartDate()
   {
     return $this->startDate;
   }
-
   public function setStreetAddress($streetAddress)
   {
     $this->streetAddress = $streetAddress;
   }
-
   public function getStreetAddress()
   {
     return $this->streetAddress;
   }
-
   public function setText($text)
   {
     $this->text = $text;
   }
-
   public function getText()
   {
     return $this->text;
   }
-
   public function setThumbnail(Google_Service_Plus_ItemScope $thumbnail)
   {
     $this->thumbnail = $thumbnail;
   }
-
   public function getThumbnail()
   {
     return $this->thumbnail;
   }
-
   public function setThumbnailUrl($thumbnailUrl)
   {
     $this->thumbnailUrl = $thumbnailUrl;
   }
-
   public function getThumbnailUrl()
   {
     return $this->thumbnailUrl;
   }
-
   public function setTickerSymbol($tickerSymbol)
   {
     $this->tickerSymbol = $tickerSymbol;
   }
-
   public function getTickerSymbol()
   {
     return $this->tickerSymbol;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
   }
-
   public function setWorstRating($worstRating)
   {
     $this->worstRating = $worstRating;
   }
-
   public function getWorstRating()
   {
     return $this->worstRating;
@@ -2700,8 +2656,12 @@ class Google_Service_Plus_ItemScope extends Google_Collection
 
 class Google_Service_Plus_Moment extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $kind;
+  protected $objectType = 'Google_Service_Plus_ItemScope';
+  protected $objectDataType = '';
   protected $resultType = 'Google_Service_Plus_ItemScope';
   protected $resultDataType = '';
   public $startDate;
@@ -2709,61 +2669,59 @@ class Google_Service_Plus_Moment extends Google_Model
   protected $targetDataType = '';
   public $type;
 
+
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
+  public function setObject(Google_Service_Plus_ItemScope $object)
+  {
+    $this->object = $object;
+  }
+  public function getObject()
+  {
+    return $this->object;
+  }
   public function setResult(Google_Service_Plus_ItemScope $result)
   {
     $this->result = $result;
   }
-
   public function getResult()
   {
     return $this->result;
   }
-
   public function setStartDate($startDate)
   {
     $this->startDate = $startDate;
   }
-
   public function getStartDate()
   {
     return $this->startDate;
   }
-
   public function setTarget(Google_Service_Plus_ItemScope $target)
   {
     $this->target = $target;
   }
-
   public function getTarget()
   {
     return $this->target;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -2772,6 +2730,9 @@ class Google_Service_Plus_Moment extends Google_Model
 
 class Google_Service_Plus_MomentsFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_Plus_Moment';
   protected $itemsDataType = 'array';
@@ -2782,81 +2743,67 @@ class Google_Service_Plus_MomentsFeed extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextLink($nextLink)
   {
     $this->nextLink = $nextLink;
   }
-
   public function getNextLink()
   {
     return $this->nextLink;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -2865,6 +2812,9 @@ class Google_Service_Plus_MomentsFeed extends Google_Collection
 
 class Google_Service_Plus_PeopleFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_Plus_Person';
   protected $itemsDataType = 'array';
@@ -2874,71 +2824,59 @@ class Google_Service_Plus_PeopleFeed extends Google_Collection
   public $title;
   public $totalItems;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2947,6 +2885,9 @@ class Google_Service_Plus_PeopleFeed extends Google_Collection
 
 class Google_Service_Plus_Person extends Google_Collection
 {
+  protected $collection_key = 'urls';
+  protected $internal_gapi_mappings = array(
+  );
   public $aboutMe;
   protected $ageRangeType = 'Google_Service_Plus_PersonAgeRange';
   protected $ageRangeDataType = '';
@@ -2986,301 +2927,243 @@ class Google_Service_Plus_Person extends Google_Collection
   protected $urlsDataType = 'array';
   public $verified;
 
+
   public function setAboutMe($aboutMe)
   {
     $this->aboutMe = $aboutMe;
   }
-
   public function getAboutMe()
   {
     return $this->aboutMe;
   }
-
   public function setAgeRange(Google_Service_Plus_PersonAgeRange $ageRange)
   {
     $this->ageRange = $ageRange;
   }
-
   public function getAgeRange()
   {
     return $this->ageRange;
   }
-
   public function setBirthday($birthday)
   {
     $this->birthday = $birthday;
   }
-
   public function getBirthday()
   {
     return $this->birthday;
   }
-
   public function setBraggingRights($braggingRights)
   {
     $this->braggingRights = $braggingRights;
   }
-
   public function getBraggingRights()
   {
     return $this->braggingRights;
   }
-
   public function setCircledByCount($circledByCount)
   {
     $this->circledByCount = $circledByCount;
   }
-
   public function getCircledByCount()
   {
     return $this->circledByCount;
   }
-
   public function setCover(Google_Service_Plus_PersonCover $cover)
   {
     $this->cover = $cover;
   }
-
   public function getCover()
   {
     return $this->cover;
   }
-
   public function setCurrentLocation($currentLocation)
   {
     $this->currentLocation = $currentLocation;
   }
-
   public function getCurrentLocation()
   {
     return $this->currentLocation;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setDomain($domain)
   {
     $this->domain = $domain;
   }
-
   public function getDomain()
   {
     return $this->domain;
   }
-
   public function setEmails($emails)
   {
     $this->emails = $emails;
   }
-
   public function getEmails()
   {
     return $this->emails;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setGender($gender)
   {
     $this->gender = $gender;
   }
-
   public function getGender()
   {
     return $this->gender;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_Plus_PersonImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setIsPlusUser($isPlusUser)
   {
     $this->isPlusUser = $isPlusUser;
   }
-
   public function getIsPlusUser()
   {
     return $this->isPlusUser;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLanguage($language)
   {
     $this->language = $language;
   }
-
   public function getLanguage()
   {
     return $this->language;
   }
-
   public function setName(Google_Service_Plus_PersonName $name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setNickname($nickname)
   {
     $this->nickname = $nickname;
   }
-
   public function getNickname()
   {
     return $this->nickname;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setOccupation($occupation)
   {
     $this->occupation = $occupation;
   }
-
   public function getOccupation()
   {
     return $this->occupation;
   }
-
   public function setOrganizations($organizations)
   {
     $this->organizations = $organizations;
   }
-
   public function getOrganizations()
   {
     return $this->organizations;
   }
-
   public function setPlacesLived($placesLived)
   {
     $this->placesLived = $placesLived;
   }
-
   public function getPlacesLived()
   {
     return $this->placesLived;
   }
-
   public function setPlusOneCount($plusOneCount)
   {
     $this->plusOneCount = $plusOneCount;
   }
-
   public function getPlusOneCount()
   {
     return $this->plusOneCount;
   }
-
   public function setRelationshipStatus($relationshipStatus)
   {
     $this->relationshipStatus = $relationshipStatus;
   }
-
   public function getRelationshipStatus()
   {
     return $this->relationshipStatus;
   }
-
   public function setSkills($skills)
   {
     $this->skills = $skills;
   }
-
   public function getSkills()
   {
     return $this->skills;
   }
-
   public function setTagline($tagline)
   {
     $this->tagline = $tagline;
   }
-
   public function getTagline()
   {
     return $this->tagline;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setUrls($urls)
   {
     $this->urls = $urls;
   }
-
   public function getUrls()
   {
     return $this->urls;
   }
-
   public function setVerified($verified)
   {
     $this->verified = $verified;
   }
-
   public function getVerified()
   {
     return $this->verified;
@@ -3289,24 +3172,24 @@ class Google_Service_Plus_Person extends Google_Collection
 
 class Google_Service_Plus_PersonAgeRange extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $max;
   public $min;
+
 
   public function setMax($max)
   {
     $this->max = $max;
   }
-
   public function getMax()
   {
     return $this->max;
   }
-
   public function setMin($min)
   {
     $this->min = $min;
   }
-
   public function getMin()
   {
     return $this->min;
@@ -3315,37 +3198,35 @@ class Google_Service_Plus_PersonAgeRange extends Google_Model
 
 class Google_Service_Plus_PersonCover extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $coverInfoType = 'Google_Service_Plus_PersonCoverCoverInfo';
   protected $coverInfoDataType = '';
   protected $coverPhotoType = 'Google_Service_Plus_PersonCoverCoverPhoto';
   protected $coverPhotoDataType = '';
   public $layout;
 
+
   public function setCoverInfo(Google_Service_Plus_PersonCoverCoverInfo $coverInfo)
   {
     $this->coverInfo = $coverInfo;
   }
-
   public function getCoverInfo()
   {
     return $this->coverInfo;
   }
-
   public function setCoverPhoto(Google_Service_Plus_PersonCoverCoverPhoto $coverPhoto)
   {
     $this->coverPhoto = $coverPhoto;
   }
-
   public function getCoverPhoto()
   {
     return $this->coverPhoto;
   }
-
   public function setLayout($layout)
   {
     $this->layout = $layout;
   }
-
   public function getLayout()
   {
     return $this->layout;
@@ -3354,24 +3235,24 @@ class Google_Service_Plus_PersonCover extends Google_Model
 
 class Google_Service_Plus_PersonCoverCoverInfo extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $leftImageOffset;
   public $topImageOffset;
+
 
   public function setLeftImageOffset($leftImageOffset)
   {
     $this->leftImageOffset = $leftImageOffset;
   }
-
   public function getLeftImageOffset()
   {
     return $this->leftImageOffset;
   }
-
   public function setTopImageOffset($topImageOffset)
   {
     $this->topImageOffset = $topImageOffset;
   }
-
   public function getTopImageOffset()
   {
     return $this->topImageOffset;
@@ -3380,35 +3261,33 @@ class Google_Service_Plus_PersonCoverCoverInfo extends Google_Model
 
 class Google_Service_Plus_PersonCoverCoverPhoto extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $url;
   public $width;
+
 
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -3417,24 +3296,24 @@ class Google_Service_Plus_PersonCoverCoverPhoto extends Google_Model
 
 class Google_Service_Plus_PersonEmails extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $type;
   public $value;
+
 
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -3443,13 +3322,24 @@ class Google_Service_Plus_PersonEmails extends Google_Model
 
 class Google_Service_Plus_PersonImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  public $isDefault;
   public $url;
 
+
+  public function setIsDefault($isDefault)
+  {
+    $this->isDefault = $isDefault;
+  }
+  public function getIsDefault()
+  {
+    return $this->isDefault;
+  }
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -3458,6 +3348,8 @@ class Google_Service_Plus_PersonImage extends Google_Model
 
 class Google_Service_Plus_PersonName extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $familyName;
   public $formatted;
   public $givenName;
@@ -3465,61 +3357,51 @@ class Google_Service_Plus_PersonName extends Google_Model
   public $honorificSuffix;
   public $middleName;
 
+
   public function setFamilyName($familyName)
   {
     $this->familyName = $familyName;
   }
-
   public function getFamilyName()
   {
     return $this->familyName;
   }
-
   public function setFormatted($formatted)
   {
     $this->formatted = $formatted;
   }
-
   public function getFormatted()
   {
     return $this->formatted;
   }
-
   public function setGivenName($givenName)
   {
     $this->givenName = $givenName;
   }
-
   public function getGivenName()
   {
     return $this->givenName;
   }
-
   public function setHonorificPrefix($honorificPrefix)
   {
     $this->honorificPrefix = $honorificPrefix;
   }
-
   public function getHonorificPrefix()
   {
     return $this->honorificPrefix;
   }
-
   public function setHonorificSuffix($honorificSuffix)
   {
     $this->honorificSuffix = $honorificSuffix;
   }
-
   public function getHonorificSuffix()
   {
     return $this->honorificSuffix;
   }
-
   public function setMiddleName($middleName)
   {
     $this->middleName = $middleName;
   }
-
   public function getMiddleName()
   {
     return $this->middleName;
@@ -3528,6 +3410,8 @@ class Google_Service_Plus_PersonName extends Google_Model
 
 class Google_Service_Plus_PersonOrganizations extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $department;
   public $description;
   public $endDate;
@@ -3538,91 +3422,75 @@ class Google_Service_Plus_PersonOrganizations extends Google_Model
   public $title;
   public $type;
 
+
   public function setDepartment($department)
   {
     $this->department = $department;
   }
-
   public function getDepartment()
   {
     return $this->department;
   }
-
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setEndDate($endDate)
   {
     $this->endDate = $endDate;
   }
-
   public function getEndDate()
   {
     return $this->endDate;
   }
-
   public function setLocation($location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setPrimary($primary)
   {
     $this->primary = $primary;
   }
-
   public function getPrimary()
   {
     return $this->primary;
   }
-
   public function setStartDate($startDate)
   {
     $this->startDate = $startDate;
   }
-
   public function getStartDate()
   {
     return $this->startDate;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -3631,24 +3499,24 @@ class Google_Service_Plus_PersonOrganizations extends Google_Model
 
 class Google_Service_Plus_PersonPlacesLived extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $primary;
   public $value;
+
 
   public function setPrimary($primary)
   {
     $this->primary = $primary;
   }
-
   public function getPrimary()
   {
     return $this->primary;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -3657,35 +3525,33 @@ class Google_Service_Plus_PersonPlacesLived extends Google_Model
 
 class Google_Service_Plus_PersonUrls extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $label;
   public $type;
   public $value;
+
 
   public function setLabel($label)
   {
     $this->label = $label;
   }
-
   public function getLabel()
   {
     return $this->label;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -3694,6 +3560,8 @@ class Google_Service_Plus_PersonUrls extends Google_Model
 
 class Google_Service_Plus_Place extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $addressType = 'Google_Service_Plus_PlaceAddress';
   protected $addressDataType = '';
   public $displayName;
@@ -3702,51 +3570,43 @@ class Google_Service_Plus_Place extends Google_Model
   protected $positionType = 'Google_Service_Plus_PlacePosition';
   protected $positionDataType = '';
 
+
   public function setAddress(Google_Service_Plus_PlaceAddress $address)
   {
     $this->address = $address;
   }
-
   public function getAddress()
   {
     return $this->address;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setPosition(Google_Service_Plus_PlacePosition $position)
   {
     $this->position = $position;
   }
-
   public function getPosition()
   {
     return $this->position;
@@ -3755,13 +3615,15 @@ class Google_Service_Plus_Place extends Google_Model
 
 class Google_Service_Plus_PlaceAddress extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $formatted;
+
 
   public function setFormatted($formatted)
   {
     $this->formatted = $formatted;
   }
-
   public function getFormatted()
   {
     return $this->formatted;
@@ -3770,24 +3632,24 @@ class Google_Service_Plus_PlaceAddress extends Google_Model
 
 class Google_Service_Plus_PlacePosition extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $latitude;
   public $longitude;
+
 
   public function setLatitude($latitude)
   {
     $this->latitude = $latitude;
   }
-
   public function getLatitude()
   {
     return $this->latitude;
   }
-
   public function setLongitude($longitude)
   {
     $this->longitude = $longitude;
   }
-
   public function getLongitude()
   {
     return $this->longitude;
@@ -3796,35 +3658,33 @@ class Google_Service_Plus_PlacePosition extends Google_Model
 
 class Google_Service_Plus_PlusAclentryResource extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $displayName;
   public $id;
   public $type;
+
 
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
