@@ -19,8 +19,7 @@
  * Service definition for PlusDomains (v1).
  *
  * <p>
- * The Google+ API enables developers to build on top of the Google+ platform.
- * </p>
+ * The Google+ API enables developers to build on top of the Google+ platform.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,25 +31,35 @@
 class Google_Service_PlusDomains extends Google_Service
 {
   /** View your circles and the people and pages in them. */
-  const PLUS_CIRCLES_READ = "https://www.googleapis.com/auth/plus.circles.read";
+  const PLUS_CIRCLES_READ =
+      "https://www.googleapis.com/auth/plus.circles.read";
   /** Manage your circles and add people and pages. People and pages you add to your circles will be notified. Others may see this information publicly. People you add to circles can use Hangouts with you.. */
-  const PLUS_CIRCLES_WRITE = "https://www.googleapis.com/auth/plus.circles.write";
+  const PLUS_CIRCLES_WRITE =
+      "https://www.googleapis.com/auth/plus.circles.write";
   /** Know your basic profile info and list of people in your circles.. */
-  const PLUS_LOGIN = "https://www.googleapis.com/auth/plus.login";
+  const PLUS_LOGIN =
+      "https://www.googleapis.com/auth/plus.login";
   /** Know who you are on Google. */
-  const PLUS_ME = "https://www.googleapis.com/auth/plus.me";
+  const PLUS_ME =
+      "https://www.googleapis.com/auth/plus.me";
   /** Send your photos and videos to Google+. */
-  const PLUS_MEDIA_UPLOAD = "https://www.googleapis.com/auth/plus.media.upload";
+  const PLUS_MEDIA_UPLOAD =
+      "https://www.googleapis.com/auth/plus.media.upload";
   /** View your own Google+ profile and profiles visible to you. */
-  const PLUS_PROFILES_READ = "https://www.googleapis.com/auth/plus.profiles.read";
+  const PLUS_PROFILES_READ =
+      "https://www.googleapis.com/auth/plus.profiles.read";
   /** View your Google+ posts, comments, and stream. */
-  const PLUS_STREAM_READ = "https://www.googleapis.com/auth/plus.stream.read";
+  const PLUS_STREAM_READ =
+      "https://www.googleapis.com/auth/plus.stream.read";
   /** Manage your Google+ posts, comments, and stream. */
-  const PLUS_STREAM_WRITE = "https://www.googleapis.com/auth/plus.stream.write";
+  const PLUS_STREAM_WRITE =
+      "https://www.googleapis.com/auth/plus.stream.write";
   /** View your email address. */
-  const USERINFO_EMAIL = "https://www.googleapis.com/auth/userinfo.email";
-  /** View basic information about your account. */
-  const USERINFO_PROFILE = "https://www.googleapis.com/auth/userinfo.profile";
+  const USERINFO_EMAIL =
+      "https://www.googleapis.com/auth/userinfo.email";
+  /** View your basic profile info. */
+  const USERINFO_PROFILE =
+      "https://www.googleapis.com/auth/userinfo.profile";
 
   public $activities;
   public $audiences;
@@ -68,6 +77,7 @@ class Google_Service_PlusDomains extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'plusDomains/v1/';
     $this->version = 'v1';
     $this->serviceName = 'plusDomains';
@@ -458,8 +468,7 @@ class Google_Service_PlusDomains_Activities_Resource extends Google_Service_Reso
   /**
    * Get an activity. (activities.get)
    *
-   * @param string $activityId
-   * The ID of the activity to get.
+   * @param string $activityId The ID of the activity to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Activity
    */
@@ -469,19 +478,18 @@ class Google_Service_PlusDomains_Activities_Resource extends Google_Service_Reso
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_PlusDomains_Activity");
   }
+
   /**
    * Create a new activity for the authenticated user. (activities.insert)
    *
-   * @param string $userId
-   * The ID of the user to create the activity on behalf of. Its value should be "me", to indicate
-    * the authenticated user.
+   * @param string $userId The ID of the user to create the activity on behalf of.
+   * Its value should be "me", to indicate the authenticated user.
    * @param Google_Activity $postBody
    * @param array $optParams Optional parameters.
    *
-   * @opt_param bool preview
-   * If "true", extract the potential media attachments for a URL. The response will include all
-    * possible attachments for a URL, including video, photos, and articles based on the content of
-    * the page.
+   * @opt_param bool preview If "true", extract the potential media attachments
+   * for a URL. The response will include all possible attachments for a URL,
+   * including video, photos, and articles based on the content of the page.
    * @return Google_Service_PlusDomains_Activity
    */
   public function insert($userId, Google_Service_PlusDomains_Activity $postBody, $optParams = array())
@@ -490,23 +498,22 @@ class Google_Service_PlusDomains_Activities_Resource extends Google_Service_Reso
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_PlusDomains_Activity");
   }
+
   /**
    * List all of the activities in the specified collection for a particular user.
    * (activities.listActivities)
    *
-   * @param string $userId
-   * The ID of the user to get activities for. The special value "me" can be used to indicate the
-    * authenticated user.
-   * @param string $collection
-   * The collection of activities to list.
+   * @param string $userId The ID of the user to get activities for. The special
+   * value "me" can be used to indicate the authenticated user.
+   * @param string $collection The collection of activities to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of activities to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of activities to include in
+   * the response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_ActivityFeed
    */
   public function listActivities($userId, $collection, $optParams = array())
@@ -532,17 +539,16 @@ class Google_Service_PlusDomains_Audiences_Resource extends Google_Service_Resou
    * List all of the audiences to which a user can share.
    * (audiences.listAudiences)
    *
-   * @param string $userId
-   * The ID of the user to get audiences for. The special value "me" can be used to indicate the
-    * authenticated user.
+   * @param string $userId The ID of the user to get audiences for. The special
+   * value "me" can be used to indicate the authenticated user.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of circles to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of circles to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_AudiencesFeed
    */
   public function listAudiences($userId, $optParams = array())
@@ -568,14 +574,13 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
    * Add a person to a circle. Google+ limits certain circle operations, including
    * the number of circle adds. Learn More. (circles.addPeople)
    *
-   * @param string $circleId
-   * The ID of the circle to add the person to.
+   * @param string $circleId The ID of the circle to add the person to.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string userId
-   * IDs of the people to add to the circle. Optional, can be repeated.
-   * @opt_param string email
-   * Email of the people to add to the circle. Optional, can be repeated.
+   * @opt_param string userId IDs of the people to add to the circle. Optional,
+   * can be repeated.
+   * @opt_param string email Email of the people to add to the circle. Optional,
+   * can be repeated.
    * @return Google_Service_PlusDomains_Circle
    */
   public function addPeople($circleId, $optParams = array())
@@ -584,11 +589,11 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('addPeople', array($params), "Google_Service_PlusDomains_Circle");
   }
+
   /**
    * Get a circle. (circles.get)
    *
-   * @param string $circleId
-   * The ID of the circle to get.
+   * @param string $circleId The ID of the circle to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Circle
    */
@@ -598,12 +603,12 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_PlusDomains_Circle");
   }
+
   /**
    * Create a new circle for the authenticated user. (circles.insert)
    *
-   * @param string $userId
-   * The ID of the user to create the circle on behalf of. The value "me" can be used to indicate the
-    * authenticated user.
+   * @param string $userId The ID of the user to create the circle on behalf of.
+   * The value "me" can be used to indicate the authenticated user.
    * @param Google_Circle $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Circle
@@ -614,20 +619,20 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_PlusDomains_Circle");
   }
+
   /**
    * List all of the circles for a user. (circles.listCircles)
    *
-   * @param string $userId
-   * The ID of the user to get circles for. The special value "me" can be used to indicate the
-    * authenticated user.
+   * @param string $userId The ID of the user to get circles for. The special
+   * value "me" can be used to indicate the authenticated user.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of circles to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of circles to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_CircleFeed
    */
   public function listCircles($userId, $optParams = array())
@@ -636,12 +641,12 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_PlusDomains_CircleFeed");
   }
+
   /**
    * Update a circle's description. This method supports patch semantics.
    * (circles.patch)
    *
-   * @param string $circleId
-   * The ID of the circle to update.
+   * @param string $circleId The ID of the circle to update.
    * @param Google_Circle $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Circle
@@ -652,11 +657,11 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_PlusDomains_Circle");
   }
+
   /**
    * Delete a circle. (circles.remove)
    *
-   * @param string $circleId
-   * The ID of the circle to delete.
+   * @param string $circleId The ID of the circle to delete.
    * @param array $optParams Optional parameters.
    */
   public function remove($circleId, $optParams = array())
@@ -665,17 +670,17 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('remove', array($params));
   }
+
   /**
    * Remove a person from a circle. (circles.removePeople)
    *
-   * @param string $circleId
-   * The ID of the circle to remove the person from.
+   * @param string $circleId The ID of the circle to remove the person from.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string userId
-   * IDs of the people to remove from the circle. Optional, can be repeated.
-   * @opt_param string email
-   * Email of the people to add to the circle. Optional, can be repeated.
+   * @opt_param string userId IDs of the people to remove from the circle.
+   * Optional, can be repeated.
+   * @opt_param string email Email of the people to add to the circle. Optional,
+   * can be repeated.
    */
   public function removePeople($circleId, $optParams = array())
   {
@@ -683,11 +688,11 @@ class Google_Service_PlusDomains_Circles_Resource extends Google_Service_Resourc
     $params = array_merge($params, $optParams);
     return $this->call('removePeople', array($params));
   }
+
   /**
    * Update a circle's description. (circles.update)
    *
-   * @param string $circleId
-   * The ID of the circle to update.
+   * @param string $circleId The ID of the circle to update.
    * @param Google_Circle $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Circle
@@ -714,8 +719,7 @@ class Google_Service_PlusDomains_Comments_Resource extends Google_Service_Resour
   /**
    * Get a comment. (comments.get)
    *
-   * @param string $commentId
-   * The ID of the comment to get.
+   * @param string $commentId The ID of the comment to get.
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Comment
    */
@@ -725,11 +729,11 @@ class Google_Service_PlusDomains_Comments_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_PlusDomains_Comment");
   }
+
   /**
    * Create a new comment in reply to an activity. (comments.insert)
    *
-   * @param string $activityId
-   * The ID of the activity to reply to.
+   * @param string $activityId The ID of the activity to reply to.
    * @param Google_Comment $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Comment
@@ -740,21 +744,20 @@ class Google_Service_PlusDomains_Comments_Resource extends Google_Service_Resour
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_PlusDomains_Comment");
   }
+
   /**
    * List all of the comments for an activity. (comments.listComments)
    *
-   * @param string $activityId
-   * The ID of the activity to get comments for.
+   * @param string $activityId The ID of the activity to get comments for.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string sortOrder
-   * The order in which to sort the list of comments.
-   * @opt_param string maxResults
-   * The maximum number of comments to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string sortOrder The order in which to sort the list of comments.
+   * @opt_param string maxResults The maximum number of comments to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_CommentFeed
    */
   public function listComments($activityId, $optParams = array())
@@ -782,10 +785,8 @@ class Google_Service_PlusDomains_Media_Resource extends Google_Service_Resource
    * photos are less than 2048 pixels on their longest side or videos are less
    * than 15 minutes in length. (media.insert)
    *
-   * @param string $userId
-   * The ID of the user to create the activity on behalf of.
+   * @param string $userId The ID of the user to create the activity on behalf of.
    * @param string $collection
-   *
    * @param Google_Media $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Media
@@ -812,9 +813,8 @@ class Google_Service_PlusDomains_People_Resource extends Google_Service_Resource
   /**
    * Get a person's profile. (people.get)
    *
-   * @param string $userId
-   * The ID of the person to get the profile for. The special value "me" can be used to indicate the
-    * authenticated user.
+   * @param string $userId The ID of the person to get the profile for. The
+   * special value "me" can be used to indicate the authenticated user.
    * @param array $optParams Optional parameters.
    * @return Google_Service_PlusDomains_Person
    */
@@ -824,24 +824,22 @@ class Google_Service_PlusDomains_People_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_PlusDomains_Person");
   }
+
   /**
    * List all of the people in the specified collection. (people.listPeople)
    *
-   * @param string $userId
-   * Get the collection of people for the person identified. Use "me" to indicate the authenticated
-    * user.
-   * @param string $collection
-   * The collection of people to list.
+   * @param string $userId Get the collection of people for the person identified.
+   * Use "me" to indicate the authenticated user.
+   * @param string $collection The collection of people to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string orderBy
-   * The order to return people in.
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of people to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string orderBy The order to return people in.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of people to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_PeopleFeed
    */
   public function listPeople($userId, $collection, $optParams = array())
@@ -850,22 +848,22 @@ class Google_Service_PlusDomains_People_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_PlusDomains_PeopleFeed");
   }
+
   /**
    * List all of the people in the specified collection for a particular activity.
    * (people.listByActivity)
    *
-   * @param string $activityId
-   * The ID of the activity to get the list of people for.
-   * @param string $collection
-   * The collection of people to list.
+   * @param string $activityId The ID of the activity to get the list of people
+   * for.
+   * @param string $collection The collection of people to list.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of people to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of people to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_PeopleFeed
    */
   public function listByActivity($activityId, $collection, $optParams = array())
@@ -874,19 +872,19 @@ class Google_Service_PlusDomains_People_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('listByActivity', array($params), "Google_Service_PlusDomains_PeopleFeed");
   }
+
   /**
    * List all of the people who are members of a circle. (people.listByCircle)
    *
-   * @param string $circleId
-   * The ID of the circle to get the members of.
+   * @param string $circleId The ID of the circle to get the members of.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * The continuation token, which is used to page through large result sets. To get the next page of
-    * results, set this parameter to the value of "nextPageToken" from the previous response.
-   * @opt_param string maxResults
-   * The maximum number of people to include in the response, which is used for paging. For any
-    * response, the actual number returned might be less than the specified maxResults.
+   * @opt_param string pageToken The continuation token, which is used to page
+   * through large result sets. To get the next page of results, set this
+   * parameter to the value of "nextPageToken" from the previous response.
+   * @opt_param string maxResults The maximum number of people to include in the
+   * response, which is used for paging. For any response, the actual number
+   * returned might be less than the specified maxResults.
    * @return Google_Service_PlusDomains_PeopleFeed
    */
   public function listByCircle($circleId, $optParams = array())
@@ -902,47 +900,44 @@ class Google_Service_PlusDomains_People_Resource extends Google_Service_Resource
 
 class Google_Service_PlusDomains_Acl extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $description;
   public $domainRestricted;
   protected $itemsType = 'Google_Service_PlusDomains_PlusDomainsAclentryResource';
   protected $itemsDataType = 'array';
   public $kind;
 
+
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setDomainRestricted($domainRestricted)
   {
     $this->domainRestricted = $domainRestricted;
   }
-
   public function getDomainRestricted()
   {
     return $this->domainRestricted;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
@@ -951,6 +946,8 @@ class Google_Service_PlusDomains_Acl extends Google_Collection
 
 class Google_Service_PlusDomains_Activity extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $accessType = 'Google_Service_PlusDomains_Acl';
   protected $accessDataType = '';
   protected $actorType = 'Google_Service_PlusDomains_ActivityActor';
@@ -977,201 +974,163 @@ class Google_Service_PlusDomains_Activity extends Google_Model
   public $url;
   public $verb;
 
+
   public function setAccess(Google_Service_PlusDomains_Acl $access)
   {
     $this->access = $access;
   }
-
   public function getAccess()
   {
     return $this->access;
   }
-
   public function setActor(Google_Service_PlusDomains_ActivityActor $actor)
   {
     $this->actor = $actor;
   }
-
   public function getActor()
   {
     return $this->actor;
   }
-
   public function setAddress($address)
   {
     $this->address = $address;
   }
-
   public function getAddress()
   {
     return $this->address;
   }
-
   public function setAnnotation($annotation)
   {
     $this->annotation = $annotation;
   }
-
   public function getAnnotation()
   {
     return $this->annotation;
   }
-
   public function setCrosspostSource($crosspostSource)
   {
     $this->crosspostSource = $crosspostSource;
   }
-
   public function getCrosspostSource()
   {
     return $this->crosspostSource;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setGeocode($geocode)
   {
     $this->geocode = $geocode;
   }
-
   public function getGeocode()
   {
     return $this->geocode;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLocation(Google_Service_PlusDomains_Place $location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setObject(Google_Service_PlusDomains_ActivityObject $object)
   {
     $this->object = $object;
   }
-
   public function getObject()
   {
     return $this->object;
   }
-
   public function setPlaceId($placeId)
   {
     $this->placeId = $placeId;
   }
-
   public function getPlaceId()
   {
     return $this->placeId;
   }
-
   public function setPlaceName($placeName)
   {
     $this->placeName = $placeName;
   }
-
   public function getPlaceName()
   {
     return $this->placeName;
   }
-
   public function setProvider(Google_Service_PlusDomains_ActivityProvider $provider)
   {
     $this->provider = $provider;
   }
-
   public function getProvider()
   {
     return $this->provider;
   }
-
   public function setPublished($published)
   {
     $this->published = $published;
   }
-
   public function getPublished()
   {
     return $this->published;
   }
-
   public function setRadius($radius)
   {
     $this->radius = $radius;
   }
-
   public function getRadius()
   {
     return $this->radius;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setVerb($verb)
   {
     $this->verb = $verb;
   }
-
   public function getVerb()
   {
     return $this->verb;
@@ -1180,6 +1139,10 @@ class Google_Service_PlusDomains_Activity extends Google_Model
 
 class Google_Service_PlusDomains_ActivityActor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  protected $clientSpecificActorInfoType = 'Google_Service_PlusDomains_ActivityActorClientSpecificActorInfo';
+  protected $clientSpecificActorInfoDataType = '';
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_PlusDomains_ActivityActorImage';
@@ -1187,67 +1150,114 @@ class Google_Service_PlusDomains_ActivityActor extends Google_Model
   protected $nameType = 'Google_Service_PlusDomains_ActivityActorName';
   protected $nameDataType = '';
   public $url;
+  protected $verificationType = 'Google_Service_PlusDomains_ActivityActorVerification';
+  protected $verificationDataType = '';
 
+
+  public function setClientSpecificActorInfo(Google_Service_PlusDomains_ActivityActorClientSpecificActorInfo $clientSpecificActorInfo)
+  {
+    $this->clientSpecificActorInfo = $clientSpecificActorInfo;
+  }
+  public function getClientSpecificActorInfo()
+  {
+    return $this->clientSpecificActorInfo;
+  }
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_PlusDomains_ActivityActorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setName(Google_Service_PlusDomains_ActivityActorName $name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
+  }
+  public function setVerification(Google_Service_PlusDomains_ActivityActorVerification $verification)
+  {
+    $this->verification = $verification;
+  }
+  public function getVerification()
+  {
+    return $this->verification;
+  }
+}
+
+class Google_Service_PlusDomains_ActivityActorClientSpecificActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  protected $youtubeActorInfoType = 'Google_Service_PlusDomains_ActivityActorClientSpecificActorInfoYoutubeActorInfo';
+  protected $youtubeActorInfoDataType = '';
+
+
+  public function setYoutubeActorInfo(Google_Service_PlusDomains_ActivityActorClientSpecificActorInfoYoutubeActorInfo $youtubeActorInfo)
+  {
+    $this->youtubeActorInfo = $youtubeActorInfo;
+  }
+  public function getYoutubeActorInfo()
+  {
+    return $this->youtubeActorInfo;
+  }
+}
+
+class Google_Service_PlusDomains_ActivityActorClientSpecificActorInfoYoutubeActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $channelId;
+
+
+  public function setChannelId($channelId)
+  {
+    $this->channelId = $channelId;
+  }
+  public function getChannelId()
+  {
+    return $this->channelId;
   }
 }
 
 class Google_Service_PlusDomains_ActivityActorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1256,32 +1266,52 @@ class Google_Service_PlusDomains_ActivityActorImage extends Google_Model
 
 class Google_Service_PlusDomains_ActivityActorName extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $familyName;
   public $givenName;
+
 
   public function setFamilyName($familyName)
   {
     $this->familyName = $familyName;
   }
-
   public function getFamilyName()
   {
     return $this->familyName;
   }
-
   public function setGivenName($givenName)
   {
     $this->givenName = $givenName;
   }
-
   public function getGivenName()
   {
     return $this->givenName;
   }
 }
 
+class Google_Service_PlusDomains_ActivityActorVerification extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $adHocVerified;
+
+
+  public function setAdHocVerified($adHocVerified)
+  {
+    $this->adHocVerified = $adHocVerified;
+  }
+  public function getAdHocVerified()
+  {
+    return $this->adHocVerified;
+  }
+}
+
 class Google_Service_PlusDomains_ActivityFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   public $id;
   protected $itemsType = 'Google_Service_PlusDomains_Activity';
@@ -1293,91 +1323,75 @@ class Google_Service_PlusDomains_ActivityFeed extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextLink($nextLink)
   {
     $this->nextLink = $nextLink;
   }
-
   public function getNextLink()
   {
     return $this->nextLink;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -1386,6 +1400,9 @@ class Google_Service_PlusDomains_ActivityFeed extends Google_Collection
 
 class Google_Service_PlusDomains_ActivityObject extends Google_Collection
 {
+  protected $collection_key = 'attachments';
+  protected $internal_gapi_mappings = array(
+  );
   protected $actorType = 'Google_Service_PlusDomains_ActivityObjectActor';
   protected $actorDataType = '';
   protected $attachmentsType = 'Google_Service_PlusDomains_ActivityObjectAttachments';
@@ -1404,111 +1421,91 @@ class Google_Service_PlusDomains_ActivityObject extends Google_Collection
   protected $statusForViewerDataType = '';
   public $url;
 
+
   public function setActor(Google_Service_PlusDomains_ActivityObjectActor $actor)
   {
     $this->actor = $actor;
   }
-
   public function getActor()
   {
     return $this->actor;
   }
-
   public function setAttachments($attachments)
   {
     $this->attachments = $attachments;
   }
-
   public function getAttachments()
   {
     return $this->attachments;
   }
-
   public function setContent($content)
   {
     $this->content = $content;
   }
-
   public function getContent()
   {
     return $this->content;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setOriginalContent($originalContent)
   {
     $this->originalContent = $originalContent;
   }
-
   public function getOriginalContent()
   {
     return $this->originalContent;
   }
-
   public function setPlusoners(Google_Service_PlusDomains_ActivityObjectPlusoners $plusoners)
   {
     $this->plusoners = $plusoners;
   }
-
   public function getPlusoners()
   {
     return $this->plusoners;
   }
-
   public function setReplies(Google_Service_PlusDomains_ActivityObjectReplies $replies)
   {
     $this->replies = $replies;
   }
-
   public function getReplies()
   {
     return $this->replies;
   }
-
   public function setResharers(Google_Service_PlusDomains_ActivityObjectResharers $resharers)
   {
     $this->resharers = $resharers;
   }
-
   public function getResharers()
   {
     return $this->resharers;
   }
-
   public function setStatusForViewer(Google_Service_PlusDomains_ActivityObjectStatusForViewer $statusForViewer)
   {
     $this->statusForViewer = $statusForViewer;
   }
-
   public function getStatusForViewer()
   {
     return $this->statusForViewer;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1517,70 +1514,143 @@ class Google_Service_PlusDomains_ActivityObject extends Google_Collection
 
 class Google_Service_PlusDomains_ActivityObjectActor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  protected $clientSpecificActorInfoType = 'Google_Service_PlusDomains_ActivityObjectActorClientSpecificActorInfo';
+  protected $clientSpecificActorInfoDataType = '';
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_PlusDomains_ActivityObjectActorImage';
   protected $imageDataType = '';
   public $url;
+  protected $verificationType = 'Google_Service_PlusDomains_ActivityObjectActorVerification';
+  protected $verificationDataType = '';
 
+
+  public function setClientSpecificActorInfo(Google_Service_PlusDomains_ActivityObjectActorClientSpecificActorInfo $clientSpecificActorInfo)
+  {
+    $this->clientSpecificActorInfo = $clientSpecificActorInfo;
+  }
+  public function getClientSpecificActorInfo()
+  {
+    return $this->clientSpecificActorInfo;
+  }
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_PlusDomains_ActivityObjectActorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
+  }
+  public function setVerification(Google_Service_PlusDomains_ActivityObjectActorVerification $verification)
+  {
+    $this->verification = $verification;
+  }
+  public function getVerification()
+  {
+    return $this->verification;
+  }
+}
+
+class Google_Service_PlusDomains_ActivityObjectActorClientSpecificActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  protected $youtubeActorInfoType = 'Google_Service_PlusDomains_ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo';
+  protected $youtubeActorInfoDataType = '';
+
+
+  public function setYoutubeActorInfo(Google_Service_PlusDomains_ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo $youtubeActorInfo)
+  {
+    $this->youtubeActorInfo = $youtubeActorInfo;
+  }
+  public function getYoutubeActorInfo()
+  {
+    return $this->youtubeActorInfo;
+  }
+}
+
+class Google_Service_PlusDomains_ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $channelId;
+
+
+  public function setChannelId($channelId)
+  {
+    $this->channelId = $channelId;
+  }
+  public function getChannelId()
+  {
+    return $this->channelId;
   }
 }
 
 class Google_Service_PlusDomains_ActivityObjectActorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
 }
 
+class Google_Service_PlusDomains_ActivityObjectActorVerification extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $adHocVerified;
+
+
+  public function setAdHocVerified($adHocVerified)
+  {
+    $this->adHocVerified = $adHocVerified;
+  }
+  public function getAdHocVerified()
+  {
+    return $this->adHocVerified;
+  }
+}
+
 class Google_Service_PlusDomains_ActivityObjectAttachments extends Google_Collection
 {
+  protected $collection_key = 'thumbnails';
+  protected $internal_gapi_mappings = array(
+  );
   public $content;
   public $displayName;
   protected $embedType = 'Google_Service_PlusDomains_ActivityObjectAttachmentsEmbed';
@@ -1597,101 +1667,83 @@ class Google_Service_PlusDomains_ActivityObjectAttachments extends Google_Collec
   protected $thumbnailsDataType = 'array';
   public $url;
 
+
   public function setContent($content)
   {
     $this->content = $content;
   }
-
   public function getContent()
   {
     return $this->content;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setEmbed(Google_Service_PlusDomains_ActivityObjectAttachmentsEmbed $embed)
   {
     $this->embed = $embed;
   }
-
   public function getEmbed()
   {
     return $this->embed;
   }
-
   public function setFullImage(Google_Service_PlusDomains_ActivityObjectAttachmentsFullImage $fullImage)
   {
     $this->fullImage = $fullImage;
   }
-
   public function getFullImage()
   {
     return $this->fullImage;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_PlusDomains_ActivityObjectAttachmentsImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setPreviewThumbnails($previewThumbnails)
   {
     $this->previewThumbnails = $previewThumbnails;
   }
-
   public function getPreviewThumbnails()
   {
     return $this->previewThumbnails;
   }
-
   public function setThumbnails($thumbnails)
   {
     $this->thumbnails = $thumbnails;
   }
-
   public function getThumbnails()
   {
     return $this->thumbnails;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1700,24 +1752,24 @@ class Google_Service_PlusDomains_ActivityObjectAttachments extends Google_Collec
 
 class Google_Service_PlusDomains_ActivityObjectAttachmentsEmbed extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $type;
   public $url;
+
 
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1726,46 +1778,42 @@ class Google_Service_PlusDomains_ActivityObjectAttachmentsEmbed extends Google_M
 
 class Google_Service_PlusDomains_ActivityObjectAttachmentsFullImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -1774,46 +1822,42 @@ class Google_Service_PlusDomains_ActivityObjectAttachmentsFullImage extends Goog
 
 class Google_Service_PlusDomains_ActivityObjectAttachmentsImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -1822,13 +1866,15 @@ class Google_Service_PlusDomains_ActivityObjectAttachmentsImage extends Google_M
 
 class Google_Service_PlusDomains_ActivityObjectAttachmentsPreviewThumbnails extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1837,36 +1883,34 @@ class Google_Service_PlusDomains_ActivityObjectAttachmentsPreviewThumbnails exte
 
 class Google_Service_PlusDomains_ActivityObjectAttachmentsThumbnails extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $description;
   protected $imageType = 'Google_Service_PlusDomains_ActivityObjectAttachmentsThumbnailsImage';
   protected $imageDataType = '';
   public $url;
 
+
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setImage(Google_Service_PlusDomains_ActivityObjectAttachmentsThumbnailsImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -1875,46 +1919,42 @@ class Google_Service_PlusDomains_ActivityObjectAttachmentsThumbnails extends Goo
 
 class Google_Service_PlusDomains_ActivityObjectAttachmentsThumbnailsImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -1923,24 +1963,24 @@ class Google_Service_PlusDomains_ActivityObjectAttachmentsThumbnailsImage extend
 
 class Google_Service_PlusDomains_ActivityObjectPlusoners extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $selfLink;
   public $totalItems;
+
 
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -1949,24 +1989,24 @@ class Google_Service_PlusDomains_ActivityObjectPlusoners extends Google_Model
 
 class Google_Service_PlusDomains_ActivityObjectReplies extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $selfLink;
   public $totalItems;
+
 
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -1975,24 +2015,24 @@ class Google_Service_PlusDomains_ActivityObjectReplies extends Google_Model
 
 class Google_Service_PlusDomains_ActivityObjectResharers extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $selfLink;
   public $totalItems;
+
 
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2001,57 +2041,51 @@ class Google_Service_PlusDomains_ActivityObjectResharers extends Google_Model
 
 class Google_Service_PlusDomains_ActivityObjectStatusForViewer extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $canComment;
   public $canPlusone;
   public $canUpdate;
   public $isPlusOned;
   public $resharingDisabled;
 
+
   public function setCanComment($canComment)
   {
     $this->canComment = $canComment;
   }
-
   public function getCanComment()
   {
     return $this->canComment;
   }
-
   public function setCanPlusone($canPlusone)
   {
     $this->canPlusone = $canPlusone;
   }
-
   public function getCanPlusone()
   {
     return $this->canPlusone;
   }
-
   public function setCanUpdate($canUpdate)
   {
     $this->canUpdate = $canUpdate;
   }
-
   public function getCanUpdate()
   {
     return $this->canUpdate;
   }
-
   public function setIsPlusOned($isPlusOned)
   {
     $this->isPlusOned = $isPlusOned;
   }
-
   public function getIsPlusOned()
   {
     return $this->isPlusOned;
   }
-
   public function setResharingDisabled($resharingDisabled)
   {
     $this->resharingDisabled = $resharingDisabled;
   }
-
   public function getResharingDisabled()
   {
     return $this->resharingDisabled;
@@ -2060,13 +2094,15 @@ class Google_Service_PlusDomains_ActivityObjectStatusForViewer extends Google_Mo
 
 class Google_Service_PlusDomains_ActivityProvider extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $title;
+
 
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
@@ -2075,6 +2111,8 @@ class Google_Service_PlusDomains_ActivityProvider extends Google_Model
 
 class Google_Service_PlusDomains_Audience extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemType = 'Google_Service_PlusDomains_PlusDomainsAclentryResource';
   protected $itemDataType = '';
@@ -2082,51 +2120,43 @@ class Google_Service_PlusDomains_Audience extends Google_Model
   public $memberCount;
   public $visibility;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItem(Google_Service_PlusDomains_PlusDomainsAclentryResource $item)
   {
     $this->item = $item;
   }
-
   public function getItem()
   {
     return $this->item;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setMemberCount($memberCount)
   {
     $this->memberCount = $memberCount;
   }
-
   public function getMemberCount()
   {
     return $this->memberCount;
   }
-
   public function setVisibility($visibility)
   {
     $this->visibility = $visibility;
   }
-
   public function getVisibility()
   {
     return $this->visibility;
@@ -2135,6 +2165,9 @@ class Google_Service_PlusDomains_Audience extends Google_Model
 
 class Google_Service_PlusDomains_AudiencesFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_PlusDomains_Audience';
   protected $itemsDataType = 'array';
@@ -2142,51 +2175,43 @@ class Google_Service_PlusDomains_AudiencesFeed extends Google_Collection
   public $nextPageToken;
   public $totalItems;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2195,6 +2220,8 @@ class Google_Service_PlusDomains_AudiencesFeed extends Google_Collection
 
 class Google_Service_PlusDomains_Circle extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $description;
   public $displayName;
   public $etag;
@@ -2204,71 +2231,59 @@ class Google_Service_PlusDomains_Circle extends Google_Model
   protected $peopleDataType = '';
   public $selfLink;
 
+
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setPeople(Google_Service_PlusDomains_CirclePeople $people)
   {
     $this->people = $people;
   }
-
   public function getPeople()
   {
     return $this->people;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
@@ -2277,6 +2292,9 @@ class Google_Service_PlusDomains_Circle extends Google_Model
 
 class Google_Service_PlusDomains_CircleFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_PlusDomains_Circle';
   protected $itemsDataType = 'array';
@@ -2287,81 +2305,67 @@ class Google_Service_PlusDomains_CircleFeed extends Google_Collection
   public $title;
   public $totalItems;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextLink($nextLink)
   {
     $this->nextLink = $nextLink;
   }
-
   public function getNextLink()
   {
     return $this->nextLink;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2370,13 +2374,15 @@ class Google_Service_PlusDomains_CircleFeed extends Google_Collection
 
 class Google_Service_PlusDomains_CirclePeople extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $totalItems;
+
 
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2385,6 +2391,9 @@ class Google_Service_PlusDomains_CirclePeople extends Google_Model
 
 class Google_Service_PlusDomains_Comment extends Google_Collection
 {
+  protected $collection_key = 'inReplyTo';
+  protected $internal_gapi_mappings = array(
+  );
   protected $actorType = 'Google_Service_PlusDomains_CommentActor';
   protected $actorDataType = '';
   public $etag;
@@ -2401,111 +2410,91 @@ class Google_Service_PlusDomains_Comment extends Google_Collection
   public $updated;
   public $verb;
 
+
   public function setActor(Google_Service_PlusDomains_CommentActor $actor)
   {
     $this->actor = $actor;
   }
-
   public function getActor()
   {
     return $this->actor;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setInReplyTo($inReplyTo)
   {
     $this->inReplyTo = $inReplyTo;
   }
-
   public function getInReplyTo()
   {
     return $this->inReplyTo;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setObject(Google_Service_PlusDomains_CommentObject $object)
   {
     $this->object = $object;
   }
-
   public function getObject()
   {
     return $this->object;
   }
-
   public function setPlusoners(Google_Service_PlusDomains_CommentPlusoners $plusoners)
   {
     $this->plusoners = $plusoners;
   }
-
   public function getPlusoners()
   {
     return $this->plusoners;
   }
-
   public function setPublished($published)
   {
     $this->published = $published;
   }
-
   public function getPublished()
   {
     return $this->published;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
   }
-
   public function setVerb($verb)
   {
     $this->verb = $verb;
   }
-
   public function getVerb()
   {
     return $this->verb;
@@ -2514,70 +2503,143 @@ class Google_Service_PlusDomains_Comment extends Google_Collection
 
 class Google_Service_PlusDomains_CommentActor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  protected $clientSpecificActorInfoType = 'Google_Service_PlusDomains_CommentActorClientSpecificActorInfo';
+  protected $clientSpecificActorInfoDataType = '';
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_PlusDomains_CommentActorImage';
   protected $imageDataType = '';
   public $url;
+  protected $verificationType = 'Google_Service_PlusDomains_CommentActorVerification';
+  protected $verificationDataType = '';
 
+
+  public function setClientSpecificActorInfo(Google_Service_PlusDomains_CommentActorClientSpecificActorInfo $clientSpecificActorInfo)
+  {
+    $this->clientSpecificActorInfo = $clientSpecificActorInfo;
+  }
+  public function getClientSpecificActorInfo()
+  {
+    return $this->clientSpecificActorInfo;
+  }
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_PlusDomains_CommentActorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
+  }
+  public function setVerification(Google_Service_PlusDomains_CommentActorVerification $verification)
+  {
+    $this->verification = $verification;
+  }
+  public function getVerification()
+  {
+    return $this->verification;
+  }
+}
+
+class Google_Service_PlusDomains_CommentActorClientSpecificActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  protected $youtubeActorInfoType = 'Google_Service_PlusDomains_CommentActorClientSpecificActorInfoYoutubeActorInfo';
+  protected $youtubeActorInfoDataType = '';
+
+
+  public function setYoutubeActorInfo(Google_Service_PlusDomains_CommentActorClientSpecificActorInfoYoutubeActorInfo $youtubeActorInfo)
+  {
+    $this->youtubeActorInfo = $youtubeActorInfo;
+  }
+  public function getYoutubeActorInfo()
+  {
+    return $this->youtubeActorInfo;
+  }
+}
+
+class Google_Service_PlusDomains_CommentActorClientSpecificActorInfoYoutubeActorInfo extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $channelId;
+
+
+  public function setChannelId($channelId)
+  {
+    $this->channelId = $channelId;
+  }
+  public function getChannelId()
+  {
+    return $this->channelId;
   }
 }
 
 class Google_Service_PlusDomains_CommentActorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
 }
 
+class Google_Service_PlusDomains_CommentActorVerification extends Google_Model
+{
+  protected $internal_gapi_mappings = array(
+  );
+  public $adHocVerified;
+
+
+  public function setAdHocVerified($adHocVerified)
+  {
+    $this->adHocVerified = $adHocVerified;
+  }
+  public function getAdHocVerified()
+  {
+    return $this->adHocVerified;
+  }
+}
+
 class Google_Service_PlusDomains_CommentFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   public $id;
   protected $itemsType = 'Google_Service_PlusDomains_Comment';
@@ -2588,81 +2650,67 @@ class Google_Service_PlusDomains_CommentFeed extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextLink($nextLink)
   {
     $this->nextLink = $nextLink;
   }
-
   public function getNextLink()
   {
     return $this->nextLink;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -2671,24 +2719,24 @@ class Google_Service_PlusDomains_CommentFeed extends Google_Collection
 
 class Google_Service_PlusDomains_CommentInReplyTo extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $id;
   public $url;
+
 
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -2697,35 +2745,33 @@ class Google_Service_PlusDomains_CommentInReplyTo extends Google_Model
 
 class Google_Service_PlusDomains_CommentObject extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $content;
   public $objectType;
   public $originalContent;
+
 
   public function setContent($content)
   {
     $this->content = $content;
   }
-
   public function getContent()
   {
     return $this->content;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setOriginalContent($originalContent)
   {
     $this->originalContent = $originalContent;
   }
-
   public function getOriginalContent()
   {
     return $this->originalContent;
@@ -2734,13 +2780,15 @@ class Google_Service_PlusDomains_CommentObject extends Google_Model
 
 class Google_Service_PlusDomains_CommentPlusoners extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $totalItems;
+
 
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -2749,6 +2797,9 @@ class Google_Service_PlusDomains_CommentPlusoners extends Google_Model
 
 class Google_Service_PlusDomains_Media extends Google_Collection
 {
+  protected $collection_key = 'streams';
+  protected $internal_gapi_mappings = array(
+  );
   protected $authorType = 'Google_Service_PlusDomains_MediaAuthor';
   protected $authorDataType = '';
   public $displayName;
@@ -2771,181 +2822,147 @@ class Google_Service_PlusDomains_Media extends Google_Collection
   public $videoStatus;
   public $width;
 
+
   public function setAuthor(Google_Service_PlusDomains_MediaAuthor $author)
   {
     $this->author = $author;
   }
-
   public function getAuthor()
   {
     return $this->author;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setExif(Google_Service_PlusDomains_MediaExif $exif)
   {
     $this->exif = $exif;
   }
-
   public function getExif()
   {
     return $this->exif;
   }
-
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setMediaCreatedTime($mediaCreatedTime)
   {
     $this->mediaCreatedTime = $mediaCreatedTime;
   }
-
   public function getMediaCreatedTime()
   {
     return $this->mediaCreatedTime;
   }
-
   public function setMediaUrl($mediaUrl)
   {
     $this->mediaUrl = $mediaUrl;
   }
-
   public function getMediaUrl()
   {
     return $this->mediaUrl;
   }
-
   public function setPublished($published)
   {
     $this->published = $published;
   }
-
   public function getPublished()
   {
     return $this->published;
   }
-
   public function setSizeBytes($sizeBytes)
   {
     $this->sizeBytes = $sizeBytes;
   }
-
   public function getSizeBytes()
   {
     return $this->sizeBytes;
   }
-
   public function setStreams($streams)
   {
     $this->streams = $streams;
   }
-
   public function getStreams()
   {
     return $this->streams;
   }
-
   public function setSummary($summary)
   {
     $this->summary = $summary;
   }
-
   public function getSummary()
   {
     return $this->summary;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setVideoDuration($videoDuration)
   {
     $this->videoDuration = $videoDuration;
   }
-
   public function getVideoDuration()
   {
     return $this->videoDuration;
   }
-
   public function setVideoStatus($videoStatus)
   {
     $this->videoStatus = $videoStatus;
   }
-
   public function getVideoStatus()
   {
     return $this->videoStatus;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -2954,47 +2971,43 @@ class Google_Service_PlusDomains_Media extends Google_Collection
 
 class Google_Service_PlusDomains_MediaAuthor extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $displayName;
   public $id;
   protected $imageType = 'Google_Service_PlusDomains_MediaAuthorImage';
   protected $imageDataType = '';
   public $url;
 
+
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_PlusDomains_MediaAuthorImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -3003,13 +3016,15 @@ class Google_Service_PlusDomains_MediaAuthor extends Google_Model
 
 class Google_Service_PlusDomains_MediaAuthorImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $url;
+
 
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -3018,13 +3033,15 @@ class Google_Service_PlusDomains_MediaAuthorImage extends Google_Model
 
 class Google_Service_PlusDomains_MediaExif extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $time;
+
 
   public function setTime($time)
   {
     $this->time = $time;
   }
-
   public function getTime()
   {
     return $this->time;
@@ -3033,6 +3050,9 @@ class Google_Service_PlusDomains_MediaExif extends Google_Model
 
 class Google_Service_PlusDomains_PeopleFeed extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_PlusDomains_Person';
   protected $itemsDataType = 'array';
@@ -3042,71 +3062,59 @@ class Google_Service_PlusDomains_PeopleFeed extends Google_Collection
   public $title;
   public $totalItems;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
@@ -3115,6 +3123,9 @@ class Google_Service_PlusDomains_PeopleFeed extends Google_Collection
 
 class Google_Service_PlusDomains_Person extends Google_Collection
 {
+  protected $collection_key = 'urls';
+  protected $internal_gapi_mappings = array(
+  );
   public $aboutMe;
   public $birthday;
   public $braggingRights;
@@ -3151,281 +3162,227 @@ class Google_Service_PlusDomains_Person extends Google_Collection
   protected $urlsDataType = 'array';
   public $verified;
 
+
   public function setAboutMe($aboutMe)
   {
     $this->aboutMe = $aboutMe;
   }
-
   public function getAboutMe()
   {
     return $this->aboutMe;
   }
-
   public function setBirthday($birthday)
   {
     $this->birthday = $birthday;
   }
-
   public function getBirthday()
   {
     return $this->birthday;
   }
-
   public function setBraggingRights($braggingRights)
   {
     $this->braggingRights = $braggingRights;
   }
-
   public function getBraggingRights()
   {
     return $this->braggingRights;
   }
-
   public function setCircledByCount($circledByCount)
   {
     $this->circledByCount = $circledByCount;
   }
-
   public function getCircledByCount()
   {
     return $this->circledByCount;
   }
-
   public function setCover(Google_Service_PlusDomains_PersonCover $cover)
   {
     $this->cover = $cover;
   }
-
   public function getCover()
   {
     return $this->cover;
   }
-
   public function setCurrentLocation($currentLocation)
   {
     $this->currentLocation = $currentLocation;
   }
-
   public function getCurrentLocation()
   {
     return $this->currentLocation;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setDomain($domain)
   {
     $this->domain = $domain;
   }
-
   public function getDomain()
   {
     return $this->domain;
   }
-
   public function setEmails($emails)
   {
     $this->emails = $emails;
   }
-
   public function getEmails()
   {
     return $this->emails;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setGender($gender)
   {
     $this->gender = $gender;
   }
-
   public function getGender()
   {
     return $this->gender;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setImage(Google_Service_PlusDomains_PersonImage $image)
   {
     $this->image = $image;
   }
-
   public function getImage()
   {
     return $this->image;
   }
-
   public function setIsPlusUser($isPlusUser)
   {
     $this->isPlusUser = $isPlusUser;
   }
-
   public function getIsPlusUser()
   {
     return $this->isPlusUser;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setName(Google_Service_PlusDomains_PersonName $name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setNickname($nickname)
   {
     $this->nickname = $nickname;
   }
-
   public function getNickname()
   {
     return $this->nickname;
   }
-
   public function setObjectType($objectType)
   {
     $this->objectType = $objectType;
   }
-
   public function getObjectType()
   {
     return $this->objectType;
   }
-
   public function setOccupation($occupation)
   {
     $this->occupation = $occupation;
   }
-
   public function getOccupation()
   {
     return $this->occupation;
   }
-
   public function setOrganizations($organizations)
   {
     $this->organizations = $organizations;
   }
-
   public function getOrganizations()
   {
     return $this->organizations;
   }
-
   public function setPlacesLived($placesLived)
   {
     $this->placesLived = $placesLived;
   }
-
   public function getPlacesLived()
   {
     return $this->placesLived;
   }
-
   public function setPlusOneCount($plusOneCount)
   {
     $this->plusOneCount = $plusOneCount;
   }
-
   public function getPlusOneCount()
   {
     return $this->plusOneCount;
   }
-
   public function setRelationshipStatus($relationshipStatus)
   {
     $this->relationshipStatus = $relationshipStatus;
   }
-
   public function getRelationshipStatus()
   {
     return $this->relationshipStatus;
   }
-
   public function setSkills($skills)
   {
     $this->skills = $skills;
   }
-
   public function getSkills()
   {
     return $this->skills;
   }
-
   public function setTagline($tagline)
   {
     $this->tagline = $tagline;
   }
-
   public function getTagline()
   {
     return $this->tagline;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setUrls($urls)
   {
     $this->urls = $urls;
   }
-
   public function getUrls()
   {
     return $this->urls;
   }
-
   public function setVerified($verified)
   {
     $this->verified = $verified;
   }
-
   public function getVerified()
   {
     return $this->verified;
@@ -3434,37 +3391,35 @@ class Google_Service_PlusDomains_Person extends Google_Collection
 
 class Google_Service_PlusDomains_PersonCover extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $coverInfoType = 'Google_Service_PlusDomains_PersonCoverCoverInfo';
   protected $coverInfoDataType = '';
   protected $coverPhotoType = 'Google_Service_PlusDomains_PersonCoverCoverPhoto';
   protected $coverPhotoDataType = '';
   public $layout;
 
+
   public function setCoverInfo(Google_Service_PlusDomains_PersonCoverCoverInfo $coverInfo)
   {
     $this->coverInfo = $coverInfo;
   }
-
   public function getCoverInfo()
   {
     return $this->coverInfo;
   }
-
   public function setCoverPhoto(Google_Service_PlusDomains_PersonCoverCoverPhoto $coverPhoto)
   {
     $this->coverPhoto = $coverPhoto;
   }
-
   public function getCoverPhoto()
   {
     return $this->coverPhoto;
   }
-
   public function setLayout($layout)
   {
     $this->layout = $layout;
   }
-
   public function getLayout()
   {
     return $this->layout;
@@ -3473,24 +3428,24 @@ class Google_Service_PlusDomains_PersonCover extends Google_Model
 
 class Google_Service_PlusDomains_PersonCoverCoverInfo extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $leftImageOffset;
   public $topImageOffset;
+
 
   public function setLeftImageOffset($leftImageOffset)
   {
     $this->leftImageOffset = $leftImageOffset;
   }
-
   public function getLeftImageOffset()
   {
     return $this->leftImageOffset;
   }
-
   public function setTopImageOffset($topImageOffset)
   {
     $this->topImageOffset = $topImageOffset;
   }
-
   public function getTopImageOffset()
   {
     return $this->topImageOffset;
@@ -3499,35 +3454,33 @@ class Google_Service_PlusDomains_PersonCoverCoverInfo extends Google_Model
 
 class Google_Service_PlusDomains_PersonCoverCoverPhoto extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $url;
   public $width;
+
 
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
@@ -3536,24 +3489,24 @@ class Google_Service_PlusDomains_PersonCoverCoverPhoto extends Google_Model
 
 class Google_Service_PlusDomains_PersonEmails extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $type;
   public $value;
+
 
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -3562,13 +3515,24 @@ class Google_Service_PlusDomains_PersonEmails extends Google_Model
 
 class Google_Service_PlusDomains_PersonImage extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
+  public $isDefault;
   public $url;
 
+
+  public function setIsDefault($isDefault)
+  {
+    $this->isDefault = $isDefault;
+  }
+  public function getIsDefault()
+  {
+    return $this->isDefault;
+  }
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
@@ -3577,6 +3541,8 @@ class Google_Service_PlusDomains_PersonImage extends Google_Model
 
 class Google_Service_PlusDomains_PersonName extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $familyName;
   public $formatted;
   public $givenName;
@@ -3584,61 +3550,51 @@ class Google_Service_PlusDomains_PersonName extends Google_Model
   public $honorificSuffix;
   public $middleName;
 
+
   public function setFamilyName($familyName)
   {
     $this->familyName = $familyName;
   }
-
   public function getFamilyName()
   {
     return $this->familyName;
   }
-
   public function setFormatted($formatted)
   {
     $this->formatted = $formatted;
   }
-
   public function getFormatted()
   {
     return $this->formatted;
   }
-
   public function setGivenName($givenName)
   {
     $this->givenName = $givenName;
   }
-
   public function getGivenName()
   {
     return $this->givenName;
   }
-
   public function setHonorificPrefix($honorificPrefix)
   {
     $this->honorificPrefix = $honorificPrefix;
   }
-
   public function getHonorificPrefix()
   {
     return $this->honorificPrefix;
   }
-
   public function setHonorificSuffix($honorificSuffix)
   {
     $this->honorificSuffix = $honorificSuffix;
   }
-
   public function getHonorificSuffix()
   {
     return $this->honorificSuffix;
   }
-
   public function setMiddleName($middleName)
   {
     $this->middleName = $middleName;
   }
-
   public function getMiddleName()
   {
     return $this->middleName;
@@ -3647,6 +3603,8 @@ class Google_Service_PlusDomains_PersonName extends Google_Model
 
 class Google_Service_PlusDomains_PersonOrganizations extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $department;
   public $description;
   public $endDate;
@@ -3657,91 +3615,75 @@ class Google_Service_PlusDomains_PersonOrganizations extends Google_Model
   public $title;
   public $type;
 
+
   public function setDepartment($department)
   {
     $this->department = $department;
   }
-
   public function getDepartment()
   {
     return $this->department;
   }
-
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setEndDate($endDate)
   {
     $this->endDate = $endDate;
   }
-
   public function getEndDate()
   {
     return $this->endDate;
   }
-
   public function setLocation($location)
   {
     $this->location = $location;
   }
-
   public function getLocation()
   {
     return $this->location;
   }
-
   public function setName($name)
   {
     $this->name = $name;
   }
-
   public function getName()
   {
     return $this->name;
   }
-
   public function setPrimary($primary)
   {
     $this->primary = $primary;
   }
-
   public function getPrimary()
   {
     return $this->primary;
   }
-
   public function setStartDate($startDate)
   {
     $this->startDate = $startDate;
   }
-
   public function getStartDate()
   {
     return $this->startDate;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -3750,24 +3692,24 @@ class Google_Service_PlusDomains_PersonOrganizations extends Google_Model
 
 class Google_Service_PlusDomains_PersonPlacesLived extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $primary;
   public $value;
+
 
   public function setPrimary($primary)
   {
     $this->primary = $primary;
   }
-
   public function getPrimary()
   {
     return $this->primary;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -3776,35 +3718,33 @@ class Google_Service_PlusDomains_PersonPlacesLived extends Google_Model
 
 class Google_Service_PlusDomains_PersonUrls extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $label;
   public $type;
   public $value;
+
 
   public function setLabel($label)
   {
     $this->label = $label;
   }
-
   public function getLabel()
   {
     return $this->label;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setValue($value)
   {
     $this->value = $value;
   }
-
   public function getValue()
   {
     return $this->value;
@@ -3813,6 +3753,8 @@ class Google_Service_PlusDomains_PersonUrls extends Google_Model
 
 class Google_Service_PlusDomains_Place extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $addressType = 'Google_Service_PlusDomains_PlaceAddress';
   protected $addressDataType = '';
   public $displayName;
@@ -3821,51 +3763,43 @@ class Google_Service_PlusDomains_Place extends Google_Model
   protected $positionType = 'Google_Service_PlusDomains_PlacePosition';
   protected $positionDataType = '';
 
+
   public function setAddress(Google_Service_PlusDomains_PlaceAddress $address)
   {
     $this->address = $address;
   }
-
   public function getAddress()
   {
     return $this->address;
   }
-
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setPosition(Google_Service_PlusDomains_PlacePosition $position)
   {
     $this->position = $position;
   }
-
   public function getPosition()
   {
     return $this->position;
@@ -3874,13 +3808,15 @@ class Google_Service_PlusDomains_Place extends Google_Model
 
 class Google_Service_PlusDomains_PlaceAddress extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $formatted;
+
 
   public function setFormatted($formatted)
   {
     $this->formatted = $formatted;
   }
-
   public function getFormatted()
   {
     return $this->formatted;
@@ -3889,24 +3825,24 @@ class Google_Service_PlusDomains_PlaceAddress extends Google_Model
 
 class Google_Service_PlusDomains_PlacePosition extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $latitude;
   public $longitude;
+
 
   public function setLatitude($latitude)
   {
     $this->latitude = $latitude;
   }
-
   public function getLatitude()
   {
     return $this->latitude;
   }
-
   public function setLongitude($longitude)
   {
     $this->longitude = $longitude;
   }
-
   public function getLongitude()
   {
     return $this->longitude;
@@ -3915,35 +3851,33 @@ class Google_Service_PlusDomains_PlacePosition extends Google_Model
 
 class Google_Service_PlusDomains_PlusDomainsAclentryResource extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $displayName;
   public $id;
   public $type;
+
 
   public function setDisplayName($displayName)
   {
     $this->displayName = $displayName;
   }
-
   public function getDisplayName()
   {
     return $this->displayName;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -3952,46 +3886,42 @@ class Google_Service_PlusDomains_PlusDomainsAclentryResource extends Google_Mode
 
 class Google_Service_PlusDomains_Videostream extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $height;
   public $type;
   public $url;
   public $width;
 
+
   public function setHeight($height)
   {
     $this->height = $height;
   }
-
   public function getHeight()
   {
     return $this->height;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
   }
-
   public function setUrl($url)
   {
     $this->url = $url;
   }
-
   public function getUrl()
   {
     return $this->url;
   }
-
   public function setWidth($width)
   {
     $this->width = $width;
   }
-
   public function getWidth()
   {
     return $this->width;
