@@ -19,8 +19,7 @@
  * Service definition for Tasks (v1).
  *
  * <p>
- * Lets you manage your tasks and task lists.
- * </p>
+ * Lets you manage your tasks and task lists.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -32,9 +31,11 @@
 class Google_Service_Tasks extends Google_Service
 {
   /** Manage your tasks. */
-  const TASKS = "https://www.googleapis.com/auth/tasks";
+  const TASKS =
+      "https://www.googleapis.com/auth/tasks";
   /** View your tasks. */
-  const TASKS_READONLY = "https://www.googleapis.com/auth/tasks.readonly";
+  const TASKS_READONLY =
+      "https://www.googleapis.com/auth/tasks.readonly";
 
   public $tasklists;
   public $tasks;
@@ -48,6 +49,7 @@ class Google_Service_Tasks extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'tasks/v1/';
     $this->version = 'v1';
     $this->serviceName = 'tasks';
@@ -308,8 +310,7 @@ class Google_Service_Tasks_Tasklists_Resource extends Google_Service_Resource
   /**
    * Deletes the authenticated user's specified task list. (tasklists.delete)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param array $optParams Optional parameters.
    */
   public function delete($tasklist, $optParams = array())
@@ -318,11 +319,11 @@ class Google_Service_Tasks_Tasklists_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Returns the authenticated user's specified task list. (tasklists.get)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Tasks_TaskList
    */
@@ -332,6 +333,7 @@ class Google_Service_Tasks_Tasklists_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Tasks_TaskList");
   }
+
   /**
    * Creates a new task list and adds it to the authenticated user's task lists.
    * (tasklists.insert)
@@ -346,15 +348,16 @@ class Google_Service_Tasks_Tasklists_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Tasks_TaskList");
   }
+
   /**
    * Returns all the authenticated user's task lists. (tasklists.listTasklists)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string pageToken
-   * Token specifying the result page to return. Optional.
-   * @opt_param string maxResults
-   * Maximum number of task lists returned on one page. Optional. The default is 100.
+   * @opt_param string pageToken Token specifying the result page to return.
+   * Optional.
+   * @opt_param string maxResults Maximum number of task lists returned on one
+   * page. Optional. The default is 100.
    * @return Google_Service_Tasks_TaskLists
    */
   public function listTasklists($optParams = array())
@@ -363,12 +366,12 @@ class Google_Service_Tasks_Tasklists_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Tasks_TaskLists");
   }
+
   /**
    * Updates the authenticated user's specified task list. This method supports
    * patch semantics. (tasklists.patch)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param Google_TaskList $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Tasks_TaskList
@@ -379,11 +382,11 @@ class Google_Service_Tasks_Tasklists_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_Tasks_TaskList");
   }
+
   /**
    * Updates the authenticated user's specified task list. (tasklists.update)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param Google_TaskList $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Tasks_TaskList
@@ -412,8 +415,7 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
    * will be marked as 'hidden' and no longer be returned by default when
    * retrieving all tasks for a task list. (tasks.clear)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param array $optParams Optional parameters.
    */
   public function clear($tasklist, $optParams = array())
@@ -422,13 +424,12 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('clear', array($params));
   }
+
   /**
    * Deletes the specified task from the task list. (tasks.delete)
    *
-   * @param string $tasklist
-   * Task list identifier.
-   * @param string $task
-   * Task identifier.
+   * @param string $tasklist Task list identifier.
+   * @param string $task Task identifier.
    * @param array $optParams Optional parameters.
    */
   public function delete($tasklist, $task, $optParams = array())
@@ -437,13 +438,12 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('delete', array($params));
   }
+
   /**
    * Returns the specified task. (tasks.get)
    *
-   * @param string $tasklist
-   * Task list identifier.
-   * @param string $task
-   * Task identifier.
+   * @param string $tasklist Task list identifier.
+   * @param string $task Task identifier.
    * @param array $optParams Optional parameters.
    * @return Google_Service_Tasks_Task
    */
@@ -453,20 +453,19 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Tasks_Task");
   }
+
   /**
    * Creates a new task on the specified task list. (tasks.insert)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param Google_Task $postBody
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string parent
-   * Parent task identifier. If the task is created at the top level, this parameter is omitted.
-    * Optional.
-   * @opt_param string previous
-   * Previous sibling task identifier. If the task is created at the first position among its
-    * siblings, this parameter is omitted. Optional.
+   * @opt_param string parent Parent task identifier. If the task is created at
+   * the top level, this parameter is omitted. Optional.
+   * @opt_param string previous Previous sibling task identifier. If the task is
+   * created at the first position among its siblings, this parameter is omitted.
+   * Optional.
    * @return Google_Service_Tasks_Task
    */
   public function insert($tasklist, Google_Service_Tasks_Task $postBody, $optParams = array())
@@ -475,40 +474,36 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Tasks_Task");
   }
+
   /**
    * Returns all tasks in the specified task list. (tasks.listTasks)
    *
-   * @param string $tasklist
-   * Task list identifier.
+   * @param string $tasklist Task list identifier.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string dueMax
-   * Upper bound for a task's due date (as a RFC 3339 timestamp) to filter by. Optional. The default
-    * is not to filter by due date.
-   * @opt_param bool showDeleted
-   * Flag indicating whether deleted tasks are returned in the result. Optional. The default is
-    * False.
-   * @opt_param string updatedMin
-   * Lower bound for a task's last modification time (as a RFC 3339 timestamp) to filter by.
-    * Optional. The default is not to filter by last modification time.
-   * @opt_param string completedMin
-   * Lower bound for a task's completion date (as a RFC 3339 timestamp) to filter by. Optional. The
-    * default is not to filter by completion date.
-   * @opt_param string maxResults
-   * Maximum number of task lists returned on one page. Optional. The default is 100.
-   * @opt_param bool showCompleted
-   * Flag indicating whether completed tasks are returned in the result. Optional. The default is
-    * True.
-   * @opt_param string pageToken
-   * Token specifying the result page to return. Optional.
-   * @opt_param string completedMax
-   * Upper bound for a task's completion date (as a RFC 3339 timestamp) to filter by. Optional. The
-    * default is not to filter by completion date.
-   * @opt_param bool showHidden
-   * Flag indicating whether hidden tasks are returned in the result. Optional. The default is False.
-   * @opt_param string dueMin
-   * Lower bound for a task's due date (as a RFC 3339 timestamp) to filter by. Optional. The default
-    * is not to filter by due date.
+   * @opt_param string dueMax Upper bound for a task's due date (as a RFC 3339
+   * timestamp) to filter by. Optional. The default is not to filter by due date.
+   * @opt_param bool showDeleted Flag indicating whether deleted tasks are
+   * returned in the result. Optional. The default is False.
+   * @opt_param string updatedMin Lower bound for a task's last modification time
+   * (as a RFC 3339 timestamp) to filter by. Optional. The default is not to
+   * filter by last modification time.
+   * @opt_param string completedMin Lower bound for a task's completion date (as a
+   * RFC 3339 timestamp) to filter by. Optional. The default is not to filter by
+   * completion date.
+   * @opt_param string maxResults Maximum number of task lists returned on one
+   * page. Optional. The default is 100.
+   * @opt_param bool showCompleted Flag indicating whether completed tasks are
+   * returned in the result. Optional. The default is True.
+   * @opt_param string pageToken Token specifying the result page to return.
+   * Optional.
+   * @opt_param string completedMax Upper bound for a task's completion date (as a
+   * RFC 3339 timestamp) to filter by. Optional. The default is not to filter by
+   * completion date.
+   * @opt_param bool showHidden Flag indicating whether hidden tasks are returned
+   * in the result. Optional. The default is False.
+   * @opt_param string dueMin Lower bound for a task's due date (as a RFC 3339
+   * timestamp) to filter by. Optional. The default is not to filter by due date.
    * @return Google_Service_Tasks_Tasks
    */
   public function listTasks($tasklist, $optParams = array())
@@ -517,23 +512,21 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('list', array($params), "Google_Service_Tasks_Tasks");
   }
+
   /**
    * Moves the specified task to another position in the task list. This can
    * include putting it as a child task under a new parent and/or move it to a
    * different position among its sibling tasks. (tasks.move)
    *
-   * @param string $tasklist
-   * Task list identifier.
-   * @param string $task
-   * Task identifier.
+   * @param string $tasklist Task list identifier.
+   * @param string $task Task identifier.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string parent
-   * New parent task identifier. If the task is moved to the top level, this parameter is omitted.
-    * Optional.
-   * @opt_param string previous
-   * New previous sibling task identifier. If the task is moved to the first position among its
-    * siblings, this parameter is omitted. Optional.
+   * @opt_param string parent New parent task identifier. If the task is moved to
+   * the top level, this parameter is omitted. Optional.
+   * @opt_param string previous New previous sibling task identifier. If the task
+   * is moved to the first position among its siblings, this parameter is omitted.
+   * Optional.
    * @return Google_Service_Tasks_Task
    */
   public function move($tasklist, $task, $optParams = array())
@@ -542,14 +535,13 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('move', array($params), "Google_Service_Tasks_Task");
   }
+
   /**
    * Updates the specified task. This method supports patch semantics.
    * (tasks.patch)
    *
-   * @param string $tasklist
-   * Task list identifier.
-   * @param string $task
-   * Task identifier.
+   * @param string $tasklist Task list identifier.
+   * @param string $task Task identifier.
    * @param Google_Task $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Tasks_Task
@@ -560,13 +552,12 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('patch', array($params), "Google_Service_Tasks_Task");
   }
+
   /**
    * Updates the specified task. (tasks.update)
    *
-   * @param string $tasklist
-   * Task list identifier.
-   * @param string $task
-   * Task identifier.
+   * @param string $tasklist Task list identifier.
+   * @param string $task Task identifier.
    * @param Google_Task $postBody
    * @param array $optParams Optional parameters.
    * @return Google_Service_Tasks_Task
@@ -584,6 +575,9 @@ class Google_Service_Tasks_Tasks_Resource extends Google_Service_Resource
 
 class Google_Service_Tasks_Task extends Google_Collection
 {
+  protected $collection_key = 'links';
+  protected $internal_gapi_mappings = array(
+  );
   public $completed;
   public $deleted;
   public $due;
@@ -601,151 +595,123 @@ class Google_Service_Tasks_Task extends Google_Collection
   public $title;
   public $updated;
 
+
   public function setCompleted($completed)
   {
     $this->completed = $completed;
   }
-
   public function getCompleted()
   {
     return $this->completed;
   }
-
   public function setDeleted($deleted)
   {
     $this->deleted = $deleted;
   }
-
   public function getDeleted()
   {
     return $this->deleted;
   }
-
   public function setDue($due)
   {
     $this->due = $due;
   }
-
   public function getDue()
   {
     return $this->due;
   }
-
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setHidden($hidden)
   {
     $this->hidden = $hidden;
   }
-
   public function getHidden()
   {
     return $this->hidden;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLinks($links)
   {
     $this->links = $links;
   }
-
   public function getLinks()
   {
     return $this->links;
   }
-
   public function setNotes($notes)
   {
     $this->notes = $notes;
   }
-
   public function getNotes()
   {
     return $this->notes;
   }
-
   public function setParent($parent)
   {
     $this->parent = $parent;
   }
-
   public function getParent()
   {
     return $this->parent;
   }
-
   public function setPosition($position)
   {
     $this->position = $position;
   }
-
   public function getPosition()
   {
     return $this->position;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setStatus($status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -754,35 +720,33 @@ class Google_Service_Tasks_Task extends Google_Collection
 
 class Google_Service_Tasks_TaskLinks extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $description;
   public $link;
   public $type;
+
 
   public function setDescription($description)
   {
     $this->description = $description;
   }
-
   public function getDescription()
   {
     return $this->description;
   }
-
   public function setLink($link)
   {
     $this->link = $link;
   }
-
   public function getLink()
   {
     return $this->link;
   }
-
   public function setType($type)
   {
     $this->type = $type;
   }
-
   public function getType()
   {
     return $this->type;
@@ -791,6 +755,8 @@ class Google_Service_Tasks_TaskLinks extends Google_Model
 
 class Google_Service_Tasks_TaskList extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   public $id;
   public $kind;
@@ -798,61 +764,51 @@ class Google_Service_Tasks_TaskList extends Google_Model
   public $title;
   public $updated;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setSelfLink($selfLink)
   {
     $this->selfLink = $selfLink;
   }
-
   public function getSelfLink()
   {
     return $this->selfLink;
   }
-
   public function setTitle($title)
   {
     $this->title = $title;
   }
-
   public function getTitle()
   {
     return $this->title;
   }
-
   public function setUpdated($updated)
   {
     $this->updated = $updated;
   }
-
   public function getUpdated()
   {
     return $this->updated;
@@ -861,47 +817,44 @@ class Google_Service_Tasks_TaskList extends Google_Model
 
 class Google_Service_Tasks_TaskLists extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_Tasks_TaskList';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
@@ -910,47 +863,44 @@ class Google_Service_Tasks_TaskLists extends Google_Collection
 
 class Google_Service_Tasks_Tasks extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   public $etag;
   protected $itemsType = 'Google_Service_Tasks_Task';
   protected $itemsDataType = 'array';
   public $kind;
   public $nextPageToken;
 
+
   public function setEtag($etag)
   {
     $this->etag = $etag;
   }
-
   public function getEtag()
   {
     return $this->etag;
   }
-
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
