@@ -19,12 +19,11 @@
  * Service definition for Urlshortener (v1).
  *
  * <p>
- * Lets you create, inspect, and manage goo.gl short URLs
- * </p>
+ * Lets you create, inspect, and manage goo.gl short URLs</p>
  *
  * <p>
  * For more information about this service, see the API
- * <a href="http://code.google.com/apis/urlshortener/v1/getting_started.html" target="_blank">Documentation</a>
+ * <a href="https://developers.google.com/url-shortener/v1/getting_started" target="_blank">Documentation</a>
  * </p>
  *
  * @author Google, Inc.
@@ -32,7 +31,8 @@
 class Google_Service_Urlshortener extends Google_Service
 {
   /** Manage your goo.gl short URLs. */
-  const URLSHORTENER = "https://www.googleapis.com/auth/urlshortener";
+  const URLSHORTENER =
+      "https://www.googleapis.com/auth/urlshortener";
 
   public $url;
   
@@ -45,6 +45,7 @@ class Google_Service_Urlshortener extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'urlshortener/v1/';
     $this->version = 'v1';
     $this->serviceName = 'urlshortener';
@@ -108,12 +109,10 @@ class Google_Service_Urlshortener_Url_Resource extends Google_Service_Resource
   /**
    * Expands a short URL or gets creation time and analytics. (url.get)
    *
-   * @param string $shortUrl
-   * The short URL, including the protocol.
+   * @param string $shortUrl The short URL, including the protocol.
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string projection
-   * Additional information to return.
+   * @opt_param string projection Additional information to return.
    * @return Google_Service_Urlshortener_Url
    */
   public function get($shortUrl, $optParams = array())
@@ -122,6 +121,7 @@ class Google_Service_Urlshortener_Url_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('get', array($params), "Google_Service_Urlshortener_Url");
   }
+
   /**
    * Creates a new short URL. (url.insert)
    *
@@ -135,15 +135,15 @@ class Google_Service_Urlshortener_Url_Resource extends Google_Service_Resource
     $params = array_merge($params, $optParams);
     return $this->call('insert', array($params), "Google_Service_Urlshortener_Url");
   }
+
   /**
    * Retrieves a list of URLs shortened by a user. (url.listUrl)
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string start-token
-   * Token for requesting successive pages of results.
-   * @opt_param string projection
-   * Additional information to return.
+   * @opt_param string start-token Token for requesting successive pages of
+   * results.
+   * @opt_param string projection Additional information to return.
    * @return Google_Service_Urlshortener_UrlHistory
    */
   public function listUrl($optParams = array())
@@ -159,6 +159,9 @@ class Google_Service_Urlshortener_Url_Resource extends Google_Service_Resource
 
 class Google_Service_Urlshortener_AnalyticsSnapshot extends Google_Collection
 {
+  protected $collection_key = 'referrers';
+  protected $internal_gapi_mappings = array(
+  );
   protected $browsersType = 'Google_Service_Urlshortener_StringCount';
   protected $browsersDataType = 'array';
   protected $countriesType = 'Google_Service_Urlshortener_StringCount';
@@ -170,61 +173,51 @@ class Google_Service_Urlshortener_AnalyticsSnapshot extends Google_Collection
   protected $referrersDataType = 'array';
   public $shortUrlClicks;
 
+
   public function setBrowsers($browsers)
   {
     $this->browsers = $browsers;
   }
-
   public function getBrowsers()
   {
     return $this->browsers;
   }
-
   public function setCountries($countries)
   {
     $this->countries = $countries;
   }
-
   public function getCountries()
   {
     return $this->countries;
   }
-
   public function setLongUrlClicks($longUrlClicks)
   {
     $this->longUrlClicks = $longUrlClicks;
   }
-
   public function getLongUrlClicks()
   {
     return $this->longUrlClicks;
   }
-
   public function setPlatforms($platforms)
   {
     $this->platforms = $platforms;
   }
-
   public function getPlatforms()
   {
     return $this->platforms;
   }
-
   public function setReferrers($referrers)
   {
     $this->referrers = $referrers;
   }
-
   public function getReferrers()
   {
     return $this->referrers;
   }
-
   public function setShortUrlClicks($shortUrlClicks)
   {
     $this->shortUrlClicks = $shortUrlClicks;
   }
-
   public function getShortUrlClicks()
   {
     return $this->shortUrlClicks;
@@ -233,6 +226,8 @@ class Google_Service_Urlshortener_AnalyticsSnapshot extends Google_Collection
 
 class Google_Service_Urlshortener_AnalyticsSummary extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $allTimeType = 'Google_Service_Urlshortener_AnalyticsSnapshot';
   protected $allTimeDataType = '';
   protected $dayType = 'Google_Service_Urlshortener_AnalyticsSnapshot';
@@ -244,51 +239,43 @@ class Google_Service_Urlshortener_AnalyticsSummary extends Google_Model
   protected $weekType = 'Google_Service_Urlshortener_AnalyticsSnapshot';
   protected $weekDataType = '';
 
+
   public function setAllTime(Google_Service_Urlshortener_AnalyticsSnapshot $allTime)
   {
     $this->allTime = $allTime;
   }
-
   public function getAllTime()
   {
     return $this->allTime;
   }
-
   public function setDay(Google_Service_Urlshortener_AnalyticsSnapshot $day)
   {
     $this->day = $day;
   }
-
   public function getDay()
   {
     return $this->day;
   }
-
   public function setMonth(Google_Service_Urlshortener_AnalyticsSnapshot $month)
   {
     $this->month = $month;
   }
-
   public function getMonth()
   {
     return $this->month;
   }
-
   public function setTwoHours(Google_Service_Urlshortener_AnalyticsSnapshot $twoHours)
   {
     $this->twoHours = $twoHours;
   }
-
   public function getTwoHours()
   {
     return $this->twoHours;
   }
-
   public function setWeek(Google_Service_Urlshortener_AnalyticsSnapshot $week)
   {
     $this->week = $week;
   }
-
   public function getWeek()
   {
     return $this->week;
@@ -297,24 +284,24 @@ class Google_Service_Urlshortener_AnalyticsSummary extends Google_Model
 
 class Google_Service_Urlshortener_StringCount extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   public $count;
   public $id;
+
 
   public function setCount($count)
   {
     $this->count = $count;
   }
-
   public function getCount()
   {
     return $this->count;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
@@ -323,6 +310,8 @@ class Google_Service_Urlshortener_StringCount extends Google_Model
 
 class Google_Service_Urlshortener_Url extends Google_Model
 {
+  protected $internal_gapi_mappings = array(
+  );
   protected $analyticsType = 'Google_Service_Urlshortener_AnalyticsSummary';
   protected $analyticsDataType = '';
   public $created;
@@ -331,61 +320,51 @@ class Google_Service_Urlshortener_Url extends Google_Model
   public $longUrl;
   public $status;
 
+
   public function setAnalytics(Google_Service_Urlshortener_AnalyticsSummary $analytics)
   {
     $this->analytics = $analytics;
   }
-
   public function getAnalytics()
   {
     return $this->analytics;
   }
-
   public function setCreated($created)
   {
     $this->created = $created;
   }
-
   public function getCreated()
   {
     return $this->created;
   }
-
   public function setId($id)
   {
     $this->id = $id;
   }
-
   public function getId()
   {
     return $this->id;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLongUrl($longUrl)
   {
     $this->longUrl = $longUrl;
   }
-
   public function getLongUrl()
   {
     return $this->longUrl;
   }
-
   public function setStatus($status)
   {
     $this->status = $status;
   }
-
   public function getStatus()
   {
     return $this->status;
@@ -394,6 +373,9 @@ class Google_Service_Urlshortener_Url extends Google_Model
 
 class Google_Service_Urlshortener_UrlHistory extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Urlshortener_Url';
   protected $itemsDataType = 'array';
   public $itemsPerPage;
@@ -401,51 +383,43 @@ class Google_Service_Urlshortener_UrlHistory extends Google_Collection
   public $nextPageToken;
   public $totalItems;
 
+
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setItemsPerPage($itemsPerPage)
   {
     $this->itemsPerPage = $itemsPerPage;
   }
-
   public function getItemsPerPage()
   {
     return $this->itemsPerPage;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setNextPageToken($nextPageToken)
   {
     $this->nextPageToken = $nextPageToken;
   }
-
   public function getNextPageToken()
   {
     return $this->nextPageToken;
   }
-
   public function setTotalItems($totalItems)
   {
     $this->totalItems = $totalItems;
   }
-
   public function getTotalItems()
   {
     return $this->totalItems;
