@@ -19,8 +19,7 @@
  * Service definition for Webfonts (v1).
  *
  * <p>
- * The Google Fonts Developer API.
- * </p>
+ * The Google Fonts Developer API.</p>
  *
  * <p>
  * For more information about this service, see the API
@@ -44,6 +43,7 @@ class Google_Service_Webfonts extends Google_Service
   public function __construct(Google_Client $client)
   {
     parent::__construct($client);
+    $this->rootUrl = 'https://www.googleapis.com/';
     $this->servicePath = 'webfonts/v1/';
     $this->version = 'v1';
     $this->serviceName = 'webfonts';
@@ -88,8 +88,7 @@ class Google_Service_Webfonts_Webfonts_Resource extends Google_Service_Resource
    *
    * @param array $optParams Optional parameters.
    *
-   * @opt_param string sort
-   * Enables sorting of the list
+   * @opt_param string sort Enables sorting of the list
    * @return Google_Service_Webfonts_WebfontList
    */
   public function listWebfonts($optParams = array())
@@ -105,6 +104,9 @@ class Google_Service_Webfonts_Webfonts_Resource extends Google_Service_Resource
 
 class Google_Service_Webfonts_Webfont extends Google_Collection
 {
+  protected $collection_key = 'variants';
+  protected $internal_gapi_mappings = array(
+  );
   public $category;
   public $family;
   public $files;
@@ -114,108 +116,99 @@ class Google_Service_Webfonts_Webfont extends Google_Collection
   public $variants;
   public $version;
 
+
   public function setCategory($category)
   {
     $this->category = $category;
   }
-
   public function getCategory()
   {
     return $this->category;
   }
-
   public function setFamily($family)
   {
     $this->family = $family;
   }
-
   public function getFamily()
   {
     return $this->family;
   }
-
   public function setFiles($files)
   {
     $this->files = $files;
   }
-
   public function getFiles()
   {
     return $this->files;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
   }
-
   public function setLastModified($lastModified)
   {
     $this->lastModified = $lastModified;
   }
-
   public function getLastModified()
   {
     return $this->lastModified;
   }
-
   public function setSubsets($subsets)
   {
     $this->subsets = $subsets;
   }
-
   public function getSubsets()
   {
     return $this->subsets;
   }
-
   public function setVariants($variants)
   {
     $this->variants = $variants;
   }
-
   public function getVariants()
   {
     return $this->variants;
   }
-
   public function setVersion($version)
   {
     $this->version = $version;
   }
-
   public function getVersion()
   {
     return $this->version;
   }
 }
 
+class Google_Service_Webfonts_WebfontFiles extends Google_Model
+{
+}
+
 class Google_Service_Webfonts_WebfontList extends Google_Collection
 {
+  protected $collection_key = 'items';
+  protected $internal_gapi_mappings = array(
+  );
   protected $itemsType = 'Google_Service_Webfonts_Webfont';
   protected $itemsDataType = 'array';
   public $kind;
+
 
   public function setItems($items)
   {
     $this->items = $items;
   }
-
   public function getItems()
   {
     return $this->items;
   }
-
   public function setKind($kind)
   {
     $this->kind = $kind;
   }
-
   public function getKind()
   {
     return $this->kind;
