@@ -16,17 +16,17 @@
  */
 
 /**
- * Implementation of levels 1-3 of the URI Template spec. 
+ * Implementation of levels 1-3 of the URI Template spec.
  * @see http://tools.ietf.org/html/rfc6570
  */
-class Google_Utils_URITemplate
+class Google_Utils_UriTemplate
 {
   const TYPE_MAP = "1";
   const TYPE_LIST = "2";
   const TYPE_SCALAR = "4";
 
   /**
-   * @var $operators array 
+   * @var $operators array
    * These are valid at the start of a template block to
    * modify the way in which the variables inside are
    * processed.
@@ -64,7 +64,7 @@ class Google_Utils_URITemplate
   /**
    * This function finds the first matching {...} block and
    * executes the replacement. It then calls itself to find
-   * subsequent blocks, if any. 
+   * subsequent blocks, if any.
    */
   private function resolveNextSection($string, $parameters)
   {
@@ -181,7 +181,7 @@ class Google_Utils_URITemplate
       );
     }
   }
- 
+
   public function combine(
       $key,
       $parameters,
@@ -200,20 +200,20 @@ class Google_Utils_URITemplate
     if (strpos($key, ":") !== false) {
       list($key, $length) = explode(":", $key);
     }
-    
+
     // Check for explode parameter.
     if ($key[strlen($key) - 1] == "*") {
       $explode = true;
       $key = substr($key, 0, -1);
       $skip_final_combine = true;
     }
-    
+
     // Define the list separator.
     $list_sep = $explode ? $sep : ",";
-    
+
     if (isset($parameters[$key])) {
       $data_type = $this->getDataType($parameters[$key]);
-      switch($data_type) {
+      switch ($data_type) {
         case self::TYPE_SCALAR:
           $value = $this->getValue($parameters[$key], $length);
           break;
@@ -267,11 +267,11 @@ class Google_Utils_URITemplate
     if (!$combine || $skip_final_combine) {
       return $value;
     }
-        
+
     // Else we combine the key name: foo=bar, if value is not the empty string.
     return $key . ($value != '' || $combine_on_empty ? $combine . $value : '');
   }
-  
+
   /**
    * Return the type of a passed in value
    */
@@ -286,7 +286,7 @@ class Google_Utils_URITemplate
     }
     return self::TYPE_SCALAR;
   }
-  
+
   /**
    * Utility function that merges multiple combine calls
    * for multi-key templates.
@@ -318,7 +318,7 @@ class Google_Utils_URITemplate
     }
     return implode($sep, $ret);
   }
-  
+
   /**
    * Utility function to encode and trim values
    */
